wx (wx v2.3.1)

A port of wxWidgets.

A port of wxWidgets.

This is the base api of wxWidgets. This module contains functions for starting and stopping the wx-server, as well as other utility functions.

wxWidgets is object oriented, and not functional. Thus, in wxErlang a module represents a class, and the object created by this class has an own type, wxCLASS(). This module represents the base class, and all other wxMODULE's are sub-classes of this class.

Objects of a class are created with wxCLASS:new(...) and destroyed with wxCLASS:destroy(). Member functions are called with wxCLASS:member(Object, ...) instead of as in C++ Object.member(...).

Sub class modules inherit (non static) functions from their parents. The inherited functions are not documented in the sub-classes.

This erlang port of wxWidgets tries to be a one-to-one mapping with the original wxWidgets library. Some things are different though, as the optional arguments use property lists and can be in any order. The main difference is the event handling which is different from the original library. See wxEvtHandler.

The following classes are implemented directly as erlang types:
wxPoint={x,y},wxSize={w,h},wxRect={x,y,w,h},wxColour={r,g,b [,a]}, wxString=unicode:chardata(), wxGBPosition={r,c},wxGBSpan={rs,cs},wxGridCellCoords={r,c}.

wxWidgets uses a process specific environment, which is created by wx:new/0. To be able to use the environment from other processes, call get_env/0 to retrieve the environment and set_env/1 to assign the environment in the other process.

Global (classless) functions are located in the wx_misc module.

DATA TYPES

wx_colour() = {R::byte(), G::byte(), B::byte()} | wx_colour4()

wx_colour4() = {R::byte(), G::byte(), B::byte(), A::byte()}

wx_datetime() = {{Year::integer(), Month::integer(), Day::integer()}, {Hour::integer(), Minute::integer(), Second::integer()}}

In Local Timezone

wx_enum() = integer()

Constant defined in wx.hrl

wx_env() = #wx_env{}

Opaque process environment

wx_memory() = binary() | #wx_mem{}

Opaque memory reference

wx_object() = #wx_ref{}

Opaque object reference

wx_wxHtmlLinkInfo() = #wxHtmlLinkInfo{href=unicode:chardata(), target=unicode:chardata()}

wx_wxMouseState() = #wxMouseState{x=integer(), y=integer(), leftDown=boolean(), middleDown=boolean(), rightDown=boolean(), controlDown=boolean(), shiftDown=boolean(), altDown=boolean(), metaDown=boolean(), cmdDown=boolean()}

See #wxMouseState{} defined in wx.hrl

Types

event()

-type event() ::
    wxActivateEvent:wxActivate() |
    wxAuiManagerEvent:wxAuiManager() |
    wxAuiNotebookEvent:wxAuiNotebook() |
    wxBookCtrlEvent:wxBookCtrl() |
    wxCalendarEvent:wxCalendar() |
    wxChildFocusEvent:wxChildFocus() |
    wxClipboardTextEvent:wxClipboardText() |
    wxCloseEvent:wxClose() |
    wxColourPickerEvent:wxColourPicker() |
    wxCommandEvent:wxCommand() |
    wxContextMenuEvent:wxContextMenu() |
    wxDateEvent:wxDate() |
    wxDisplayChangedEvent:wxDisplayChanged() |
    wxDropFilesEvent:wxDropFiles() |
    wxEraseEvent:wxErase() |
    wxFileDirPickerEvent:wxFileDirPicker() |
    wxFocusEvent:wxFocus() |
    wxFontPickerEvent:wxFontPicker() |
    wxGridEvent:wxGrid() |
    wxHelpEvent:wxHelp() |
    wxHtmlLinkEvent:wxHtmlLink() |
    wxIconizeEvent:wxIconize() |
    wxIdleEvent:wxIdle() |
    wxInitDialogEvent:wxInitDialog() |
    wxJoystickEvent:wxJoystick() |
    wxKeyEvent:wxKey() |
    wxListEvent:wxList() |
    wxMaximizeEvent:wxMaximize() |
    wxMenuEvent:wxMenu() |
    wxMouseCaptureChangedEvent:wxMouseCaptureChanged() |
    wxMouseCaptureLostEvent:wxMouseCaptureLost() |
    wxMouseEvent:wxMouse() |
    wxMoveEvent:wxMove() |
    wxNavigationKeyEvent:wxNavigationKey() |
    wxPaintEvent:wxPaint() |
    wxPaletteChangedEvent:wxPaletteChanged() |
    wxQueryNewPaletteEvent:wxQueryNewPalette() |
    wxSashEvent:wxSash() |
    wxScrollEvent:wxScroll() |
    wxScrollWinEvent:wxScrollWin() |
    wxSetCursorEvent:wxSetCursor() |
    wxShowEvent:wxShow() |
    wxSizeEvent:wxSize() |
    wxSpinEvent:wxSpin() |
    wxSplitterEvent:wxSplitter() |
    wxStyledTextEvent:wxStyledText() |
    wxSysColourChangedEvent:wxSysColourChanged() |
    wxTaskBarIconEvent:wxTaskBarIcon() |
    wxTreeEvent:wxTree() |
    wxUpdateUIEvent:wxUpdateUI() |
    wxWebViewEvent:wxWebView() |
    wxWindowCreateEvent:wxWindowCreate() |
    wxWindowDestroyEvent:wxWindowDestroy().

wx()

-type wx() :: #wx{}.

wx_colour4/0

-type wx_colour4() :: {R :: byte(), G :: byte(), B :: byte(), A :: byte()}.

wx_colour/0

-type wx_colour() :: {R :: byte(), G :: byte(), B :: byte()} | wx_colour4().

wx_datetime/0

-type wx_datetime() ::
    {{Year :: integer(), Month :: integer(), Day :: integer()},
     {Hour :: integer(), Minute :: integer(), Second :: integer()}}.

wx_enum/0

-type wx_enum() :: integer().

wx_env/0

-type wx_env() :: #wx_env{}.

wx_memory/0

-type wx_memory() :: binary() | #wx_mem{}.

wx_object/0

-type wx_object() :: #wx_ref{}.

wx_wxHtmlLinkInfo/0

-type wx_wxHtmlLinkInfo() :: #wxHtmlLinkInfo{}.

wx_wxMouseState/0

-type wx_wxMouseState() :: #wxMouseState{}.

wxActivate()

-type wxActivate() :: #wxActivate{}.

wxAuiManager()

-type wxAuiManager() :: #wxAuiManager{}.

wxAuiNotebook()

-type wxAuiNotebook() :: #wxAuiNotebook{}.

wxBookCtrl()

-type wxBookCtrl() :: #wxBookCtrl{}.

wxCalendar()

-type wxCalendar() :: #wxCalendar{}.

wxChildFocus()

-type wxChildFocus() :: #wxChildFocus{}.

wxClipboardText()

-type wxClipboardText() :: #wxClipboardText{}.

wxClose()

-type wxClose() :: #wxClose{}.

wxColourPicker()

-type wxColourPicker() :: #wxColourPicker{}.

wxCommand()

-type wxCommand() :: #wxCommand{}.

wxContextMenu()

-type wxContextMenu() :: #wxContextMenu{}.

wxDate()

-type wxDate() :: #wxDate{}.

wxDisplayChanged()

-type wxDisplayChanged() :: #wxDisplayChanged{}.

wxDropFiles()

-type wxDropFiles() :: #wxDropFiles{}.

wxErase()

-type wxErase() :: #wxErase{}.

wxEventType()

-type wxEventType() ::
    wxActivateEvent:wxActivateEventType() |
    wxAuiManagerEvent:wxAuiManagerEventType() |
    wxAuiNotebookEvent:wxAuiNotebookEventType() |
    wxBookCtrlEvent:wxBookCtrlEventType() |
    wxCalendarEvent:wxCalendarEventType() |
    wxChildFocusEvent:wxChildFocusEventType() |
    wxClipboardTextEvent:wxClipboardTextEventType() |
    wxCloseEvent:wxCloseEventType() |
    wxColourPickerEvent:wxColourPickerEventType() |
    wxCommandEvent:wxCommandEventType() |
    wxContextMenuEvent:wxContextMenuEventType() |
    wxDateEvent:wxDateEventType() |
    wxDisplayChangedEvent:wxDisplayChangedEventType() |
    wxDropFilesEvent:wxDropFilesEventType() |
    wxEraseEvent:wxEraseEventType() |
    wxFileDirPickerEvent:wxFileDirPickerEventType() |
    wxFocusEvent:wxFocusEventType() |
    wxFontPickerEvent:wxFontPickerEventType() |
    wxGridEvent:wxGridEventType() |
    wxHelpEvent:wxHelpEventType() |
    wxHtmlLinkEvent:wxHtmlLinkEventType() |
    wxIconizeEvent:wxIconizeEventType() |
    wxIdleEvent:wxIdleEventType() |
    wxInitDialogEvent:wxInitDialogEventType() |
    wxJoystickEvent:wxJoystickEventType() |
    wxKeyEvent:wxKeyEventType() |
    wxListEvent:wxListEventType() |
    wxMaximizeEvent:wxMaximizeEventType() |
    wxMenuEvent:wxMenuEventType() |
    wxMouseCaptureChangedEvent:wxMouseCaptureChangedEventType() |
    wxMouseCaptureLostEvent:wxMouseCaptureLostEventType() |
    wxMouseEvent:wxMouseEventType() |
    wxMoveEvent:wxMoveEventType() |
    wxNavigationKeyEvent:wxNavigationKeyEventType() |
    wxPaintEvent:wxPaintEventType() |
    wxPaletteChangedEvent:wxPaletteChangedEventType() |
    wxQueryNewPaletteEvent:wxQueryNewPaletteEventType() |
    wxSashEvent:wxSashEventType() |
    wxScrollEvent:wxScrollEventType() |
    wxScrollWinEvent:wxScrollWinEventType() |
    wxSetCursorEvent:wxSetCursorEventType() |
    wxShowEvent:wxShowEventType() |
    wxSizeEvent:wxSizeEventType() |
    wxSpinEvent:wxSpinEventType() |
    wxSplitterEvent:wxSplitterEventType() |
    wxStyledTextEvent:wxStyledTextEventType() |
    wxSysColourChangedEvent:wxSysColourChangedEventType() |
    wxTaskBarIconEvent:wxTaskBarIconEventType() |
    wxTreeEvent:wxTreeEventType() |
    wxUpdateUIEvent:wxUpdateUIEventType() |
    wxWebViewEvent:wxWebViewEventType() |
    wxWindowCreateEvent:wxWindowCreateEventType() |
    wxWindowDestroyEvent:wxWindowDestroyEventType().

wxFileDirPicker()

-type wxFileDirPicker() :: #wxFileDirPicker{}.

wxFocus()

-type wxFocus() :: #wxFocus{}.

wxFontPicker()

-type wxFontPicker() :: #wxFontPicker{}.

wxGrid()

-type wxGrid() :: #wxGrid{}.

wxHelp()

-type wxHelp() :: #wxHelp{}.

wxHtmlLink()

-type wxHtmlLink() :: #wxHtmlLink{}.

wxIconize()

-type wxIconize() :: #wxIconize{}.

wxIdle()

-type wxIdle() :: #wxIdle{}.

wxInitDialog()

-type wxInitDialog() :: #wxInitDialog{}.

wxJoystick()

-type wxJoystick() :: #wxJoystick{}.

wxKey()

-type wxKey() :: #wxKey{}.

wxList()

-type wxList() :: #wxList{}.

wxMaximize()

-type wxMaximize() :: #wxMaximize{}.

wxMenu()

-type wxMenu() :: #wxMenu{}.

wxMouse()

-type wxMouse() :: #wxMouse{}.

wxMouseCaptureChanged()

-type wxMouseCaptureChanged() :: #wxMouseCaptureChanged{}.

wxMouseCaptureLost()

-type wxMouseCaptureLost() :: #wxMouseCaptureLost{}.

wxMove()

-type wxMove() :: #wxMove{}.

wxNavigationKey()

-type wxNavigationKey() :: #wxNavigationKey{}.

wxPaint()

-type wxPaint() :: #wxPaint{}.

wxPaletteChanged()

-type wxPaletteChanged() :: #wxPaletteChanged{}.

wxQueryNewPalette()

-type wxQueryNewPalette() :: #wxQueryNewPalette{}.

wxSash()

-type wxSash() :: #wxSash{}.

wxScroll()

-type wxScroll() :: #wxScroll{}.

wxScrollWin()

-type wxScrollWin() :: #wxScrollWin{}.

wxSetCursor()

-type wxSetCursor() :: #wxSetCursor{}.

wxShow()

-type wxShow() :: #wxShow{}.

wxSize()

-type wxSize() :: #wxSize{}.

wxSpin()

-type wxSpin() :: #wxSpin{}.

wxSplitter()

-type wxSplitter() :: #wxSplitter{}.

wxStyledText()

-type wxStyledText() :: #wxStyledText{}.

wxSysColourChanged()

-type wxSysColourChanged() :: #wxSysColourChanged{}.

wxTaskBarIcon()

-type wxTaskBarIcon() :: #wxTaskBarIcon{}.

wxTree()

-type wxTree() :: #wxTree{}.

wxUpdateUI()

-type wxUpdateUI() :: #wxUpdateUI{}.

wxWebView()

-type wxWebView() :: #wxWebView{}.

wxWindowCreate()

-type wxWindowCreate() :: #wxWindowCreate{}.

wxWindowDestroy()

-type wxWindowDestroy() :: #wxWindowDestroy{}.

Functions

batch(Fun)

-spec batch(function()) -> term().

batch(Fun::function()) -> term()

Batches all wx commands used in the fun. Improves performance of the command processing by grabbing the wxWidgets thread so that no event processing will be done before the complete batch of commands is invoked.

See also: foldl/3, foldr/3, foreach/2, map/2.

create_memory(Size)

-spec create_memory(integer()) -> wx_memory().

create_memory(Size::integer()) -> wx_memory()

Creates a memory area (of Size in bytes) which can be used by an external library (i.e. opengl). It is up to the client to keep a reference to this object so it does not get garbage collected by erlang while still in use by the external library.

This is far from erlang's intentional usage and can crash the erlang emulator. Use it carefully.

debug(Debug)

-spec debug(Level | [Level]) -> ok when Level :: none | verbose | trace | driver | integer().

debug(Debug::Level | [Level]) -> ok

Sets debug level. If debug level is 'verbose' or 'trace' each call is printed on console. If Level is 'driver' each allocated object and deletion is printed on the console.

demo()

-spec demo() -> ok | {error, atom()}.

demo() -> ok | {error, atom()}

Starts a wxErlang demo if examples directory exists and is compiled

destroy()

-spec destroy() -> ok.

destroy() -> ok

Stops a wx server.

equal/2

-spec equal(wx_object(), wx_object()) -> boolean().

equal(Wx_ref::wx_object(), X2::wx_object()) -> boolean()

Returns true if both arguments references the same object, false otherwise

foldl(Fun, Acc, List)

-spec foldl(function(), term(), list()) -> term().

foldl(Fun::function(), Acc::term(), List::list()) -> term()

Behaves like lists:foldl/3 but batches wx commands. See batch/1.

foldr(Fun, Acc, List)

-spec foldr(function(), term(), list()) -> term().

foldr(Fun::function(), Acc::term(), List::list()) -> term()

Behaves like lists:foldr/3 but batches wx commands. See batch/1.

foreach(Fun, List)

-spec foreach(function(), list()) -> ok.

foreach(Fun::function(), List::list()) -> ok

Behaves like lists:foreach/2 but batches wx commands. See batch/1.

get_env()

-spec get_env() -> wx_env().

get_env() -> wx_env()

Gets this process's current wx environment. Can be sent to other processes to allow them use this process wx environment.

See also: set_env/1.

get_memory_bin/1

-spec get_memory_bin(wx_memory()) -> binary().

get_memory_bin(Wx_mem::wx_memory()) -> binary()

Returns the memory area as a binary.

getObjectType/1

-spec getObjectType(wx_object()) -> atom().

getObjectType(Wx_ref::wx_object()) -> atom()

Returns the object type

is_null/1

-spec is_null(wx_object()) -> boolean().

is_null(Wx_ref::wx_object()) -> boolean()

Returns true if object is null, false otherwise

map(Fun, List)

-spec map(function(), list()) -> list().

map(Fun::function(), List::list()) -> list()

Behaves like lists:map/2 but batches wx commands. See batch/1.

new()

-spec new() -> wx_object().

new() -> wx_object()

Starts a wx server.

new(Options)

-spec new([Option]) -> wx_object() when Option :: {debug, list() | atom()} | {silent_start, boolean()}.

new(Options::[Option]) -> wx_object()

Starts a wx server. Option may be {debug, Level}, see debug/1. Or {silent_start, Bool}, which causes error messages at startup to be suppressed. The latter can be used as a silent test of whether wx is properly installed or not.

null()

-spec null() -> wx_object().

null() -> wx_object()

Returns the null object

parent_class/1

parent_class(X1) -> term()

release_memory/1

-spec release_memory(wx_memory()) -> ok.

release_memory(Wx_mem::wx_memory()) -> ok

retain_memory/1

-spec retain_memory(wx_memory()) -> ok.

retain_memory(Wx_mem::wx_memory()) -> ok

Saves the memory from deletion until release_memory/1 is called. If release_memory/1 is not called the memory will not be garbage collected.

set_env/1

-spec set_env(wx_env()) -> ok.

set_env(Wx_env::wx_env()) -> ok

Sets the process wx environment, allows this process to use another process wx environment.

subscribe_events()

-spec subscribe_events() -> ok.

subscribe_events() -> ok

Adds the calling process to the list of of processes that are listening to wx application events.

At the moment these are all MacOSX specific events corresponding to MacNewFile() and friends from wxWidgets wxApp:

{new_file, ""}

{open_file, Filename}

{print_file, Filename}

{open_url, Url}

{reopen_app, ""}

The call always returns ok but will have sent any already received events to the calling process.

typeCast/2

-spec typeCast(wx_object(), atom()) -> wx_object().

typeCast(Old::wx_object(), NewType::atom()) -> wx_object()

Casts the object to class NewType. It is needed when using functions like wxWindow:findWindow/2, which returns a generic wxObject type.