wxAuiManagerEvent (wx v2.4.2)

Functions for wxAuiManagerEvent class

Event used to indicate various actions taken with wxAuiManager.

See wxAuiManager for available event types.

See: wxAuiManager, wxAuiPaneInfo

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxAuiManagerEvent

Events

Use wxEvtHandler:connect/3 with wxAuiManagerEventType to subscribe to events of this type.

Types

wxAuiManager()

-type wxAuiManager() ::
    #wxAuiManager{type :: wxAuiManagerEvent:wxAuiManagerEventType(),
                  manager :: wxAuiManager:wxAuiManager(),
                  pane :: wxAuiPaneInfo:wxAuiPaneInfo(),
                  button :: integer(),
                  veto_flag :: boolean(),
                  canveto_flag :: boolean(),
                  dc :: wxDC:wxDC()}.

wxAuiManagerEvent()

-type wxAuiManagerEvent() :: wx:wx_object().

wxAuiManagerEventType()

-type wxAuiManagerEventType() ::
    aui_pane_button | aui_pane_close | aui_pane_maximize | aui_pane_restore | aui_pane_activated |
    aui_render | aui_find_manager.

Functions

canVeto(This)

-spec canVeto(This) -> boolean() when This :: wxAuiManagerEvent().

Return: true if this event can be vetoed.

See: veto/2

getButton(This)

-spec getButton(This) -> integer() when This :: wxAuiManagerEvent().

Return: The ID of the button that was clicked.

getDC(This)

-spec getDC(This) -> wxDC:wxDC() when This :: wxAuiManagerEvent().

getManager(This)

-spec getManager(This) -> wxAuiManager:wxAuiManager() when This :: wxAuiManagerEvent().

Return: The wxAuiManager this event is associated with.

getPane(This)

-spec getPane(This) -> wxAuiPaneInfo:wxAuiPaneInfo() when This :: wxAuiManagerEvent().

Return: The pane this event is associated with.

getVeto(This)

-spec getVeto(This) -> boolean() when This :: wxAuiManagerEvent().

Return: true if this event was vetoed.

See: veto/2

setButton(This, Button)

-spec setButton(This, Button) -> ok when This :: wxAuiManagerEvent(), Button :: integer().

Sets the ID of the button clicked that triggered this event.

setCanVeto(This, Can_veto)

-spec setCanVeto(This, Can_veto) -> ok when This :: wxAuiManagerEvent(), Can_veto :: boolean().

Sets whether or not this event can be vetoed.

setDC(This, Pdc)

-spec setDC(This, Pdc) -> ok when This :: wxAuiManagerEvent(), Pdc :: wxDC:wxDC().

setManager(This, Manager)

-spec setManager(This, Manager) -> ok
              when This :: wxAuiManagerEvent(), Manager :: wxAuiManager:wxAuiManager().

Sets the wxAuiManager this event is associated with.

setPane(This, Pane)

-spec setPane(This, Pane) -> ok when This :: wxAuiManagerEvent(), Pane :: wxAuiPaneInfo:wxAuiPaneInfo().

Sets the pane this event is associated with.

veto(This)

-spec veto(This) -> ok when This :: wxAuiManagerEvent().

veto/2

-spec veto(This, [Option]) -> ok when This :: wxAuiManagerEvent(), Option :: {veto, boolean()}.

Cancels the action indicated by this event if canVeto/1 is true.