wxAuiNotebookEvent (wx v2.4.3)

This class is used by the events generated by wxAuiNotebook.

See:

wxAuiNotebook

wxBookCtrlEvent

This class is derived, and can use functions, from:

wxBookCtrlEvent

wxNotifyEvent

wxCommandEvent

wxEvent

wxWidgets docs: wxAuiNotebookEvent

Events

Use wxEvtHandler:connect/3 with wxAuiNotebookEventType to subscribe to events of this type.

Types

wxAuiNotebook()

-type wxAuiNotebook() ::
          #wxAuiNotebook{type :: wxAuiNotebookEvent:wxAuiNotebookEventType(),
                         old_selection :: integer(),
                         selection :: integer(),
                         drag_source :: wxAuiNotebook:wxAuiNotebook()}.

wxAuiNotebookEvent()

-type wxAuiNotebookEvent() :: wx:wx_object().

wxAuiNotebookEventType()

-type wxAuiNotebookEventType() ::
          command_auinotebook_page_close | command_auinotebook_page_changed |
          command_auinotebook_page_changing | command_auinotebook_button |
          command_auinotebook_begin_drag | command_auinotebook_end_drag |
          command_auinotebook_drag_motion | command_auinotebook_allow_dnd |
          command_auinotebook_tab_middle_down | command_auinotebook_tab_middle_up |
          command_auinotebook_tab_right_down | command_auinotebook_tab_right_up |
          command_auinotebook_page_closed | command_auinotebook_drag_done |
          command_auinotebook_bg_dclick.

Functions

getDragSource(This)

-spec getDragSource(This) -> wxAuiNotebook:wxAuiNotebook() when This :: wxAuiNotebookEvent().

getOldSelection(This)

-spec getOldSelection(This) -> integer() when This :: wxAuiNotebookEvent().

Returns the page that was selected before the change, wxNOT\_FOUND if none was selected.

getSelection(This)

-spec getSelection(This) -> integer() when This :: wxAuiNotebookEvent().

Returns the currently selected page, or wxNOT\_FOUND if none was selected.

Note: under Windows, getSelection/1 will return the same value as getOldSelection/1 when called from the EVT_BOOKCTRL_PAGE_CHANGING handler and not the page which is going to be selected.

setDragSource(This, S)

-spec setDragSource(This, S) -> ok when This :: wxAuiNotebookEvent(), S :: wxAuiNotebook:wxAuiNotebook().

setOldSelection(This, Page)

-spec setOldSelection(This, Page) -> ok when This :: wxAuiNotebookEvent(), Page :: integer().

Sets the id of the page selected before the change.

setSelection(This, Page)

-spec setSelection(This, Page) -> ok when This :: wxAuiNotebookEvent(), Page :: integer().

Sets the selection member variable.