wxAuiSimpleTabArt (wx v2.4.2)

Functions for wxAuiSimpleTabArt class

Another standard tab art provider for wxAuiNotebook.

wxAuiSimpleTabArt is derived from wxAuiTabArt demonstrating how to write a completely new tab art class. It can also be used as alternative to wxAuiDefaultTabArt (not implemented in wx).

This class is derived (and can use functions) from: wxAuiTabArt

wxWidgets docs: wxAuiSimpleTabArt

Types

wxAuiSimpleTabArt()

-type wxAuiSimpleTabArt() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxAuiSimpleTabArt()) -> ok.

Destroys the object.

new()

-spec new() -> wxAuiSimpleTabArt().