wxAuiTabArt (wx v2.3.1)

Functions for wxAuiTabArt class

Tab art provider defines all the drawing functions used by wxAuiNotebook.

This allows the wxAuiNotebook to have a pluggable look-and-feel.

By default, a wxAuiNotebook uses an instance of this class called wxAuiDefaultTabArt (not implemented in wx) which provides bitmap art and a colour scheme that is adapted to the major platforms' look. You can either derive from that class to alter its behaviour or write a completely new tab art class.

Another example of creating a new wxAuiNotebook tab bar is wxAuiSimpleTabArt.

Call wxAuiNotebook:setArtProvider/2 to make use of this new tab art.

wxWidgets docs: wxAuiTabArt

Types

wxAuiTabArt()

-type wxAuiTabArt() :: wx:wx_object().

Functions

setActiveColour/2

-spec setActiveColour(This, Colour) -> ok when This :: wxAuiTabArt(), Colour :: wx:wx_colour().

Sets the colour of the selected tab.

Since: 2.9.2

setColour/2

-spec setColour(This, Colour) -> ok when This :: wxAuiTabArt(), Colour :: wx:wx_colour().

Sets the colour of the inactive tabs.

Since: 2.9.2

setFlags/2

-spec setFlags(This, Flags) -> ok when This :: wxAuiTabArt(), Flags :: integer().

Sets flags.

setMeasuringFont/2

-spec setMeasuringFont(This, Font) -> ok when This :: wxAuiTabArt(), Font :: wxFont:wxFont().

Sets the font used for calculating measurements.

setNormalFont/2

-spec setNormalFont(This, Font) -> ok when This :: wxAuiTabArt(), Font :: wxFont:wxFont().

Sets the normal font for drawing labels.

setSelectedFont/2

-spec setSelectedFont(This, Font) -> ok when This :: wxAuiTabArt(), Font :: wxFont:wxFont().

Sets the font for drawing text for selected UI elements.