wxBookCtrlEvent (wx v2.4.2)

Functions for wxBookCtrlEvent class

This class represents the events generated by book controls (wxNotebook, wxListbook, wxChoicebook, wxTreebook, wxAuiNotebook).

The PAGE_CHANGING events are sent before the current page is changed. It allows the program to examine the current page (which can be retrieved with getOldSelection/1) and to veto the page change by calling wxNotifyEvent:veto/1 if, for example, the current values in the controls of the old page are invalid.

The PAGE_CHANGED events are sent after the page has been changed and the program cannot veto it any more, it just informs it about the page change.

To summarize, if the program is interested in validating the page values before allowing the user to change it, it should process the PAGE_CHANGING event, otherwise PAGE_CHANGED is probably enough. In any case, it is probably unnecessary to process both events at once.

See: wxNotebook, wxListbook, wxChoicebook, wxTreebook, wxToolbook, wxAuiNotebook

This class is derived (and can use functions) from: wxNotifyEvent wxCommandEvent wxEvent

wxWidgets docs: wxBookCtrlEvent

Types

wxBookCtrl()

-type wxBookCtrl() ::
    #wxBookCtrl{type :: wxBookCtrlEvent:wxBookCtrlEventType(),
                nSel :: integer(),
                nOldSel :: integer()}.

wxBookCtrlEvent()

-type wxBookCtrlEvent() :: wx:wx_object().

wxBookCtrlEventType()

-type wxBookCtrlEventType() ::
    command_notebook_page_changed | command_notebook_page_changing | choicebook_page_changed |
    choicebook_page_changing | treebook_page_changed | treebook_page_changing |
    toolbook_page_changed | toolbook_page_changing | listbook_page_changed |
    listbook_page_changing.

Functions

getOldSelection(This)

-spec getOldSelection(This) -> integer() when This :: wxBookCtrlEvent().

Returns the page that was selected before the change, wxNOT_FOUND if none was selected.

getSelection(This)

-spec getSelection(This) -> integer() when This :: wxBookCtrlEvent().

Returns the currently selected page, or wxNOT_FOUND if none was selected.

Note: under Windows, getSelection/1 will return the same value as getOldSelection/1 when called from the EVT_BOOKCTRL_PAGE_CHANGING handler and not the page which is going to be selected.

setOldSelection(This, Page)

-spec setOldSelection(This, Page) -> ok when This :: wxBookCtrlEvent(), Page :: integer().

Sets the id of the page selected before the change.

setSelection(This, Page)

-spec setSelection(This, Page) -> ok when This :: wxBookCtrlEvent(), Page :: integer().

Sets the selection member variable.