wxBufferedPaintDC (wx v2.4.3)

This is a subclass of wxBufferedDC which can be used inside of an EVT\_PAINT() event handler to achieve double-buffered drawing.

Just use this class instead of wxPaintDC and make sure wxWindow:setBackgroundStyle/2 is called with wxBG_STYLE_PAINT somewhere in the class initialization code, and that's all you have to do to (mostly) avoid flicker. The only thing to watch out for is that if you are using this class together with wxScrolled (not implemented in wx), you probably do not want to call wxScrolledWindow:prepareDC/2 on it as it already does this internally for the real underlying wxPaintDC.

See:

wxDC

wxBufferedDC

wxPaintDC

This class is derived, and can use functions, from:

wxBufferedDC

wxMemoryDC

wxDC

wxWidgets docs: wxBufferedPaintDC

Types

wxBufferedPaintDC()

-type wxBufferedPaintDC() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxBufferedPaintDC()) -> ok.

Destroys the object

new(Window)

-spec new(Window) -> wxBufferedPaintDC() when Window :: wxWindow:wxWindow().

Equivalent to new(Window, []).

new/2

-spec new(Window, Buffer) -> wxBufferedPaintDC()
             when Window :: wxWindow:wxWindow(), Buffer :: wxBitmap:wxBitmap();
         (Window, [Option]) -> wxBufferedPaintDC()
             when Window :: wxWindow:wxWindow(), Option :: {style, integer()}.

new/3

-spec new(Window, Buffer, [Option]) -> wxBufferedPaintDC()
             when
                 Window :: wxWindow:wxWindow(),
                 Buffer :: wxBitmap:wxBitmap(),
                 Option :: {style, integer()}.

As with wxBufferedDC, you may either provide the bitmap to be used for buffering or let this object create one internally (in the latter case, the size of the client part of the window is used).

Pass wxBUFFER_CLIENT_AREA for the style parameter to indicate that just the client area of the window is buffered, or wxBUFFER_VIRTUAL_AREA to indicate that the buffer bitmap covers the virtual area.