wxCalendarDateAttr (wx v2.4.2)

Functions for wxCalendarDateAttr class

wxCalendarDateAttr is a custom attributes for a calendar date. The objects of this class are used with wxCalendarCtrl.

See: wxCalendarCtrl

wxWidgets docs: wxCalendarDateAttr

Types

wxCalendarDateAttr()

-type wxCalendarDateAttr() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxCalendarDateAttr()) -> ok.

Destroys the object.

getBackgroundColour(This)

-spec getBackgroundColour(This) -> wx:wx_colour4() when This :: wxCalendarDateAttr().

Returns the background colour set for the calendar date.

getBorder(This)

-spec getBorder(This) -> wx:wx_enum() when This :: wxCalendarDateAttr().

Returns the border set for the calendar date.

getBorderColour(This)

-spec getBorderColour(This) -> wx:wx_colour4() when This :: wxCalendarDateAttr().

Returns the border colour set for the calendar date.

getFont(This)

-spec getFont(This) -> wxFont:wxFont() when This :: wxCalendarDateAttr().

Returns the font set for the calendar date.

getTextColour(This)

-spec getTextColour(This) -> wx:wx_colour4() when This :: wxCalendarDateAttr().

Returns the text colour set for the calendar date.

hasBackgroundColour(This)

-spec hasBackgroundColour(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default text background colour is set.

hasBorder(This)

-spec hasBorder(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default (i.e. any) border is set.

hasBorderColour(This)

-spec hasBorderColour(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default border colour is set.

hasFont(This)

-spec hasFont(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default font is set.

hasTextColour(This)

-spec hasTextColour(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default text foreground colour is set.

isHoliday(This)

-spec isHoliday(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if this calendar day is displayed as a holiday.

new()

-spec new() -> wxCalendarDateAttr().

new/1

-spec new(Border) -> wxCalendarDateAttr() when Border :: wx:wx_enum();
   ([Option]) -> wxCalendarDateAttr()
       when
           Option ::
               {colText, wx:wx_colour()} |
               {colBack, wx:wx_colour()} |
               {colBorder, wx:wx_colour()} |
               {font, wxFont:wxFont()} |
               {border, wx:wx_enum()}.

Constructor for specifying all wxCalendarDateAttr properties.

new(Border, Options)

-spec new(Border, [Option]) -> wxCalendarDateAttr()
       when Border :: wx:wx_enum(), Option :: {colBorder, wx:wx_colour()}.

Constructor using default properties except the given border.

setBackgroundColour(This, ColBack)

-spec setBackgroundColour(This, ColBack) -> ok
                       when This :: wxCalendarDateAttr(), ColBack :: wx:wx_colour().

Sets the text background colour to use.

setBorder(This, Border)

-spec setBorder(This, Border) -> ok when This :: wxCalendarDateAttr(), Border :: wx:wx_enum().

Sets the border to use.

setBorderColour(This, Col)

-spec setBorderColour(This, Col) -> ok when This :: wxCalendarDateAttr(), Col :: wx:wx_colour().

Sets the border colour to use.

setFont(This, Font)

-spec setFont(This, Font) -> ok when This :: wxCalendarDateAttr(), Font :: wxFont:wxFont().

Sets the font to use.

setHoliday(This, Holiday)

-spec setHoliday(This, Holiday) -> ok when This :: wxCalendarDateAttr(), Holiday :: boolean().

If holiday is true, this calendar day will be displayed as a holiday.

setTextColour(This, ColText)

-spec setTextColour(This, ColText) -> ok when This :: wxCalendarDateAttr(), ColText :: wx:wx_colour().

Sets the text (foreground) colour to use.