wxCalendarEvent (wx v2.0.1)

The wxCalendarEvent class is used together with wxCalendarCtrl.

See: wxCalendarCtrl

This class is derived (and can use functions) from: wxDateEvent wxCommandEvent wxEvent

wxWidgets docs: wxCalendarEvent

Types

-type wxCalendar() :: term().

Specs

wxCalendar() :: #wxCalendar{}.

-type wxCalendarEvent() :: term().

Specs

wxCalendarEvent() :: wx:wx_object().

-type wxCalendarEventType() :: term().

Specs

wxCalendarEventType() ::
    calendar_sel_changed | calendar_day_changed | calendar_month_changed | calendar_year_changed |
    calendar_doubleclicked | calendar_weekday_clicked.

Functions

getDate/1

Specs

getDate(This) -> wx:wx_datetime() when This :: wxCalendarEvent().

Returns the date.

getWeekDay/1

Specs

getWeekDay(This) -> wx:wx_enum() when This :: wxCalendarEvent().

Returns the week day on which the user clicked in EVT_CALENDAR_WEEKDAY_CLICKED handler.

It doesn't make sense to call this function in other handlers.