wxCaret (wx v2.4.2)

Functions for wxCaret class

A caret is a blinking cursor showing the position where the typed text will appear. Text controls usually have their own caret but wxCaret provides a way to use a caret in other windows.

Currently, the caret appears as a rectangle of the given size. In the future, it will be possible to specify a bitmap to be used for the caret shape.

A caret is always associated with a window and the current caret can be retrieved using wxWindow:getCaret/1. The same caret can't be reused in two different windows.

wxWidgets docs: wxCaret

Types

wxCaret()

-type wxCaret() :: wx:wx_object().

Functions

create(This, Window, Size)

-spec create(This, Window, Size) -> boolean()
          when
              This :: wxCaret(),
              Window :: wxWindow:wxWindow(),
              Size :: {W :: integer(), H :: integer()}.

create(This, Window, Width, Height)

-spec create(This, Window, Width, Height) -> boolean()
          when
              This :: wxCaret(),
              Window :: wxWindow:wxWindow(),
              Width :: integer(),
              Height :: integer().

Creates a caret with the given size (in pixels) and associates it with the window (same as the equivalent constructors).

destroy(This)

-spec destroy(This :: wxCaret()) -> ok.

Destroys the object.

getBlinkTime()

-spec getBlinkTime() -> integer().

Returns the blink time which is measured in milliseconds and is the time elapsed between 2 inversions of the caret (blink time of the caret is the same for all carets, so this functions is static).

getPosition(This)

-spec getPosition(This) -> {X :: integer(), Y :: integer()} when This :: wxCaret().

getSize(This)

-spec getSize(This) -> {W :: integer(), H :: integer()} when This :: wxCaret().

getWindow(This)

-spec getWindow(This) -> wxWindow:wxWindow() when This :: wxCaret().

Get the window the caret is associated with.

hide(This)

-spec hide(This) -> ok when This :: wxCaret().

Hides the caret, same as Show(false).

isOk(This)

-spec isOk(This) -> boolean() when This :: wxCaret().

Returns true if the caret was created successfully.

isVisible(This)

-spec isVisible(This) -> boolean() when This :: wxCaret().

Returns true if the caret is visible and false if it is permanently hidden (if it is blinking and not shown currently but will be after the next blink, this method still returns true).

move(This, Pt)

-spec move(This, Pt) -> ok when This :: wxCaret(), Pt :: {X :: integer(), Y :: integer()}.

move(This, X, Y)

-spec move(This, X, Y) -> ok when This :: wxCaret(), X :: integer(), Y :: integer().

Move the caret to given position (in logical coordinates).

new(Window, Size)

-spec new(Window, Size) -> wxCaret()
       when Window :: wxWindow:wxWindow(), Size :: {W :: integer(), H :: integer()}.

new(Window, Width, Height)

-spec new(Window, Width, Height) -> wxCaret()
       when Window :: wxWindow:wxWindow(), Width :: integer(), Height :: integer().

Creates a caret with the given size (in pixels) and associates it with the window.

setBlinkTime(Milliseconds)

-spec setBlinkTime(Milliseconds) -> ok when Milliseconds :: integer().

Sets the blink time for all the carets.

Warning: Under Windows, this function will change the blink time for all carets permanently (until the next time it is called), even for carets in other applications.

See: getBlinkTime/0

setSize(This, Size)

-spec setSize(This, Size) -> ok when This :: wxCaret(), Size :: {W :: integer(), H :: integer()}.

setSize(This, Width, Height)

-spec setSize(This, Width, Height) -> ok when This :: wxCaret(), Width :: integer(), Height :: integer().

Changes the size of the caret.

show(This)

-spec show(This) -> ok when This :: wxCaret().

show/2

-spec show(This, [Option]) -> ok when This :: wxCaret(), Option :: {show, boolean()}.

Shows or hides the caret.

Notice that if the caret was hidden N times, it must be shown N times as well to reappear on the screen.