wxCheckListBox (wx v2.3.1)

Functions for wxCheckListBox class

A wxCheckListBox is like a wxListBox, but allows items to be checked or unchecked.

When using this class under Windows wxWidgets must be compiled with wxUSE_OWNER_DRAWN set to 1.

See: wxListBox, wxChoice, wxComboBox, wxListCtrl, wxCommandEvent

This class is derived (and can use functions) from: wxListBox wxControlWithItems wxControl wxWindow wxEvtHandler

wxWidgets docs: wxCheckListBox

Events

Event types emitted from this class: command_checklistbox_toggled

Types

wxCheckListBox()

-type wxCheckListBox() :: wx:wx_object().

Functions

check/2

-spec check(This, Item) -> ok when This :: wxCheckListBox(), Item :: integer().

check/3

-spec check(This, Item, [Option]) -> ok
         when This :: wxCheckListBox(), Item :: integer(), Option :: {check, boolean()}.

Checks the given item.

Note that calling this method does not result in a wxEVT_CHECKLISTBOX event being emitted.

destroy/1

-spec destroy(This :: wxCheckListBox()) -> ok.

Destructor, destroying the list box.

isChecked/2

-spec isChecked(This, Item) -> boolean() when This :: wxCheckListBox(), Item :: integer().

Returns true if the given item is checked, false otherwise.

new()

-spec new() -> wxCheckListBox().

Default constructor.

new/2

-spec new(Parent, Id) -> wxCheckListBox() when Parent :: wxWindow:wxWindow(), Id :: integer().

new/3

-spec new(Parent, Id, [Option]) -> wxCheckListBox()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Option ::
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {choices, [unicode:chardata()]} |
               {style, integer()} |
               {validator, wx:wx_object()}.

Constructor, creating and showing a list box.