wxChildFocusEvent (wx v2.4.2)

Functions for wxChildFocusEvent class

A child focus event is sent to a (parent-)window when one of its child windows gains focus, so that the window could restore the focus back to its corresponding child if it loses it now and regains later.

Notice that child window is the direct child of the window receiving event. Use wxWindow:findFocus/0 to retrieve the window which is actually getting focus.

See: Overview events

This class is derived (and can use functions) from: wxCommandEvent wxEvent

wxWidgets docs: wxChildFocusEvent

Events

Use wxEvtHandler:connect/3 with wxChildFocusEventType to subscribe to events of this type.

Types

wxChildFocus()

-type wxChildFocus() :: #wxChildFocus{type :: wxChildFocusEvent:wxChildFocusEventType()}.

wxChildFocusEvent()

-type wxChildFocusEvent() :: wx:wx_object().

wxChildFocusEventType()

-type wxChildFocusEventType() :: child_focus.

Functions

getWindow(This)

-spec getWindow(This) -> wxWindow:wxWindow() when This :: wxChildFocusEvent().

Returns the direct child which receives the focus, or a (grand-)parent of the control receiving the focus.

To get the actually focused control use wxWindow:findFocus/0.