wxClipboardTextEvent (wx v2.4.3)

This class represents the events generated by a control (typically a wxTextCtrl but other windows can generate these events as well) when its content gets copied or cut to, or pasted from the clipboard.

There are three types of corresponding events wxEVT_TEXT_COPY, wxEVT_TEXT_CUT and wxEVT_TEXT_PASTE.

If any of these events is processed (without being skipped) by an event handler, the corresponding operation doesn't take place which allows preventing the text from being copied from or pasted to a control. It is also possible to examine the clipboard contents in the PASTE event handler and transform it in some way before inserting in a control - for example, changing its case or removing invalid characters.

Finally notice that a CUT event is always preceded by the COPY event which makes it possible to only process the latter if it doesn't matter if the text was copied or cut.

Note: These events are currently only generated by wxTextCtrl in wxGTK and wxOSX but are also generated by wxComboBox without wxCB_READONLY style in wxMSW.

See: wxClipboard

This class is derived, and can use functions, from:

wxCommandEvent

wxEvent

wxWidgets docs: wxClipboardTextEvent

Events

Use wxEvtHandler:connect/3 with wxClipboardTextEventType to subscribe to events of this type.

Types

wxClipboardText()

-type wxClipboardText() :: #wxClipboardText{type :: wxClipboardTextEvent:wxClipboardTextEventType()}.

wxClipboardTextEvent()

-type wxClipboardTextEvent() :: wx:wx_object().

wxClipboardTextEventType()

-type wxClipboardTextEventType() :: command_text_copy | command_text_cut | command_text_paste.