wxColourData (wx v2.4.2)

Functions for wxColourData class

This class holds a variety of information related to colour dialogs.

See: wx_color(), wxColourDialog, Overview cmndlg

wxWidgets docs: wxColourData

Types

wxColourData()

-type wxColourData() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxColourData()) -> ok.

Destructor.

getChooseFull(This)

-spec getChooseFull(This) -> boolean() when This :: wxColourData().

Under Windows, determines whether the Windows colour dialog will display the full dialog with custom colour selection controls.

Has no meaning under other platforms.

The default value is true.

getColour(This)

-spec getColour(This) -> wx:wx_colour4() when This :: wxColourData().

Gets the current colour associated with the colour dialog.

The default colour is black.

getCustomColour(This, I)

-spec getCustomColour(This, I) -> wx:wx_colour4() when This :: wxColourData(), I :: integer().

Returns custom colours associated with the colour dialog.

new()

-spec new() -> wxColourData().

Constructor.

Initializes the custom colours to wxNullColour, the data colour setting to black, and the choose full setting to true.

setChooseFull(This, Flag)

-spec setChooseFull(This, Flag) -> ok when This :: wxColourData(), Flag :: boolean().

Under Windows, tells the Windows colour dialog to display the full dialog with custom colour selection controls.

Under other platforms, has no effect.

The default value is true.

setColour(This, Colour)

-spec setColour(This, Colour) -> ok when This :: wxColourData(), Colour :: wx:wx_colour().

Sets the default colour for the colour dialog.

The default colour is black.

setCustomColour(This, I, Colour)

-spec setCustomColour(This, I, Colour) -> ok
                   when This :: wxColourData(), I :: integer(), Colour :: wx:wx_colour().

Sets custom colours for the colour dialog.