wxDateEvent (wx v2.0.1)

This event class holds information about a date change and is used together with wxDatePickerCtrl. It also serves as a base class for wxCalendarEvent.

This class is derived (and can use functions) from: wxCommandEvent wxEvent

wxWidgets docs: wxDateEvent

Types

-type wxDate() :: term().

Specs

wxDate() :: #wxDate{}.

-type wxDateEvent() :: term().

Specs

wxDateEvent() :: wx:wx_object().

-type wxDateEventType() :: term().

Specs

wxDateEventType() :: date_changed.

Functions

getDate/1

Specs

getDate(This) -> wx:wx_datetime() when This :: wxDateEvent().

Returns the date.