wxEraseEvent (wx v2.5.2-rc0)

An erase event is sent when a window's background needs to be repainted.

On some platforms, such as GTK+, this event is simulated (simply generated just before the paint event) and may cause flicker. It is therefore recommended that you set the text background colour explicitly in order to prevent flicker. The default background colour under GTK+ is grey.

To intercept this event, use the EVT_ERASE_BACKGROUND macro in an event table definition.

You must use the device context returned by getDC/1 to draw on, don't create a wxPaintDC in the event handler.

See: Overview events

This class is derived, and can use functions, from:

wxEvent

wxWidgets docs: wxEraseEvent

Events

Use wxEvtHandler:connect/3 with wxEraseEventType to subscribe to events of this type.

Types

wxErase()

-type wxErase() :: #wxErase{type :: wxEraseEvent:wxEraseEventType(), dc :: wxDC:wxDC()}.

wxEraseEvent()

-type wxEraseEvent() :: wx:wx_object().

wxEraseEventType()

-type wxEraseEventType() :: erase_background.

Functions

getDC(This)

-spec getDC(This) -> wxDC:wxDC() when This :: wxEraseEvent().

Returns the device context associated with the erase event to draw on.

The returned pointer is never NULL.