wxFileDataObject (wx v2.0.1)

wxFileDataObject is a specialization of wxDataObject for file names. The program works with it just as if it were a list of absolute file names, but internally it uses the same format as Explorer and other compatible programs under Windows or GNOME/KDE file manager under Unix which makes it possible to receive files from them using this class.

See: wxDataObject, wxDataObjectSimple (not implemented in wx), wxTextDataObject, wxBitmapDataObject, wxDataObject

This class is derived (and can use functions) from: wxDataObject

wxWidgets docs: wxFileDataObject

Types

-type wxFileDataObject() :: term().

Specs

wxFileDataObject() :: wx:wx_object().

Functions

addFile/2

Specs

addFile(This, File) -> ok when This :: wxFileDataObject(), File :: unicode:chardata().

Adds a file to the file list represented by this data object (Windows only).

destroy/1

Specs

destroy(This :: wxFileDataObject()) -> ok.

Destroys the object.

getFilenames/1

Specs

getFilenames(This) -> [unicode:charlist()] when This :: wxFileDataObject().

Returns the array of file names.

new/0

Specs

new() -> wxFileDataObject().

Constructor.