wxFilePickerCtrl (wx v2.0.1)

This control allows the user to select a file. The generic implementation is a button which brings up a wxFileDialog when clicked. Native implementation may differ but this is usually a (small) widget which give access to the file-chooser dialog. It is only available if wxUSE_FILEPICKERCTRL is set to 1 (the default).

Styles

This class supports the following styles:

See: wxFileDialog, wxFileDirPickerEvent

This class is derived (and can use functions) from: wxPickerBase wxControl wxWindow wxEvtHandler

wxWidgets docs: wxFilePickerCtrl

Events

Event types emitted from this class: command_filepicker_changed

Types

-type wxFilePickerCtrl() :: term().

Specs

wxFilePickerCtrl() :: wx:wx_object().

Functions

create/3

Specs

create(This, Parent, Id) -> boolean()
          when This :: wxFilePickerCtrl(), Parent :: wxWindow:wxWindow(), Id :: integer().

create/4

Specs

create(This, Parent, Id, [Option]) -> boolean()
          when
              This :: wxFilePickerCtrl(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Option ::
                  {path, unicode:chardata()} |
                  {message, unicode:chardata()} |
                  {wildcard, unicode:chardata()} |
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()} |
                  {validator, wx:wx_object()}.

Creates this widget with the given parameters.

Return: true if the control was successfully created or false if creation failed.

destroy/1

Specs

destroy(This :: wxFilePickerCtrl()) -> ok.

Destroys the object.

getPath/1

Specs

getPath(This) -> unicode:charlist() when This :: wxFilePickerCtrl().

Returns the absolute path of the currently selected file.

new/0

Specs

new() -> wxFilePickerCtrl().

new/2

Specs

new(Parent, Id) -> wxFilePickerCtrl() when Parent :: wxWindow:wxWindow(), Id :: integer().

new/3

Specs

new(Parent, Id, [Option]) -> wxFilePickerCtrl()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Option ::
               {path, unicode:chardata()} |
               {message, unicode:chardata()} |
               {wildcard, unicode:chardata()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()} |
               {validator, wx:wx_object()}.

Initializes the object and calls create/4 with all the parameters.

setPath/2

Specs

setPath(This, Filename) -> ok when This :: wxFilePickerCtrl(), Filename :: unicode:chardata().

Sets the absolute path of the currently selected file.

If the control uses wxFLP_FILE_MUST_EXIST and does not use wxFLP_USE_TEXTCTRL style, the filename must be a name of an existing file and will be simply ignored by the native wxGTK implementation if this is not the case (the generic implementation used under the other platforms accepts even invalid file names currently, but this is subject to change in the future, don't rely on being able to use non-existent paths with it).