wxFontPickerCtrl (wx v2.3.1)

Functions for wxFontPickerCtrl class

This control allows the user to select a font. The generic implementation is a button which brings up a wxFontDialog when clicked. Native implementation may differ but this is usually a (small) widget which give access to the font-chooser dialog. It is only available if wxUSE_FONTPICKERCTRL is set to 1 (the default).

Styles

This class supports the following styles:

See: wxFontDialog, wxFontPickerEvent

This class is derived (and can use functions) from: wxPickerBase wxControl wxWindow wxEvtHandler

wxWidgets docs: wxFontPickerCtrl

Events

Event types emitted from this class: command_fontpicker_changed

Types

wxFontPickerCtrl()

-type wxFontPickerCtrl() :: wx:wx_object().

Functions

create/3

-spec create(This, Parent, Id) -> boolean()
          when This :: wxFontPickerCtrl(), Parent :: wxWindow:wxWindow(), Id :: integer().

create/4

-spec create(This, Parent, Id, [Option]) -> boolean()
          when
              This :: wxFontPickerCtrl(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Option ::
                  {initial, wxFont:wxFont()} |
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()} |
                  {validator, wx:wx_object()}.

Creates this widget with given parameters.

Return: true if the control was successfully created or false if creation failed.

destroy/1

-spec destroy(This :: wxFontPickerCtrl()) -> ok.

Destroys the object.

getMaxPointSize/1

-spec getMaxPointSize(This) -> integer() when This :: wxFontPickerCtrl().

Returns the maximum point size value allowed for the user-chosen font.

getSelectedFont/1

-spec getSelectedFont(This) -> wxFont:wxFont() when This :: wxFontPickerCtrl().

Returns the currently selected font.

Note that this function is completely different from wxWindow:getFont/1.

new/0

-spec new() -> wxFontPickerCtrl().

new/2

-spec new(Parent, Id) -> wxFontPickerCtrl() when Parent :: wxWindow:wxWindow(), Id :: integer().

new/3

-spec new(Parent, Id, [Option]) -> wxFontPickerCtrl()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Option ::
               {initial, wxFont:wxFont()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()} |
               {validator, wx:wx_object()}.

Initializes the object and calls create/4 with all the parameters.

setMaxPointSize/2

-spec setMaxPointSize(This, Max) -> ok when This :: wxFontPickerCtrl(), Max :: integer().

Sets the maximum point size value allowed for the user-chosen font.

The default value is 100. Note that big fonts can require a lot of memory and CPU time both for creation and for rendering; thus, specially because the user has the option to specify the fontsize through a text control (see wxFNTP_USE_TEXTCTRL), it's a good idea to put a limit to the maximum font size when huge fonts do not make much sense.

setSelectedFont/2

-spec setSelectedFont(This, Font) -> ok when This :: wxFontPickerCtrl(), Font :: wxFont:wxFont().

Sets the currently selected font.

Note that this function is completely different from wxWindow:setFont/2.