wxGraphicsGradientStops (wx v2.6-rc0)

Represents a collection of wxGraphicGradientStop values for use with CreateLinearGradientBrush and CreateRadialGradientBrush.

The stops are maintained in order of position. If two or more stops are added with the same position then the one(s) added later come later. This can be useful for producing discontinuities in the colour gradient.

Notice that this class is write-once, you can't modify the stops once they had been added.

Since: 2.9.1

wxWidgets docs: wxGraphicsGradientStops

Types

wxGraphicsGradientStops()

-type wxGraphicsGradientStops() :: wx:wx_object().

Functions

add(This, Col, Pos)

-spec add(This, Col, Pos) -> ok
             when This :: wxGraphicsGradientStops(), Col :: wx:wx_colour(), Pos :: number().

Add a new stop.

destroy(This)

-spec destroy(This :: wxGraphicsGradientStops()) -> ok.

Destroys the object

getCount(This)

-spec getCount(This) -> integer() when This :: wxGraphicsGradientStops().

Returns the number of stops.

getEndColour(This)

-spec getEndColour(This) -> wx:wx_colour4() when This :: wxGraphicsGradientStops().

Returns the end colour.

getStartColour(This)

-spec getStartColour(This) -> wx:wx_colour4() when This :: wxGraphicsGradientStops().

Returns the start colour.

item(This, N)

-spec item(This, N) -> {wx:wx_colour4(), float()} when This :: wxGraphicsGradientStops(), N :: integer().

Returns the stop at the given index.

new()

-spec new() -> wxGraphicsGradientStops().

Equivalent to new([]).

new(Options)

-spec new([Option]) -> wxGraphicsGradientStops()
             when Option :: {startCol, wx:wx_colour()} | {endCol, wx:wx_colour()}.

Initializes the gradient stops with the given boundary colours.

Creates a wxGraphicsGradientStops instance with start colour given by startCol and end colour given by endCol.

setEndColour(This, Col)

-spec setEndColour(This, Col) -> ok when This :: wxGraphicsGradientStops(), Col :: wx:wx_colour().

Set the end colour to col.

setStartColour(This, Col)

-spec setStartColour(This, Col) -> ok when This :: wxGraphicsGradientStops(), Col :: wx:wx_colour().

Set the start colour to col.