wxGridCellBoolEditor (wx v2.5.2)

Grid cell editor for boolean data.

See:

wxGridCellEditor

wxGridCellChoiceEditor

wxGridCellFloatEditor

wxGridCellNumberEditor

wxGridCellTextEditor

This class is derived, and can use functions, from:

wxGridCellEditor

wxWidgets docs: wxGridCellBoolEditor

Types

wxGridCellBoolEditor()

-type wxGridCellBoolEditor() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxGridCellBoolEditor()) -> ok.

Destroys the object

isTrueValue(Value)

-spec isTrueValue(Value) -> boolean() when Value :: unicode:chardata().

Returns true if the given value is equal to the string representation of the truth value we currently use (see useStringValues/1).

new()

-spec new() -> wxGridCellBoolEditor().

Default constructor.

useStringValues()

-spec useStringValues() -> ok.

Equivalent to useStringValues([]).

useStringValues(Options)

-spec useStringValues([Option]) -> ok
                         when
                             Option ::
                                 {valueTrue, unicode:chardata()} | {valueFalse, unicode:chardata()}.

This method allows you to customize the values returned by wxGridCellNumberEditor:getValue/1 for the cell using this editor.

By default, the default values of the arguments are used, i.e. "1" is returned if the cell is checked and an empty string otherwise.