wxGridCellChoiceEditor (wx v2.5)

Grid cell editor for string data providing the user a choice from a list of strings.

See:

wxGridCellEditor

wxGridCellBoolEditor

wxGridCellFloatEditor

wxGridCellNumberEditor

wxGridCellTextEditor

This class is derived, and can use functions, from:

wxGridCellEditor

wxWidgets docs: wxGridCellChoiceEditor

Types

wxGridCellChoiceEditor()

-type wxGridCellChoiceEditor() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxGridCellChoiceEditor()) -> ok.

Destroys the object

new(Choices)

-spec new(Choices) -> wxGridCellChoiceEditor() when Choices :: [unicode:chardata()].

Equivalent to new(Choices, []).

new(Choices, Options)

-spec new(Choices, [Option]) -> wxGridCellChoiceEditor()
             when Choices :: [unicode:chardata()], Option :: {allowOthers, boolean()}.

Choice cell renderer ctor.

setParameters(This, Params)

-spec setParameters(This, Params) -> ok
                       when This :: wxGridCellChoiceEditor(), Params :: unicode:chardata().

Parameters string format is "item1[,item2[...,itemN]]".

This method can be called before the editor is used for the first time, or later, in which case it replaces the previously specified strings with the new ones.