wxGridSizer (wx v2.4.3)

A grid sizer is a sizer which lays out its children in a two-dimensional table with all table fields having the same size, i.e.

the width of each field is the width of the widest child, the height of each field is the height of the tallest child.

See:

wxSizer

Overview sizer

This class is derived, and can use functions, from:

wxSizer

wxWidgets docs: wxGridSizer

Types

wxGridSizer()

-type wxGridSizer() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxGridSizer()) -> ok.

Destroys the object

getCols(This)

-spec getCols(This) -> integer() when This :: wxGridSizer().

Returns the number of columns that has been specified for the sizer.

Returns zero if the sizer is automatically adjusting the number of columns depending on number of its children. To get the effective number of columns or rows being currently used, see GetEffectiveColsCount() (not implemented in wx)

getHGap(This)

-spec getHGap(This) -> integer() when This :: wxGridSizer().

Returns the horizontal gap (in pixels) between cells in the sizer.

getRows(This)

-spec getRows(This) -> integer() when This :: wxGridSizer().

Returns the number of rows that has been specified for the sizer.

Returns zero if the sizer is automatically adjusting the number of rows depending on number of its children. To get the effective number of columns or rows being currently used, see GetEffectiveRowsCount() (not implemented in wx).

getVGap(This)

-spec getVGap(This) -> integer() when This :: wxGridSizer().

Returns the vertical gap (in pixels) between the cells in the sizer.

new(Cols)

-spec new(Cols) -> wxGridSizer() when Cols :: integer().

Equivalent to new(Cols, []).

new(Cols, Options)

-spec new(Cols, [Option]) -> wxGridSizer()
             when Cols :: integer(), Option :: {gap, {W :: integer(), H :: integer()}}.

new/3

-spec new(Cols, Vgap, Hgap) -> wxGridSizer()
             when Cols :: integer(), Vgap :: integer(), Hgap :: integer();
         (Rows, Cols, Gap) -> wxGridSizer()
             when Rows :: integer(), Cols :: integer(), Gap :: {W :: integer(), H :: integer()}.

new(Rows, Cols, Vgap, Hgap)

-spec new(Rows, Cols, Vgap, Hgap) -> wxGridSizer()
             when Rows :: integer(), Cols :: integer(), Vgap :: integer(), Hgap :: integer().

setCols(This, Cols)

-spec setCols(This, Cols) -> ok when This :: wxGridSizer(), Cols :: integer().

Sets the number of columns in the sizer.

setHGap(This, Gap)

-spec setHGap(This, Gap) -> ok when This :: wxGridSizer(), Gap :: integer().

Sets the horizontal gap (in pixels) between cells in the sizer.

setRows(This, Rows)

-spec setRows(This, Rows) -> ok when This :: wxGridSizer(), Rows :: integer().

Sets the number of rows in the sizer.

setVGap(This, Gap)

-spec setVGap(This, Gap) -> ok when This :: wxGridSizer(), Gap :: integer().

Sets the vertical gap (in pixels) between the cells in the sizer.