wxHtmlLinkEvent (wx v2.4.2)

Functions for wxHtmlLinkEvent class

This event class is used for the events generated by wxHtmlWindow.

This class is derived (and can use functions) from: wxCommandEvent wxEvent

wxWidgets docs: wxHtmlLinkEvent

Events

Use wxEvtHandler:connect/3 with wxHtmlLinkEventType to subscribe to events of this type.

Types

wxHtmlLink()

-type wxHtmlLink() ::
    #wxHtmlLink{type :: wxHtmlLinkEvent:wxHtmlLinkEventType(), linkInfo :: wx:wx_wxHtmlLinkInfo()}.

wxHtmlLinkEvent()

-type wxHtmlLinkEvent() :: wx:wx_object().

wxHtmlLinkEventType()

-type wxHtmlLinkEventType() :: command_html_link_clicked | html_cell_clicked | html_cell_hover.

Functions

getLinkInfo(This)

-spec getLinkInfo(This) -> wx:wx_wxHtmlLinkInfo() when This :: wxHtmlLinkEvent().

Returns the wx_wxHtmlLinkInfo() which contains info about the cell clicked and the hyperlink it contains.