wxIconizeEvent (wx v2.4.3.1)

An event being sent when the frame is iconized (minimized) or restored.

See:

Overview events

wxTopLevelWindow:iconize/2

wxTopLevelWindow:isIconized/1

This class is derived, and can use functions, from:

wxEvent

wxWidgets docs: wxIconizeEvent

Events

Use wxEvtHandler:connect/3 with wxIconizeEventType to subscribe to events of this type.

Types

wxIconize()

-type wxIconize() :: #wxIconize{type :: wxIconizeEvent:wxIconizeEventType(), iconized :: boolean()}.

wxIconizeEvent()

-type wxIconizeEvent() :: wx:wx_object().

wxIconizeEventType()

-type wxIconizeEventType() :: iconize.

Functions

isIconized(This)

-spec isIconized(This) -> boolean() when This :: wxIconizeEvent().

Returns true if the frame has been iconized, false if it has been restored.