wxMiniFrame (wx v2.4.2)

Functions for wxMiniFrame class

A miniframe is a frame with a small title bar. It is suitable for floating toolbars that must not take up too much screen area.

An example of mini frame can be seen in the page_samples_dialogs using the "Mini frame" command of the "Generic dialogs" submenu.

Styles

This class supports the following styles:

Remark: This class has miniframe functionality under Windows and GTK, i.e. the presence of mini frame will not be noted in the task bar and focus behaviour is different. On other platforms, it behaves like a normal frame.

See: wxMDIParentFrame, wxMDIChildFrame, wxFrame, wxDialog

This class is derived (and can use functions) from: wxFrame wxTopLevelWindow wxWindow wxEvtHandler

wxWidgets docs: wxMiniFrame

Types

wxMiniFrame()

-type wxMiniFrame() :: wx:wx_object().

Functions

create(This, Parent, Id, Title)

-spec create(This, Parent, Id, Title) -> boolean()
          when
              This :: wxMiniFrame(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Title :: unicode:chardata().

create/5

-spec create(This, Parent, Id, Title, [Option]) -> boolean()
          when
              This :: wxMiniFrame(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Title :: unicode:chardata(),
              Option ::
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()}.

Used in two-step frame construction.

See new/4 for further details.

destroy(This)

-spec destroy(This :: wxMiniFrame()) -> ok.

Destructor.

Destroys all child windows and menu bar if present.

new()

-spec new() -> wxMiniFrame().

Default ctor.

new(Parent, Id, Title)

-spec new(Parent, Id, Title) -> wxMiniFrame()
       when Parent :: wxWindow:wxWindow(), Id :: integer(), Title :: unicode:chardata().

new/4

-spec new(Parent, Id, Title, [Option]) -> wxMiniFrame()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Title :: unicode:chardata(),
           Option ::
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Constructor, creating the window.

Remark: The frame behaves like a normal frame on non-Windows platforms.

See: create/5