wxMouseCaptureChangedEvent (wx v2.5)

An mouse capture changed event is sent to a window that loses its mouse capture.

This is called even if wxWindow:releaseMouse/1 was called by the application code. Handling this event allows an application to cater for unexpected capture releases which might otherwise confuse mouse handling code.

Only for:wxmsw

See:

wxMouseCaptureLostEvent

Overview events

wxWindow:captureMouse/1

wxWindow:releaseMouse/1

wxWindow:getCapture/0

This class is derived, and can use functions, from:

wxEvent

wxWidgets docs: wxMouseCaptureChangedEvent

Events

Use wxEvtHandler:connect/3 with wxMouseCaptureChangedEventType to subscribe to events of this type.

Types

wxMouseCaptureChanged()

-type wxMouseCaptureChanged() ::
          #wxMouseCaptureChanged{type :: wxMouseCaptureChangedEvent:wxMouseCaptureChangedEventType()}.

wxMouseCaptureChangedEvent()

-type wxMouseCaptureChangedEvent() :: wx:wx_object().

wxMouseCaptureChangedEventType()

-type wxMouseCaptureChangedEventType() :: mouse_capture_changed.

Functions

getCapturedWindow(This)

-spec getCapturedWindow(This) -> wxWindow:wxWindow() when This :: wxMouseCaptureChangedEvent().

Returns the window that gained the capture, or NULL if it was a non-wxWidgets window.