wxMouseCaptureLostEvent (wx v2.4.2)

Functions for wxMouseCaptureLostEvent class

A mouse capture lost event is sent to a window that had obtained mouse capture, which was subsequently lost due to an "external" event (for example, when a dialog box is shown or if another application captures the mouse).

If this happens, this event is sent to all windows that are on the capture stack (i.e. called CaptureMouse, but didn't call ReleaseMouse yet). The event is not sent if the capture changes because of a call to CaptureMouse or ReleaseMouse.

This event is currently emitted under Windows only.

Only for:wxmsw

See: wxMouseCaptureChangedEvent, Overview events, wxWindow:captureMouse/1, wxWindow:releaseMouse/1, wxWindow:getCapture/0

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxMouseCaptureLostEvent

Events

Use wxEvtHandler:connect/3 with wxMouseCaptureLostEventType to subscribe to events of this type.

Types

wxMouseCaptureLost()

-type wxMouseCaptureLost() ::
    #wxMouseCaptureLost{type :: wxMouseCaptureLostEvent:wxMouseCaptureLostEventType()}.

wxMouseCaptureLostEvent()

-type wxMouseCaptureLostEvent() :: wx:wx_object().

wxMouseCaptureLostEventType()

-type wxMouseCaptureLostEventType() :: mouse_capture_lost.