wxNavigationKeyEvent (wx v2.4.2)

Functions for wxNavigationKeyEvent class

This event class contains information about navigation events, generated by navigation keys such as tab and page down.

This event is mainly used by wxWidgets implementations. A wxNavigationKeyEvent handler is automatically provided by wxWidgets when you enable keyboard navigation inside a window by inheriting it from wxNavigationEnabled<>.

See: wxWindow:navigate/2, wxWindow::NavigateIn (not implemented in wx)

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxNavigationKeyEvent

Events

Use wxEvtHandler:connect/3 with wxNavigationKeyEventType to subscribe to events of this type.

Types

wxNavigationKey()

-type wxNavigationKey() ::
    #wxNavigationKey{type :: wxNavigationKeyEvent:wxNavigationKeyEventType(),
                     dir :: boolean(),
                     focus :: wxWindow:wxWindow()}.

wxNavigationKeyEvent()

-type wxNavigationKeyEvent() :: wx:wx_object().

wxNavigationKeyEventType()

-type wxNavigationKeyEventType() :: navigation_key.

Functions

getCurrentFocus(This)

-spec getCurrentFocus(This) -> wxWindow:wxWindow() when This :: wxNavigationKeyEvent().

Returns the child that has the focus, or NULL.

getDirection(This)

-spec getDirection(This) -> boolean() when This :: wxNavigationKeyEvent().

Returns true if the navigation was in the forward direction.

isFromTab(This)

-spec isFromTab(This) -> boolean() when This :: wxNavigationKeyEvent().

Returns true if the navigation event was from a tab key.

This is required for proper navigation over radio buttons.

isWindowChange(This)

-spec isWindowChange(This) -> boolean() when This :: wxNavigationKeyEvent().

Returns true if the navigation event represents a window change (for example, from Ctrl-Page Down in a notebook).

setCurrentFocus(This, CurrentFocus)

-spec setCurrentFocus(This, CurrentFocus) -> ok
                   when This :: wxNavigationKeyEvent(), CurrentFocus :: wxWindow:wxWindow().

Sets the current focus window member.

setDirection(This, Direction)

-spec setDirection(This, Direction) -> ok when This :: wxNavigationKeyEvent(), Direction :: boolean().

Sets the direction to forward if direction is true, or backward if false.

setFromTab(This, FromTab)

-spec setFromTab(This, FromTab) -> ok when This :: wxNavigationKeyEvent(), FromTab :: boolean().

Marks the navigation event as from a tab key.

setWindowChange(This, WindowChange)

-spec setWindowChange(This, WindowChange) -> ok
                   when This :: wxNavigationKeyEvent(), WindowChange :: boolean().

Marks the event as a window change event.