wxPageSetupDialog (wx v2.5)

This class represents the page setup common dialog.

The page setup dialog contains controls for paper size (letter, A4, A5 etc.), orientation (landscape or portrait), and, only under Windows currently, controls for setting left, top, right and bottom margin sizes in millimetres.

The exact appearance of this dialog varies among the platforms as a native dialog is used when available (currently the case for all major platforms).

When the dialog has been closed, you need to query the wxPageSetupDialogData object associated with the dialog.

Note that the OK and Cancel buttons do not destroy the dialog; this must be done by the application.

See:

Overview printing

wxPrintDialog

wxPageSetupDialogData

wxWidgets docs: wxPageSetupDialog

Types

wxPageSetupDialog()

-type wxPageSetupDialog() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxPageSetupDialog()) -> ok.

Destroys the object

getPageSetupData(This)

-spec getPageSetupData(This) -> wxPageSetupDialogData:wxPageSetupDialogData()
                          when This :: wxPageSetupDialog().

Returns the wxPageSetupDialogData object associated with the dialog.

new(Parent)

-spec new(Parent) -> wxPageSetupDialog() when Parent :: wxWindow:wxWindow().

Equivalent to new(Parent, []).

new/2

-spec new(Parent, [Option]) -> wxPageSetupDialog()
             when
                 Parent :: wxWindow:wxWindow(),
                 Option :: {data, wxPageSetupDialogData:wxPageSetupDialogData()}.

Constructor.

Pass a parent window, and optionally a pointer to a block of page setup data, which will be copied to the print dialog's internal data.

showModal(This)

-spec showModal(This) -> integer() when This :: wxPageSetupDialog().

Shows the dialog, returning wxID\_OK if the user pressed OK, and wxID\_CANCEL otherwise.