wxPageSetupDialogData (wx v2.4.2)

Functions for wxPageSetupDialogData class

This class holds a variety of information related to wxPageSetupDialog.

It contains a wxPrintData member which is used to hold basic printer configuration data (as opposed to the user-interface configuration settings stored by wxPageSetupDialogData).

See: Overview printing, wxPageSetupDialog

wxWidgets docs: wxPageSetupDialogData

Types

wxPageSetupDialogData()

-type wxPageSetupDialogData() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxPageSetupDialogData()) -> ok.

Destructor.

enableHelp(This, Flag)

-spec enableHelp(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the "Help" button (Windows only).

enableMargins(This, Flag)

-spec enableMargins(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the margin controls (Windows only).

enableOrientation(This, Flag)

-spec enableOrientation(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the orientation control (Windows only).

enablePaper(This, Flag)

-spec enablePaper(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the paper size control (Windows only).

enablePrinter(This, Flag)

-spec enablePrinter(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the "Printer" button, which invokes a printer setup dialog.

getDefaultInfo(This)

-spec getDefaultInfo(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the dialog will simply return default printer information (such as orientation) instead of showing a dialog (Windows only).

getDefaultMinMargins(This)

-spec getDefaultMinMargins(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the page setup dialog will take its minimum margin values from the currently selected printer properties (Windows only).

getEnableHelp(This)

-spec getEnableHelp(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the printer setup button is enabled.

getEnableMargins(This)

-spec getEnableMargins(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the margin controls are enabled (Windows only).

getEnableOrientation(This)

-spec getEnableOrientation(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the orientation control is enabled (Windows only).

getEnablePaper(This)

-spec getEnablePaper(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the paper size control is enabled (Windows only).

getEnablePrinter(This)

-spec getEnablePrinter(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the printer setup button is enabled.

getMarginBottomRight(This)

-spec getMarginBottomRight(This) -> {X :: integer(), Y :: integer()}
                        when This :: wxPageSetupDialogData().

Returns the right (x) and bottom (y) margins in millimetres.

getMarginTopLeft(This)

-spec getMarginTopLeft(This) -> {X :: integer(), Y :: integer()} when This :: wxPageSetupDialogData().

Returns the left (x) and top (y) margins in millimetres.

getMinMarginBottomRight(This)

-spec getMinMarginBottomRight(This) -> {X :: integer(), Y :: integer()}
                           when This :: wxPageSetupDialogData().

Returns the right (x) and bottom (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

getMinMarginTopLeft(This)

-spec getMinMarginTopLeft(This) -> {X :: integer(), Y :: integer()} when This :: wxPageSetupDialogData().

Returns the left (x) and top (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

getPaperId(This)

-spec getPaperId(This) -> wx:wx_enum() when This :: wxPageSetupDialogData().

Returns the paper id (stored in the internal wxPrintData object).

See: wxPrintData:setPaperId/2

getPaperSize(This)

-spec getPaperSize(This) -> {W :: integer(), H :: integer()} when This :: wxPageSetupDialogData().

Returns the paper size in millimetres.

getPrintData(This)

-spec getPrintData(This) -> wxPrintData:wxPrintData() when This :: wxPageSetupDialogData().

isOk(This)

-spec isOk(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the print data associated with the dialog data is valid.

This can return false on Windows if the current printer is not set, for example. On all other platforms, it returns true.

new()

-spec new() -> wxPageSetupDialogData().

Default constructor.

new(PrintData)

-spec new(PrintData) -> wxPageSetupDialogData()
       when PrintData :: wxPrintData:wxPrintData() | wxPageSetupDialogData:wxPageSetupDialogData().

Construct an object from a print data object.

setDefaultInfo(This, Flag)

-spec setDefaultInfo(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Pass true if the dialog will simply return default printer information (such as orientation) instead of showing a dialog (Windows only).

setDefaultMinMargins(This, Flag)

-spec setDefaultMinMargins(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Pass true if the page setup dialog will take its minimum margin values from the currently selected printer properties (Windows only).

Units are in millimetres.

setMarginBottomRight(This, Pt)

-spec setMarginBottomRight(This, Pt) -> ok
                        when
                            This :: wxPageSetupDialogData(),
                            Pt :: {X :: integer(), Y :: integer()}.

Sets the right (x) and bottom (y) margins in millimetres.

setMarginTopLeft(This, Pt)

-spec setMarginTopLeft(This, Pt) -> ok
                    when This :: wxPageSetupDialogData(), Pt :: {X :: integer(), Y :: integer()}.

Sets the left (x) and top (y) margins in millimetres.

setMinMarginBottomRight(This, Pt)

-spec setMinMarginBottomRight(This, Pt) -> ok
                           when
                               This :: wxPageSetupDialogData(),
                               Pt :: {X :: integer(), Y :: integer()}.

Sets the right (x) and bottom (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

setMinMarginTopLeft(This, Pt)

-spec setMinMarginTopLeft(This, Pt) -> ok
                       when
                           This :: wxPageSetupDialogData(), Pt :: {X :: integer(), Y :: integer()}.

Sets the left (x) and top (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

setPaperId(This, Id)

-spec setPaperId(This, Id) -> ok when This :: wxPageSetupDialogData(), Id :: wx:wx_enum().

Sets the paper size id.

Calling this function overrides the explicit paper dimensions passed in setPaperSize/2.

See: wxPrintData:setPaperId/2

setPaperSize(This, Size)

-spec setPaperSize(This, Size) -> ok
                when This :: wxPageSetupDialogData(), Size :: {W :: integer(), H :: integer()}.

Sets the paper size in millimetres.

If a corresponding paper id is found, it will be set in the internal wxPrintData object, otherwise the paper size overrides the paper id.

setPrintData(This, PrintData)

-spec setPrintData(This, PrintData) -> ok
                when This :: wxPageSetupDialogData(), PrintData :: wxPrintData:wxPrintData().

Sets the print data associated with this object.