wxPaintDC (wx v2.4.2)

Functions for wxPaintDC class

A wxPaintDC must be constructed if an application wishes to paint on the client area of a window from within an EVT_PAINT() event handler. This should normally be constructed as a temporary stack object; don't store a wxPaintDC object. If you have an EVT_PAINT() handler, you must create a wxPaintDC object within it even if you don't actually use it.

Using wxPaintDC within your EVT_PAINT() handler is important because it automatically sets the clipping area to the damaged area of the window. Attempts to draw outside this area do not appear.

A wxPaintDC object is initialized to use the same font and colours as the window it is associated with.

See: wxDC, wxClientDC, wxMemoryDC, wxWindowDC, wxScreenDC

This class is derived (and can use functions) from: wxWindowDC wxDC

wxWidgets docs: wxPaintDC

Types

wxPaintDC()

-type wxPaintDC() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxPaintDC()) -> ok.

Destroys the object.

new(Window)

-spec new(Window) -> wxPaintDC() when Window :: wxWindow:wxWindow().

Constructor.

Pass a pointer to the window on which you wish to paint.