wxPaintEvent (wx v2.3.1)

Functions for wxPaintEvent class

A paint event is sent when a window's contents needs to be repainted.

The handler of this event must create a wxPaintDC object and use it for painting the window contents. For example:

Notice that you must not create other kinds of wxDC (e.g. wxClientDC or wxWindowDC) in EVT_PAINT handlers and also don't create wxPaintDC outside of this event handlers.

You can optimize painting by retrieving the rectangles that have been damaged and only repainting these. The rectangles are in terms of the client area, and are unscrolled, so you will need to do some calculations using the current view position to obtain logical, scrolled units. Here is an example of using the wxRegionIterator (not implemented in wx) class:

Remark: Please notice that in general it is impossible to change the drawing of a standard control (such as wxButton) and so you shouldn't attempt to handle paint events for them as even if it might work on some platforms, this is inherently not portable and won't work everywhere.

See: Overview events

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxPaintEvent

Events

Use wxEvtHandler:connect/3 with wxPaintEventType to subscribe to events of this type.

Types

event()

-type event() ::
    wxActivateEvent:wxActivate() |
    wxAuiManagerEvent:wxAuiManager() |
    wxAuiNotebookEvent:wxAuiNotebook() |
    wxBookCtrlEvent:wxBookCtrl() |
    wxCalendarEvent:wxCalendar() |
    wxChildFocusEvent:wxChildFocus() |
    wxClipboardTextEvent:wxClipboardText() |
    wxCloseEvent:wxClose() |
    wxColourPickerEvent:wxColourPicker() |
    wxCommandEvent:wxCommand() |
    wxContextMenuEvent:wxContextMenu() |
    wxDateEvent:wxDate() |
    wxDisplayChangedEvent:wxDisplayChanged() |
    wxDropFilesEvent:wxDropFiles() |
    wxEraseEvent:wxErase() |
    wxFileDirPickerEvent:wxFileDirPicker() |
    wxFocusEvent:wxFocus() |
    wxFontPickerEvent:wxFontPicker() |
    wxGridEvent:wxGrid() |
    wxHelpEvent:wxHelp() |
    wxHtmlLinkEvent:wxHtmlLink() |
    wxIconizeEvent:wxIconize() |
    wxIdleEvent:wxIdle() |
    wxInitDialogEvent:wxInitDialog() |
    wxJoystickEvent:wxJoystick() |
    wxKeyEvent:wxKey() |
    wxListEvent:wxList() |
    wxMaximizeEvent:wxMaximize() |
    wxMenuEvent:wxMenu() |
    wxMouseCaptureChangedEvent:wxMouseCaptureChanged() |
    wxMouseCaptureLostEvent:wxMouseCaptureLost() |
    wxMouseEvent:wxMouse() |
    wxMoveEvent:wxMove() |
    wxNavigationKeyEvent:wxNavigationKey() |
    wxPaintEvent:wxPaint() |
    wxPaletteChangedEvent:wxPaletteChanged() |
    wxQueryNewPaletteEvent:wxQueryNewPalette() |
    wxSashEvent:wxSash() |
    wxScrollEvent:wxScroll() |
    wxScrollWinEvent:wxScrollWin() |
    wxSetCursorEvent:wxSetCursor() |
    wxShowEvent:wxShow() |
    wxSizeEvent:wxSize() |
    wxSpinEvent:wxSpin() |
    wxSplitterEvent:wxSplitter() |
    wxStyledTextEvent:wxStyledText() |
    wxSysColourChangedEvent:wxSysColourChanged() |
    wxTaskBarIconEvent:wxTaskBarIcon() |
    wxTreeEvent:wxTree() |
    wxUpdateUIEvent:wxUpdateUI() |
    wxWebViewEvent:wxWebView() |
    wxWindowCreateEvent:wxWindowCreate() |
    wxWindowDestroyEvent:wxWindowDestroy().

wx()

-type wx() :: #wx{}.

wxActivate()

-type wxActivate() :: #wxActivate{}.

wxAuiManager()

-type wxAuiManager() :: #wxAuiManager{}.

wxAuiNotebook()

-type wxAuiNotebook() :: #wxAuiNotebook{}.

wxBookCtrl()

-type wxBookCtrl() :: #wxBookCtrl{}.

wxCalendar()

-type wxCalendar() :: #wxCalendar{}.

wxChildFocus()

-type wxChildFocus() :: #wxChildFocus{}.

wxClipboardText()

-type wxClipboardText() :: #wxClipboardText{}.

wxClose()

-type wxClose() :: #wxClose{}.

wxColourPicker()

-type wxColourPicker() :: #wxColourPicker{}.

wxCommand()

-type wxCommand() :: #wxCommand{}.

wxContextMenu()

-type wxContextMenu() :: #wxContextMenu{}.

wxDate()

-type wxDate() :: #wxDate{}.

wxDisplayChanged()

-type wxDisplayChanged() :: #wxDisplayChanged{}.

wxDropFiles()

-type wxDropFiles() :: #wxDropFiles{}.

wxErase()

-type wxErase() :: #wxErase{}.

wxEventType()

-type wxEventType() ::
    wxActivateEvent:wxActivateEventType() |
    wxAuiManagerEvent:wxAuiManagerEventType() |
    wxAuiNotebookEvent:wxAuiNotebookEventType() |
    wxBookCtrlEvent:wxBookCtrlEventType() |
    wxCalendarEvent:wxCalendarEventType() |
    wxChildFocusEvent:wxChildFocusEventType() |
    wxClipboardTextEvent:wxClipboardTextEventType() |
    wxCloseEvent:wxCloseEventType() |
    wxColourPickerEvent:wxColourPickerEventType() |
    wxCommandEvent:wxCommandEventType() |
    wxContextMenuEvent:wxContextMenuEventType() |
    wxDateEvent:wxDateEventType() |
    wxDisplayChangedEvent:wxDisplayChangedEventType() |
    wxDropFilesEvent:wxDropFilesEventType() |
    wxEraseEvent:wxEraseEventType() |
    wxFileDirPickerEvent:wxFileDirPickerEventType() |
    wxFocusEvent:wxFocusEventType() |
    wxFontPickerEvent:wxFontPickerEventType() |
    wxGridEvent:wxGridEventType() |
    wxHelpEvent:wxHelpEventType() |
    wxHtmlLinkEvent:wxHtmlLinkEventType() |
    wxIconizeEvent:wxIconizeEventType() |
    wxIdleEvent:wxIdleEventType() |
    wxInitDialogEvent:wxInitDialogEventType() |
    wxJoystickEvent:wxJoystickEventType() |
    wxKeyEvent:wxKeyEventType() |
    wxListEvent:wxListEventType() |
    wxMaximizeEvent:wxMaximizeEventType() |
    wxMenuEvent:wxMenuEventType() |
    wxMouseCaptureChangedEvent:wxMouseCaptureChangedEventType() |
    wxMouseCaptureLostEvent:wxMouseCaptureLostEventType() |
    wxMouseEvent:wxMouseEventType() |
    wxMoveEvent:wxMoveEventType() |
    wxNavigationKeyEvent:wxNavigationKeyEventType() |
    wxPaintEvent:wxPaintEventType() |
    wxPaletteChangedEvent:wxPaletteChangedEventType() |
    wxQueryNewPaletteEvent:wxQueryNewPaletteEventType() |
    wxSashEvent:wxSashEventType() |
    wxScrollEvent:wxScrollEventType() |
    wxScrollWinEvent:wxScrollWinEventType() |
    wxSetCursorEvent:wxSetCursorEventType() |
    wxShowEvent:wxShowEventType() |
    wxSizeEvent:wxSizeEventType() |
    wxSpinEvent:wxSpinEventType() |
    wxSplitterEvent:wxSplitterEventType() |
    wxStyledTextEvent:wxStyledTextEventType() |
    wxSysColourChangedEvent:wxSysColourChangedEventType() |
    wxTaskBarIconEvent:wxTaskBarIconEventType() |
    wxTreeEvent:wxTreeEventType() |
    wxUpdateUIEvent:wxUpdateUIEventType() |
    wxWebViewEvent:wxWebViewEventType() |
    wxWindowCreateEvent:wxWindowCreateEventType() |
    wxWindowDestroyEvent:wxWindowDestroyEventType().

wxFileDirPicker()

-type wxFileDirPicker() :: #wxFileDirPicker{}.

wxFocus()

-type wxFocus() :: #wxFocus{}.

wxFontPicker()

-type wxFontPicker() :: #wxFontPicker{}.

wxGrid()

-type wxGrid() :: #wxGrid{}.

wxHelp()

-type wxHelp() :: #wxHelp{}.

wxHtmlLink()

-type wxHtmlLink() :: #wxHtmlLink{}.

wxIconize()

-type wxIconize() :: #wxIconize{}.

wxIdle()

-type wxIdle() :: #wxIdle{}.

wxInitDialog()

-type wxInitDialog() :: #wxInitDialog{}.

wxJoystick()

-type wxJoystick() :: #wxJoystick{}.

wxKey()

-type wxKey() :: #wxKey{}.

wxList()

-type wxList() :: #wxList{}.

wxMaximize()

-type wxMaximize() :: #wxMaximize{}.

wxMenu()

-type wxMenu() :: #wxMenu{}.

wxMouse()

-type wxMouse() :: #wxMouse{}.

wxMouseCaptureChanged()

-type wxMouseCaptureChanged() :: #wxMouseCaptureChanged{}.

wxMouseCaptureLost()

-type wxMouseCaptureLost() :: #wxMouseCaptureLost{}.

wxMove()

-type wxMove() :: #wxMove{}.

wxNavigationKey()

-type wxNavigationKey() :: #wxNavigationKey{}.

wxPaint()

-type wxPaint() :: #wxPaint{}.

wxPaintEvent()

-type wxPaintEvent() :: wx:wx_object().

wxPaintEventType()

-type wxPaintEventType() :: paint.

wxPaletteChanged()

-type wxPaletteChanged() :: #wxPaletteChanged{}.

wxQueryNewPalette()

-type wxQueryNewPalette() :: #wxQueryNewPalette{}.

wxSash()

-type wxSash() :: #wxSash{}.

wxScroll()

-type wxScroll() :: #wxScroll{}.

wxScrollWin()

-type wxScrollWin() :: #wxScrollWin{}.

wxSetCursor()

-type wxSetCursor() :: #wxSetCursor{}.

wxShow()

-type wxShow() :: #wxShow{}.

wxSize()

-type wxSize() :: #wxSize{}.

wxSpin()

-type wxSpin() :: #wxSpin{}.

wxSplitter()

-type wxSplitter() :: #wxSplitter{}.

wxStyledText()

-type wxStyledText() :: #wxStyledText{}.

wxSysColourChanged()

-type wxSysColourChanged() :: #wxSysColourChanged{}.

wxTaskBarIcon()

-type wxTaskBarIcon() :: #wxTaskBarIcon{}.

wxTree()

-type wxTree() :: #wxTree{}.

wxUpdateUI()

-type wxUpdateUI() :: #wxUpdateUI{}.

wxWebView()

-type wxWebView() :: #wxWebView{}.

wxWindowCreate()

-type wxWindowCreate() :: #wxWindowCreate{}.

wxWindowDestroy()

-type wxWindowDestroy() :: #wxWindowDestroy{}.