wxPopupTransientWindow (wx v2.4.2)

Functions for wxPopupTransientWindow class

A wxPopupWindow which disappears automatically when the user clicks mouse outside it or if it loses focus in any other way.

This window can be useful for implementing custom combobox-like controls for example.

See: wxPopupWindow

This class is derived (and can use functions) from: wxPopupWindow wxWindow wxEvtHandler

wxWidgets docs: wxPopupTransientWindow

Types

wxPopupTransientWindow()

-type wxPopupTransientWindow() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxPopupTransientWindow()) -> ok.

Destroys the object.

dismiss(This)

-spec dismiss(This) -> ok when This :: wxPopupTransientWindow().

Hide the window.

new()

-spec new() -> wxPopupTransientWindow().

Default constructor.

new(Parent)

-spec new(Parent) -> wxPopupTransientWindow() when Parent :: wxWindow:wxWindow().

new/2

-spec new(Parent, [Option]) -> wxPopupTransientWindow()
       when Parent :: wxWindow:wxWindow(), Option :: {style, integer()}.

Constructor.

popup(This)

-spec popup(This) -> ok when This :: wxPopupTransientWindow().

popup/2

-spec popup(This, [Option]) -> ok
         when This :: wxPopupTransientWindow(), Option :: {focus, wxWindow:wxWindow()}.

Popup the window (this will show it too).

If focus is non-NULL, it will be kept focused while this window is shown if supported by the current platform, otherwise the popup itself will receive focus. In any case, the popup will disappear automatically if it loses focus because of a user action.

See: dismiss/1