wxPopupWindow (wx v2.3.1)

Functions for wxPopupWindow class

A special kind of top level window used for popup menus, combobox popups and such.

Styles

This class supports the following styles:

See: wxDialog, wxFrame

This class is derived (and can use functions) from: wxWindow wxEvtHandler

wxWidgets docs: wxPopupWindow

Types

wxPopupWindow()

-type wxPopupWindow() :: wx:wx_object().

Functions

create/2

-spec create(This, Parent) -> boolean() when This :: wxPopupWindow(), Parent :: wxWindow:wxWindow().

create/3

-spec create(This, Parent, [Option]) -> boolean()
          when
              This :: wxPopupWindow(), Parent :: wxWindow:wxWindow(), Option :: {flags, integer()}.

Create method for two-step creation.

destroy/1

-spec destroy(This :: wxPopupWindow()) -> ok.

Destroys the object.

new()

-spec new() -> wxPopupWindow().

Default constructor.

new/1

-spec new(Parent) -> wxPopupWindow() when Parent :: wxWindow:wxWindow().

new/2

-spec new(Parent, [Option]) -> wxPopupWindow()
       when Parent :: wxWindow:wxWindow(), Option :: {flags, integer()}.

Constructor.

position/3

-spec position(This, PtOrigin, SizePopup) -> ok
            when
                This :: wxPopupWindow(),
                PtOrigin :: {X :: integer(), Y :: integer()},
                SizePopup :: {W :: integer(), H :: integer()}.

Move the popup window to the right position, i.e. such that it is entirely visible.

The popup is positioned at ptOrigin + size if it opens below and to the right (default), at ptOrigin - sizePopup if it opens above and to the left etc.