wxPreviewCanvas (wx v2.4.2)

Functions for wxPreviewCanvas class

A preview canvas is the default canvas used by the print preview system to display the preview.

See: wxPreviewFrame, wxPreviewControlBar, wxPrintPreview

This class is derived (and can use functions) from: wxScrolledWindow wxPanel wxWindow wxEvtHandler

wxWidgets docs: wxPreviewCanvas

Types

wxPreviewCanvas()

-type wxPreviewCanvas() :: wx:wx_object().