wxScrollWinEvent (wx v2.0.1)

A scroll event holds information about events sent from scrolling windows.

Note that you can use the EVT_SCROLLWIN* macros for intercepting scroll window events from the receiving window.

See: wxScrollEvent, Overview events

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxScrollWinEvent

Events

Use wxEvtHandler:connect/3 with wxScrollWinEventType to subscribe to events of this type.

Types

-type wxScrollWin() :: term().

Specs

wxScrollWin() :: #wxScrollWin{}.

-type wxScrollWinEvent() :: term().

Specs

wxScrollWinEvent() :: wx:wx_object().

-type wxScrollWinEventType() :: term().

Specs

wxScrollWinEventType() ::
    scrollwin_top | scrollwin_bottom | scrollwin_lineup | scrollwin_linedown | scrollwin_pageup |
    scrollwin_pagedown | scrollwin_thumbtrack | scrollwin_thumbrelease.

Functions

getOrientation/1

Specs

getOrientation(This) -> integer() when This :: wxScrollWinEvent().

Returns wxHORIZONTAL or wxVERTICAL, depending on the orientation of the scrollbar.

getPosition/1

Specs

getPosition(This) -> integer() when This :: wxScrollWinEvent().

Returns the position of the scrollbar for the thumb track and release events.

Note that this field can't be used for the other events, you need to query the window itself for the current position in that case.