wxSpinEvent (wx v2.5.2-rc0)

This event class is used for the events generated by wxSpinButton and wxSpinCtrl.

See:

wxSpinButton

wxSpinCtrl

This class is derived, and can use functions, from:

wxNotifyEvent

wxCommandEvent

wxEvent

wxWidgets docs: wxSpinEvent

Events

Use wxEvtHandler:connect/3 with wxSpinEventType to subscribe to events of this type.

Types

wxSpin()

-type wxSpin() :: #wxSpin{type :: wxSpinEvent:wxSpinEventType(), commandInt :: integer()}.

wxSpinEvent()

-type wxSpinEvent() :: wx:wx_object().

wxSpinEventType()

-type wxSpinEventType() :: command_spinctrl_updated | spin_up | spin_down | spin.

Functions

getPosition(This)

-spec getPosition(This) -> integer() when This :: wxSpinEvent().

Retrieve the current spin button or control value.

setPosition(This, Pos)

-spec setPosition(This, Pos) -> ok when This :: wxSpinEvent(), Pos :: integer().

Set the value associated with the event.