wxSplashScreen (wx v2.0.1)

wxSplashScreen shows a window with a thin border, displaying a bitmap describing your application.

Show it in application initialisation, and then either explicitly destroy it or let it time-out.

Example usage:

This class is derived (and can use functions) from: wxFrame wxTopLevelWindow wxWindow wxEvtHandler

wxWidgets docs: wxSplashScreen

Types

-type wxSplashScreen() :: term().

Specs

wxSplashScreen() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxSplashScreen()) -> ok.

Destroys the splash screen.

getSplashStyle/1

Specs

getSplashStyle(This) -> integer() when This :: wxSplashScreen().

Returns the splash style (see new/6 for details).

getTimeout/1

Specs

getTimeout(This) -> integer() when This :: wxSplashScreen().

Returns the timeout in milliseconds.

new/5

Specs

new(Bitmap, SplashStyle, Milliseconds, Parent, Id) -> wxSplashScreen()
       when
           Bitmap :: wxBitmap:wxBitmap(),
           SplashStyle :: integer(),
           Milliseconds :: integer(),
           Parent :: wxWindow:wxWindow(),
           Id :: integer().

new/6

Specs

new(Bitmap, SplashStyle, Milliseconds, Parent, Id, [Option]) -> wxSplashScreen()
       when
           Bitmap :: wxBitmap:wxBitmap(),
           SplashStyle :: integer(),
           Milliseconds :: integer(),
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Option ::
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Construct the splash screen passing a bitmap, a style, a timeout, a window id, optional position and size, and a window style.

splashStyle is a bitlist of some of the following:

milliseconds is the timeout in milliseconds.