wxStaticBitmap (wx v2.0.1)

A static bitmap control displays a bitmap. Native implementations on some platforms are only meant for display of the small icons in the dialog boxes.

If you want to display larger images portably, you may use generic implementation wxGenericStaticBitmap declared in <wx/generic/statbmpg.h>.

Notice that for the best results, the size of the control should be the same as the size of the image displayed in it, as happens by default if it's not resized explicitly. Otherwise, behaviour depends on the platform: under MSW, the bitmap is drawn centred inside the control, while elsewhere it is drawn at the origin of the control. You can use SetScaleMode() (not implemented in wx) to control how the image is scaled inside the control.

See: wxBitmap

This class is derived (and can use functions) from: wxControl wxWindow wxEvtHandler

wxWidgets docs: wxStaticBitmap

Types

-type wxStaticBitmap() :: term().

Specs

wxStaticBitmap() :: wx:wx_object().

Functions

create/4

Specs

create(This, Parent, Id, Label) -> boolean()
          when
              This :: wxStaticBitmap(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: wxBitmap:wxBitmap().

create/5

Specs

create(This, Parent, Id, Label, [Option]) -> boolean()
          when
              This :: wxStaticBitmap(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: wxBitmap:wxBitmap(),
              Option ::
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()}.

Creation function, for two-step construction.

For details see new/4.

destroy/1

Specs

destroy(This :: wxStaticBitmap()) -> ok.

Destroys the object.

getBitmap/1

Specs

getBitmap(This) -> wxBitmap:wxBitmap() when This :: wxStaticBitmap().

Returns the bitmap currently used in the control.

Notice that this method can be called even if SetIcon() (not implemented in wx) had been used.

See: setBitmap/2

new/0

Specs

new() -> wxStaticBitmap().

Default constructor.

new/3

Specs

new(Parent, Id, Label) -> wxStaticBitmap()
       when Parent :: wxWindow:wxWindow(), Id :: integer(), Label :: wxBitmap:wxBitmap().

new/4

Specs

new(Parent, Id, Label, [Option]) -> wxStaticBitmap()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Label :: wxBitmap:wxBitmap(),
           Option ::
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Constructor, creating and showing a static bitmap control.

See: create/5

setBitmap/2

Specs

setBitmap(This, Label) -> ok when This :: wxStaticBitmap(), Label :: wxBitmap:wxBitmap().

Sets the bitmap label.

See: getBitmap/1