wxStaticBox (wx v2.4.2)

Functions for wxStaticBox class

A static box is a rectangle drawn around other windows to denote a logical grouping of items.

Note that while the previous versions required that windows appearing inside a static box be created as its siblings (i.e. use the same parent as the static box itself), since wxWidgets 2.9.1 it is also possible to create them as children of wxStaticBox itself and you are actually encouraged to do it like this if compatibility with the previous versions is not important.

So the new recommended way to create static box is:

While the compatible - and now deprecated - way is

Also note that there is a specialized wxSizer class (wxStaticBoxSizer) which can be used as an easier way to pack items into a static box.

See: wxStaticText, wxStaticBoxSizer

This class is derived (and can use functions) from: wxControl wxWindow wxEvtHandler

wxWidgets docs: wxStaticBox

Types

wxStaticBox()

-type wxStaticBox() :: wx:wx_object().

Functions

create(This, Parent, Id, Label)

-spec create(This, Parent, Id, Label) -> boolean()
          when
              This :: wxStaticBox(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: unicode:chardata().

create/5

-spec create(This, Parent, Id, Label, [Option]) -> boolean()
          when
              This :: wxStaticBox(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: unicode:chardata(),
              Option ::
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()}.

Creates the static box for two-step construction.

See new/4 for further details.

destroy(This)

-spec destroy(This :: wxStaticBox()) -> ok.

Constructor for a static box using the given window as label.

This constructor takes a pointer to an arbitrary window (although usually a wxCheckBox or a wxRadioButton) instead of just the usual text label and puts this window at the top of the box at the place where the label would be shown.

The label window must be a non-null, fully created window and will become a child of this wxStaticBox, i.e. it will be owned by this control and will be deleted when the wxStaticBox itself is deleted.

An example of creating a wxStaticBox with window as a label:

Currently this constructor is only available in wxGTK and wxMSW, use wxHAS_WINDOW_LABEL_IN_STATIC_BOX to check whether it can be used at compile-time.

Since: 3.1.1 Destructor, destroying the group box.

new()

-spec new() -> wxStaticBox().

Default constructor.

new(Parent, Id, Label)

-spec new(Parent, Id, Label) -> wxStaticBox()
       when Parent :: wxWindow:wxWindow(), Id :: integer(), Label :: unicode:chardata().

new/4

-spec new(Parent, Id, Label, [Option]) -> wxStaticBox()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Label :: unicode:chardata(),
           Option ::
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Constructor, creating and showing a static box.

See: create/5