wxStaticBoxSizer (wx v2.4.2)

Functions for wxStaticBoxSizer class

wxStaticBoxSizer is a sizer derived from wxBoxSizer but adds a static box around the sizer.

The static box may be either created independently or the sizer may create it itself as a convenience. In any case, the sizer owns the wxStaticBox control and will delete it in the wxStaticBoxSizer destructor.

Note that since wxWidgets 2.9.1 you are encouraged to create the windows which are added to wxStaticBoxSizer as children of wxStaticBox itself, see this class documentation for more details.

Example of use of this class:

See: wxSizer, wxStaticBox, wxBoxSizer, Overview sizer

This class is derived (and can use functions) from: wxBoxSizer wxSizer

wxWidgets docs: wxStaticBoxSizer

Types

wxStaticBoxSizer()

-type wxStaticBoxSizer() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxStaticBoxSizer()) -> ok.

Destroys the object.

getStaticBox(This)

-spec getStaticBox(This) -> wxStaticBox:wxStaticBox() when This :: wxStaticBoxSizer().

Returns the static box associated with the sizer.

new/2

-spec new(Orient, Parent) -> wxStaticBoxSizer() when Orient :: integer(), Parent :: wxWindow:wxWindow();
   (Box, Orient) -> wxStaticBoxSizer() when Box :: wxStaticBox:wxStaticBox(), Orient :: integer().

This constructor uses an already existing static box.

new/3

-spec new(Orient, Parent, [Option]) -> wxStaticBoxSizer()
       when
           Orient :: integer(),
           Parent :: wxWindow:wxWindow(),
           Option :: {label, unicode:chardata()}.

This constructor creates a new static box with the given label and parent window.