wxStaticLine (wx v2.4.3)

A static line is just a line which may be used in a dialog to separate the groups of controls.

The line may be only vertical or horizontal. Moreover, not all ports (notably not wxGTK) support specifying the transversal direction of the line (e.g. height for a horizontal line) so for maximal portability you should specify it as wxDefaultCoord.

Styles

This class supports the following styles:

wxLI_HORIZONTAL: Creates a horizontal line.

wxLI_VERTICAL: Creates a vertical line.

See: wxStaticBox

This class is derived, and can use functions, from:

wxControl

wxWindow

wxEvtHandler

wxWidgets docs: wxStaticLine

Types

wxStaticLine()

-type wxStaticLine() :: wx:wx_object().

Functions

create(This, Parent)

-spec create(This, Parent) -> boolean() when This :: wxStaticLine(), Parent :: wxWindow:wxWindow().

Equivalent to create(This, Parent, []).

create/3

-spec create(This, Parent, [Option]) -> boolean()
                when
                    This :: wxStaticLine(),
                    Parent :: wxWindow:wxWindow(),
                    Option ::
                        {id, integer()} |
                        {pos, {X :: integer(), Y :: integer()}} |
                        {size, {W :: integer(), H :: integer()}} |
                        {style, integer()}.

Creates the static line for two-step construction.

See new/2 for further details.

destroy(This)

-spec destroy(This :: wxStaticLine()) -> ok.

Destroys the object

getDefaultSize()

-spec getDefaultSize() -> integer().

This static function returns the size which will be given to the smaller dimension of the static line, i.e.

its height for a horizontal line or its width for a vertical one.

isVertical(This)

-spec isVertical(This) -> boolean() when This :: wxStaticLine().

Returns true if the line is vertical, false if horizontal.

new()

-spec new() -> wxStaticLine().

Default constructor.

new(Parent)

-spec new(Parent) -> wxStaticLine() when Parent :: wxWindow:wxWindow().

Equivalent to new(Parent, []).

new/2

-spec new(Parent, [Option]) -> wxStaticLine()
             when
                 Parent :: wxWindow:wxWindow(),
                 Option ::
                     {id, integer()} |
                     {pos, {X :: integer(), Y :: integer()}} |
                     {size, {W :: integer(), H :: integer()}} |
                     {style, integer()}.

Constructor, creating and showing a static line.

See: create/3