wxStyledTextEvent (wx v2.4.3)

The type of events sent from wxStyledTextCtrl.

This class is derived, and can use functions, from:

wxCommandEvent

wxEvent

wxWidgets docs: wxStyledTextEvent

Events

Use wxEvtHandler:connect/3 with wxStyledTextEventType to subscribe to events of this type.

Types

wxStyledText()

-type wxStyledText() ::
          #wxStyledText{type :: wxStyledTextEvent:wxStyledTextEventType(),
                        position :: integer(),
                        key :: integer(),
                        modifiers :: integer(),
                        modificationType :: integer(),
                        text :: unicode:chardata(),
                        length :: integer(),
                        linesAdded :: integer(),
                        line :: integer(),
                        foldLevelNow :: integer(),
                        foldLevelPrev :: integer(),
                        margin :: integer(),
                        message :: integer(),
                        wParam :: integer(),
                        lParam :: integer(),
                        listType :: integer(),
                        x :: integer(),
                        y :: integer(),
                        dragText :: unicode:chardata(),
                        dragAllowMove :: boolean(),
                        dragResult :: wx:wx_enum()}.

wxStyledTextEvent()

-type wxStyledTextEvent() :: wx:wx_object().

wxStyledTextEventType()

-type wxStyledTextEventType() ::
          stc_autocomp_cancelled | stc_autocomp_char_deleted | stc_autocomp_selection |
          stc_calltip_click | stc_change | stc_charadded | stc_do_drop | stc_doubleclick |
          stc_drag_over | stc_dwellend | stc_dwellstart | stc_hotspot_click | stc_hotspot_dclick |
          stc_hotspot_release_click | stc_indicator_click | stc_indicator_release | stc_macrorecord |
          stc_marginclick | stc_modified | stc_needshown | stc_painted | stc_romodifyattempt |
          stc_savepointleft | stc_savepointreached | stc_start_drag | stc_styleneeded | stc_updateui |
          stc_userlistselection | stc_zoom.

Functions

getAlt(This)

-spec getAlt(This) -> boolean() when This :: wxStyledTextEvent().

Returns true if the Alt key is pressed.

This method is valid for the following event types:

wxEVT_STC_DOUBLECLICK

wxEVT_STC_MARGINCLICK

wxEVT_STC_HOTSPOT_CLICK

wxEVT_STC_HOTSPOT_DCLICK

wxEVT_STC_HOTSPOT_RELEASE_CLICK

wxEVT_STC_INDICATOR_CLICK

wxEVT_STC_INDICATOR_RELEASE

wxEVT_STC_MARGIN_RIGHT_CLICK

getControl(This)

-spec getControl(This) -> boolean() when This :: wxStyledTextEvent().

Returns true if the Control key is pressed.

This method is valid for the following event types:

wxEVT_STC_DOUBLECLICK

wxEVT_STC_MARGINCLICK

wxEVT_STC_HOTSPOT_CLICK

wxEVT_STC_HOTSPOT_DCLICK

wxEVT_STC_HOTSPOT_RELEASE_CLICK

wxEVT_STC_INDICATOR_CLICK

wxEVT_STC_INDICATOR_RELEASE

wxEVT_STC_MARGIN_RIGHT_CLICK

getDragAllowMove(This)

-spec getDragAllowMove(This) -> boolean() when This :: wxStyledTextEvent().

getDragResult(This)

-spec getDragResult(This) -> wx:wx_enum() when This :: wxStyledTextEvent().

Returns drag result for this event.

This method is valid for wxEVT_STC_DRAG_OVER and wxEVT_STC_DO_DROP events.

getDragText(This)

-spec getDragText(This) -> unicode:charlist() when This :: wxStyledTextEvent().

Deprecated:

Use wxCommandEvent:getString/1 instead.

getFoldLevelNow(This)

-spec getFoldLevelNow(This) -> integer() when This :: wxStyledTextEvent().

Returns the current fold level for the line.

This method is valid for wxEVT_STC_MODIFIED events when the result of getModificationType/1 includes ?wxSTC_MOD_CHANGEFOLD.

getFoldLevelPrev(This)

-spec getFoldLevelPrev(This) -> integer() when This :: wxStyledTextEvent().

Returns previous fold level for the line.

This method is valid for wxEVT_STC_MODIFIED events when the result of getModificationType/1 includes ?wxSTC_MOD_CHANGEFOLD.

getKey(This)

-spec getKey(This) -> integer() when This :: wxStyledTextEvent().

Returns the key code of the key that generated this event.

This method is valid for the following event types:

wxEVT_STC_CHARADDED

wxEVT_STC_USERLISTSELECTION

wxEVT_STC_AUTOCOMP_SELECTION

wxEVT_STC_AUTOCOMP_COMPLETED

getLength(This)

-spec getLength(This) -> integer() when This :: wxStyledTextEvent().

Returns the length (number of characters) of this event.

This method is valid for wxEVT_STC_MODIFIED and wxEVT_STC_NEEDSHOWN events.

getLine(This)

-spec getLine(This) -> integer() when This :: wxStyledTextEvent().

Returns zero-based line number for this event.

This method is valid for wxEVT_STC_DOUBLECLICK and wxEVT_STC_MODIFIED events.

getLinesAdded(This)

-spec getLinesAdded(This) -> integer() when This :: wxStyledTextEvent().

Returns the number of lines added or deleted with this event.

This method is valid for wxEVT_STC_MODIFIED events when the result of getModificationType/1 includes ?wxSTC_MOD_INSERTTEXT or ?wxSTC_MOD_DELETETEXT.

getListType(This)

-spec getListType(This) -> integer() when This :: wxStyledTextEvent().

Returns the list type for this event.

The list type is an integer passed to a list when it is created with the wxStyledTextCtrl:userListShow/3 method and can be used to distinguish lists if more than one is used.

This method is valid for wxEVT_STC_AUTOCOMP_SELECTION_CHANGE and wxEVT_STC_USERLISTSELECTION events.

getLParam(This)

-spec getLParam(This) -> integer() when This :: wxStyledTextEvent().

Returns the value of the LParam field for this event.

This method is valid for wxEVT_STC_MACRORECORD events.

getMargin(This)

-spec getMargin(This) -> integer() when This :: wxStyledTextEvent().

Returns the zero-based index of the margin that generated this event.

This method is valid for wxEVT_STC_MARGINCLICK and wxEVT_STC_MARGIN_RIGHT_CLICK events.

getMessage(This)

-spec getMessage(This) -> integer() when This :: wxStyledTextEvent().

Returns a message number while a macro is being recorded.

Many of the wxStyledTextCtrl methods such as wxStyledTextCtrl:insertText/3 and wxStyledTextCtrl:paste/1 have an event number associated with them. This method returns that number while a macro is being recorded so that the macro can be played back later.

This method is valid for wxEVT_STC_MACRORECORD events.

getModificationType(This)

-spec getModificationType(This) -> integer() when This :: wxStyledTextEvent().

Returns the modification type for this event.

The modification type is a bit list that describes the change that generated this event. It may contain one or more of the following values:

?wxSTC_MOD_INSERTTEXT

?wxSTC_MOD_DELETETEXT

?wxSTC_MOD_CHANGESTYLE

?wxSTC_MOD_CHANGEFOLD

?wxSTC_PERFORMED_USER

?wxSTC_PERFORMED_UNDO

?wxSTC_PERFORMED_REDO

?wxSTC_MULTISTEPUNDOREDO

?wxSTC_LASTSTEPINUNDOREDO

?wxSTC_MOD_CHANGEMARKER

?wxSTC_MOD_BEFOREINSERT

?wxSTC_MOD_BEFOREDELETE

?wxSTC_MULTILINEUNDOREDO

?wxSTC_STARTACTION

?wxSTC_MOD_CHANGEINDICATOR

?wxSTC_MOD_CHANGELINESTATE

?wxSTC_MOD_CHANGEMARGIN

?wxSTC_MOD_CHANGEANNOTATION

?wxSTC_MOD_CONTAINER

?wxSTC_MOD_LEXERSTATE

?wxSTC_MOD_INSERTCHECK

?wxSTC_MOD_CHANGETABSTOPS

This method is valid for wxEVT_STC_MODIFIED events.

getModifiers(This)

-spec getModifiers(This) -> integer() when This :: wxStyledTextEvent().

Returns the modifiers of the key press or mouse click for this event.

The returned value is a bit list that may contain one or more of the following values:

?wxSTC_KEYMOD_SHIFT

?wxSTC_KEYMOD_CTRL

?wxSTC_KEYMOD_ALT

?wxSTC_KEYMOD_SUPER

?wxSTC_KEYMOD_META

In addition, the value can be checked for equality with ?wxSTC_KEYMOD_NORM to test if no modifiers are present.

This method is valid for the following event types:

wxEVT_STC_DOUBLECLICK

wxEVT_STC_MARGINCLICK

wxEVT_STC_HOTSPOT_CLICK

wxEVT_STC_HOTSPOT_DCLICK

wxEVT_STC_HOTSPOT_RELEASE_CLICK

wxEVT_STC_INDICATOR_CLICK

wxEVT_STC_INDICATOR_RELEASE

wxEVT_STC_MARGIN_RIGHT_CLICK

getPosition(This)

-spec getPosition(This) -> integer() when This :: wxStyledTextEvent().

Returns the zero-based text position associated this event.

This method is valid for the following event types:

wxEVT_STC_STYLENEEDED

wxEVT_STC_DOUBLECLICK

wxEVT_STC_MODIFIED

wxEVT_STC_MARGINCLICK

wxEVT_STC_NEEDSHOWN

wxEVT_STC_USERLISTSELECTION

wxEVT_STC_DWELLSTART

wxEVT_STC_DWELLEND

wxEVT_STC_HOTSPOT_CLICK

wxEVT_STC_HOTSPOT_DCLICK

wxEVT_STC_HOTSPOT_RELEASE_CLICK

wxEVT_STC_INDICATOR_CLICK

wxEVT_STC_INDICATOR_RELEASE

wxEVT_STC_CALLTIP_CLICK

wxEVT_STC_AUTOCOMP_SELECTION

wxEVT_STC_AUTOCOMP_SELECTION_CHANGE

wxEVT_STC_AUTOCOMP_COMPLETED

wxEVT_STC_MARGIN_RIGHT_CLICK

getShift(This)

-spec getShift(This) -> boolean() when This :: wxStyledTextEvent().

Returns true if the Shift key is pressed.

This method is valid for the following event types:

wxEVT_STC_DOUBLECLICK

wxEVT_STC_MARGINCLICK

wxEVT_STC_HOTSPOT_CLICK

wxEVT_STC_HOTSPOT_DCLICK

wxEVT_STC_HOTSPOT_RELEASE_CLICK

wxEVT_STC_INDICATOR_CLICK

wxEVT_STC_INDICATOR_RELEASE

wxEVT_STC_MARGIN_RIGHT_CLICK

getText(This)

-spec getText(This) -> unicode:charlist() when This :: wxStyledTextEvent().

Deprecated:

Use wxCommandEvent:getString/1 instead.

getWParam(This)

-spec getWParam(This) -> integer() when This :: wxStyledTextEvent().

Returns value of the WParam field for this event.

This method is valid for wxEVT_STC_MACRORECORD events.

getX(This)

-spec getX(This) -> integer() when This :: wxStyledTextEvent().

Returns the X coordinate of the mouse for this event.

This method is valid for the following event types:

wxEVT_STC_DWELLSTART

wxEVT_STC_DWELLEND

wxEVT_STC_START_DRAG

wxEVT_STC_DRAG_OVER

wxEVT_STC_DO_DROP

getY(This)

-spec getY(This) -> integer() when This :: wxStyledTextEvent().

Returns the Y coordinate of the mouse for this event.

This method is valid for the following event types:

wxEVT_STC_DWELLSTART

wxEVT_STC_DWELLEND

wxEVT_STC_START_DRAG

wxEVT_STC_DRAG_OVER

wxEVT_STC_DO_DROP