wxSysColourChangedEvent (wx v2.5)

This class is used for system colour change events, which are generated when the user changes the colour settings using the control panel.

This is only appropriate under Windows.

Remark: The default event handler for this event propagates the event to child windows, since Windows only sends the events to top-level windows. If intercepting this event for a top-level window, remember to call the base class handler, or to pass the event on to the window's children explicitly.

See: Overview events

This class is derived, and can use functions, from:

wxEvent

wxWidgets docs: wxSysColourChangedEvent

Events

Use wxEvtHandler:connect/3 with wxSysColourChangedEventType to subscribe to events of this type.

Types

wxSysColourChanged()

-type wxSysColourChanged() ::
          #wxSysColourChanged{type :: wxSysColourChangedEvent:wxSysColourChangedEventType()}.

wxSysColourChangedEvent()

-type wxSysColourChangedEvent() :: wx:wx_object().

wxSysColourChangedEventType()

-type wxSysColourChangedEventType() :: sys_colour_changed.