wxUpdateUIEvent (wx v2.3.1)

Functions for wxUpdateUIEvent class

This class is used for pseudo-events which are called by wxWidgets to give an application the chance to update various user interface elements.

Without update UI events, an application has to work hard to check/uncheck, enable/disable, show/hide, and set the text for elements such as menu items and toolbar buttons. The code for doing this has to be mixed up with the code that is invoked when an action is invoked for a menu item or button.

With update UI events, you define an event handler to look at the state of the application and change UI elements accordingly. wxWidgets will call your member functions in idle time, so you don't have to worry where to call this code.

In addition to being a clearer and more declarative method, it also means you don't have to worry whether you're updating a toolbar or menubar identifier. The same handler can update a menu item and toolbar button, if the identifier is the same. Instead of directly manipulating the menu or button, you call functions in the event object, such as check/2. wxWidgets will determine whether such a call has been made, and which UI element to update.

These events will work for popup menus as well as menubars. Just before a menu is popped up, wxMenu::UpdateUI (not implemented in wx) is called to process any UI events for the window that owns the menu.

If you find that the overhead of UI update processing is affecting your application, you can do one or both of the following:

Note that although events are sent in idle time, defining a wxIdleEvent handler for a window does not affect this because the events are sent from wxWindow::OnInternalIdle (not implemented in wx) which is always called in idle time.

wxWidgets tries to optimize update events on some platforms. On Windows and GTK+, events for menubar items are only sent when the menu is about to be shown, and not in idle time.

See: Overview events

This class is derived (and can use functions) from: wxCommandEvent wxEvent

wxWidgets docs: wxUpdateUIEvent

Events

Use wxEvtHandler:connect/3 with wxUpdateUIEventType to subscribe to events of this type.

Types

event()

-type event() ::
    wxActivateEvent:wxActivate() |
    wxAuiManagerEvent:wxAuiManager() |
    wxAuiNotebookEvent:wxAuiNotebook() |
    wxBookCtrlEvent:wxBookCtrl() |
    wxCalendarEvent:wxCalendar() |
    wxChildFocusEvent:wxChildFocus() |
    wxClipboardTextEvent:wxClipboardText() |
    wxCloseEvent:wxClose() |
    wxColourPickerEvent:wxColourPicker() |
    wxCommandEvent:wxCommand() |
    wxContextMenuEvent:wxContextMenu() |
    wxDateEvent:wxDate() |
    wxDisplayChangedEvent:wxDisplayChanged() |
    wxDropFilesEvent:wxDropFiles() |
    wxEraseEvent:wxErase() |
    wxFileDirPickerEvent:wxFileDirPicker() |
    wxFocusEvent:wxFocus() |
    wxFontPickerEvent:wxFontPicker() |
    wxGridEvent:wxGrid() |
    wxHelpEvent:wxHelp() |
    wxHtmlLinkEvent:wxHtmlLink() |
    wxIconizeEvent:wxIconize() |
    wxIdleEvent:wxIdle() |
    wxInitDialogEvent:wxInitDialog() |
    wxJoystickEvent:wxJoystick() |
    wxKeyEvent:wxKey() |
    wxListEvent:wxList() |
    wxMaximizeEvent:wxMaximize() |
    wxMenuEvent:wxMenu() |
    wxMouseCaptureChangedEvent:wxMouseCaptureChanged() |
    wxMouseCaptureLostEvent:wxMouseCaptureLost() |
    wxMouseEvent:wxMouse() |
    wxMoveEvent:wxMove() |
    wxNavigationKeyEvent:wxNavigationKey() |
    wxPaintEvent:wxPaint() |
    wxPaletteChangedEvent:wxPaletteChanged() |
    wxQueryNewPaletteEvent:wxQueryNewPalette() |
    wxSashEvent:wxSash() |
    wxScrollEvent:wxScroll() |
    wxScrollWinEvent:wxScrollWin() |
    wxSetCursorEvent:wxSetCursor() |
    wxShowEvent:wxShow() |
    wxSizeEvent:wxSize() |
    wxSpinEvent:wxSpin() |
    wxSplitterEvent:wxSplitter() |
    wxStyledTextEvent:wxStyledText() |
    wxSysColourChangedEvent:wxSysColourChanged() |
    wxTaskBarIconEvent:wxTaskBarIcon() |
    wxTreeEvent:wxTree() |
    wxUpdateUIEvent:wxUpdateUI() |
    wxWebViewEvent:wxWebView() |
    wxWindowCreateEvent:wxWindowCreate() |
    wxWindowDestroyEvent:wxWindowDestroy().

wx()

-type wx() :: #wx{}.

wxActivate()

-type wxActivate() :: #wxActivate{}.

wxAuiManager()

-type wxAuiManager() :: #wxAuiManager{}.

wxAuiNotebook()

-type wxAuiNotebook() :: #wxAuiNotebook{}.

wxBookCtrl()

-type wxBookCtrl() :: #wxBookCtrl{}.

wxCalendar()

-type wxCalendar() :: #wxCalendar{}.

wxChildFocus()

-type wxChildFocus() :: #wxChildFocus{}.

wxClipboardText()

-type wxClipboardText() :: #wxClipboardText{}.

wxClose()

-type wxClose() :: #wxClose{}.

wxColourPicker()

-type wxColourPicker() :: #wxColourPicker{}.

wxCommand()

-type wxCommand() :: #wxCommand{}.

wxContextMenu()

-type wxContextMenu() :: #wxContextMenu{}.

wxDate()

-type wxDate() :: #wxDate{}.

wxDisplayChanged()

-type wxDisplayChanged() :: #wxDisplayChanged{}.

wxDropFiles()

-type wxDropFiles() :: #wxDropFiles{}.

wxErase()

-type wxErase() :: #wxErase{}.

wxEventType()

-type wxEventType() ::
    wxActivateEvent:wxActivateEventType() |
    wxAuiManagerEvent:wxAuiManagerEventType() |
    wxAuiNotebookEvent:wxAuiNotebookEventType() |
    wxBookCtrlEvent:wxBookCtrlEventType() |
    wxCalendarEvent:wxCalendarEventType() |
    wxChildFocusEvent:wxChildFocusEventType() |
    wxClipboardTextEvent:wxClipboardTextEventType() |
    wxCloseEvent:wxCloseEventType() |
    wxColourPickerEvent:wxColourPickerEventType() |
    wxCommandEvent:wxCommandEventType() |
    wxContextMenuEvent:wxContextMenuEventType() |
    wxDateEvent:wxDateEventType() |
    wxDisplayChangedEvent:wxDisplayChangedEventType() |
    wxDropFilesEvent:wxDropFilesEventType() |
    wxEraseEvent:wxEraseEventType() |
    wxFileDirPickerEvent:wxFileDirPickerEventType() |
    wxFocusEvent:wxFocusEventType() |
    wxFontPickerEvent:wxFontPickerEventType() |
    wxGridEvent:wxGridEventType() |
    wxHelpEvent:wxHelpEventType() |
    wxHtmlLinkEvent:wxHtmlLinkEventType() |
    wxIconizeEvent:wxIconizeEventType() |
    wxIdleEvent:wxIdleEventType() |
    wxInitDialogEvent:wxInitDialogEventType() |
    wxJoystickEvent:wxJoystickEventType() |
    wxKeyEvent:wxKeyEventType() |
    wxListEvent:wxListEventType() |
    wxMaximizeEvent:wxMaximizeEventType() |
    wxMenuEvent:wxMenuEventType() |
    wxMouseCaptureChangedEvent:wxMouseCaptureChangedEventType() |
    wxMouseCaptureLostEvent:wxMouseCaptureLostEventType() |
    wxMouseEvent:wxMouseEventType() |
    wxMoveEvent:wxMoveEventType() |
    wxNavigationKeyEvent:wxNavigationKeyEventType() |
    wxPaintEvent:wxPaintEventType() |
    wxPaletteChangedEvent:wxPaletteChangedEventType() |
    wxQueryNewPaletteEvent:wxQueryNewPaletteEventType() |
    wxSashEvent:wxSashEventType() |
    wxScrollEvent:wxScrollEventType() |
    wxScrollWinEvent:wxScrollWinEventType() |
    wxSetCursorEvent:wxSetCursorEventType() |
    wxShowEvent:wxShowEventType() |
    wxSizeEvent:wxSizeEventType() |
    wxSpinEvent:wxSpinEventType() |
    wxSplitterEvent:wxSplitterEventType() |
    wxStyledTextEvent:wxStyledTextEventType() |
    wxSysColourChangedEvent:wxSysColourChangedEventType() |
    wxTaskBarIconEvent:wxTaskBarIconEventType() |
    wxTreeEvent:wxTreeEventType() |
    wxUpdateUIEvent:wxUpdateUIEventType() |
    wxWebViewEvent:wxWebViewEventType() |
    wxWindowCreateEvent:wxWindowCreateEventType() |
    wxWindowDestroyEvent:wxWindowDestroyEventType().

wxFileDirPicker()

-type wxFileDirPicker() :: #wxFileDirPicker{}.

wxFocus()

-type wxFocus() :: #wxFocus{}.

wxFontPicker()

-type wxFontPicker() :: #wxFontPicker{}.

wxGrid()

-type wxGrid() :: #wxGrid{}.

wxHelp()

-type wxHelp() :: #wxHelp{}.

wxHtmlLink()

-type wxHtmlLink() :: #wxHtmlLink{}.

wxIconize()

-type wxIconize() :: #wxIconize{}.

wxIdle()

-type wxIdle() :: #wxIdle{}.

wxInitDialog()

-type wxInitDialog() :: #wxInitDialog{}.

wxJoystick()

-type wxJoystick() :: #wxJoystick{}.

wxKey()

-type wxKey() :: #wxKey{}.

wxList()

-type wxList() :: #wxList{}.

wxMaximize()

-type wxMaximize() :: #wxMaximize{}.

wxMenu()

-type wxMenu() :: #wxMenu{}.

wxMouse()

-type wxMouse() :: #wxMouse{}.

wxMouseCaptureChanged()

-type wxMouseCaptureChanged() :: #wxMouseCaptureChanged{}.

wxMouseCaptureLost()

-type wxMouseCaptureLost() :: #wxMouseCaptureLost{}.

wxMove()

-type wxMove() :: #wxMove{}.

wxNavigationKey()

-type wxNavigationKey() :: #wxNavigationKey{}.

wxPaint()

-type wxPaint() :: #wxPaint{}.

wxPaletteChanged()

-type wxPaletteChanged() :: #wxPaletteChanged{}.

wxQueryNewPalette()

-type wxQueryNewPalette() :: #wxQueryNewPalette{}.

wxSash()

-type wxSash() :: #wxSash{}.

wxScroll()

-type wxScroll() :: #wxScroll{}.

wxScrollWin()

-type wxScrollWin() :: #wxScrollWin{}.

wxSetCursor()

-type wxSetCursor() :: #wxSetCursor{}.

wxShow()

-type wxShow() :: #wxShow{}.

wxSize()

-type wxSize() :: #wxSize{}.

wxSpin()

-type wxSpin() :: #wxSpin{}.

wxSplitter()

-type wxSplitter() :: #wxSplitter{}.

wxStyledText()

-type wxStyledText() :: #wxStyledText{}.

wxSysColourChanged()

-type wxSysColourChanged() :: #wxSysColourChanged{}.

wxTaskBarIcon()

-type wxTaskBarIcon() :: #wxTaskBarIcon{}.

wxTree()

-type wxTree() :: #wxTree{}.

wxUpdateUI()

-type wxUpdateUI() :: #wxUpdateUI{}.

wxUpdateUIEvent()

-type wxUpdateUIEvent() :: wx:wx_object().

wxUpdateUIEventType()

-type wxUpdateUIEventType() :: update_ui.

wxWebView()

-type wxWebView() :: #wxWebView{}.

wxWindowCreate()

-type wxWindowCreate() :: #wxWindowCreate{}.

wxWindowDestroy()

-type wxWindowDestroy() :: #wxWindowDestroy{}.

Functions

canUpdate/1

-spec canUpdate(Window) -> boolean() when Window :: wxWindow:wxWindow().

Returns true if it is appropriate to update (send UI update events to) this window.

This function looks at the mode used (see setMode/1), the wxWS_EX_PROCESS_UI_UPDATES flag in window, the time update events were last sent in idle time, and the update interval, to determine whether events should be sent to this window now. By default this will always return true because the update mode is initially wxUPDATE_UI_PROCESS_ALL and the interval is set to 0; so update events will be sent as often as possible. You can reduce the frequency that events are sent by changing the mode and/or setting an update interval.

See: resetUpdateTime/0, setUpdateInterval/1, setMode/1

check/2

-spec check(This, Check) -> ok when This :: wxUpdateUIEvent(), Check :: boolean().

Check or uncheck the UI element.

enable/2

-spec enable(This, Enable) -> ok when This :: wxUpdateUIEvent(), Enable :: boolean().

Enable or disable the UI element.

getChecked/1

-spec getChecked(This) -> boolean() when This :: wxUpdateUIEvent().

Returns true if the UI element should be checked.

getEnabled/1

-spec getEnabled(This) -> boolean() when This :: wxUpdateUIEvent().

Returns true if the UI element should be enabled.

getMode()

-spec getMode() -> wx:wx_enum().

Static function returning a value specifying how wxWidgets will send update events: to all windows, or only to those which specify that they will process the events.

See: setMode/1

getSetChecked/1

-spec getSetChecked(This) -> boolean() when This :: wxUpdateUIEvent().

Returns true if the application has called check/2.

For wxWidgets internal use only.

getSetEnabled/1

-spec getSetEnabled(This) -> boolean() when This :: wxUpdateUIEvent().

Returns true if the application has called enable/2.

For wxWidgets internal use only.

getSetShown/1

-spec getSetShown(This) -> boolean() when This :: wxUpdateUIEvent().

Returns true if the application has called show/2.

For wxWidgets internal use only.

getSetText/1

-spec getSetText(This) -> boolean() when This :: wxUpdateUIEvent().

Returns true if the application has called setText/2.

For wxWidgets internal use only.

getShown/1

-spec getShown(This) -> boolean() when This :: wxUpdateUIEvent().

Returns true if the UI element should be shown.

getText/1

-spec getText(This) -> unicode:charlist() when This :: wxUpdateUIEvent().

Returns the text that should be set for the UI element.

getUpdateInterval()

-spec getUpdateInterval() -> integer().

Returns the current interval between updates in milliseconds.

The value -1 disables updates, 0 updates as frequently as possible.

See: setUpdateInterval/1

resetUpdateTime()

-spec resetUpdateTime() -> ok.

Used internally to reset the last-updated time to the current time.

It is assumed that update events are normally sent in idle time, so this is called at the end of idle processing.

See: canUpdate/1, setUpdateInterval/1, setMode/1

setMode(Mode)

-spec setMode(Mode) -> ok when Mode :: wx:wx_enum().

Specify how wxWidgets will send update events: to all windows, or only to those which specify that they will process the events.

setText/2

-spec setText(This, Text) -> ok when This :: wxUpdateUIEvent(), Text :: unicode:chardata().

Sets the text for this UI element.

setUpdateInterval(UpdateInterval)

-spec setUpdateInterval(UpdateInterval) -> ok when UpdateInterval :: integer().

Sets the interval between updates in milliseconds.

Set to -1 to disable updates, or to 0 to update as frequently as possible. The default is 0.

Use this to reduce the overhead of UI update events if your application has a lot of windows. If you set the value to -1 or greater than 0, you may also need to call wxWindow:updateWindowUI/2 at appropriate points in your application, such as when a dialog is about to be shown.

show/2

-spec show(This, Show) -> ok when This :: wxUpdateUIEvent(), Show :: boolean().

Show or hide the UI element.