wxWindowDC (wx v2.5)

A wxWindowDC must be constructed if an application wishes to paint on the whole area of a window (client and decorations).

This should normally be constructed as a temporary stack object; don't store a wxWindowDC object.

To draw on a window from inside an EVT_PAINT() handler, construct a wxPaintDC object instead.

To draw on the client area of a window from outside an EVT_PAINT() handler, construct a wxClientDC object.

A wxWindowDC object is initialized to use the same font and colours as the window it is associated with.

See:

wxDC

wxMemoryDC

wxPaintDC

wxClientDC

wxScreenDC

This class is derived, and can use functions, from:

wxDC

wxWidgets docs: wxWindowDC

Types

wxWindowDC()

-type wxWindowDC() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxWindowDC()) -> ok.

Destroys the object

new(Window)

-spec new(Window) -> wxWindowDC() when Window :: wxWindow:wxWindow().

Constructor.

Pass a pointer to the window on which you wish to paint.