wxAcceleratorEntry (wx v2.4.3)

An object used by an application wishing to create an accelerator table (see wxAcceleratorTable).

See:

wxAcceleratorTable

wxWindow:setAcceleratorTable/2

wxWidgets docs: wxAcceleratorEntry

Types

wxAcceleratorEntry()

-type wxAcceleratorEntry() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxAcceleratorEntry()) -> ok.

Destroys the object

getCommand(This)

-spec getCommand(This) -> integer() when This :: wxAcceleratorEntry().

Returns the command identifier for the accelerator table entry.

getFlags(This)

-spec getFlags(This) -> integer() when This :: wxAcceleratorEntry().

Returns the flags for the accelerator table entry.

getKeyCode(This)

-spec getKeyCode(This) -> integer() when This :: wxAcceleratorEntry().

Returns the keycode for the accelerator table entry.

new()

-spec new() -> wxAcceleratorEntry().

Equivalent to new([]).

new/1

-spec new([Option]) -> wxAcceleratorEntry()
             when
                 Option ::
                     {flags, integer()} |
                     {keyCode, integer()} |
                     {cmd, integer()} |
                     {item, wxMenuItem:wxMenuItem()};
         (Entry) -> wxAcceleratorEntry() when Entry :: wxAcceleratorEntry().

Copy ctor.

set(This, Flags, KeyCode, Cmd)

-spec set(This, Flags, KeyCode, Cmd) -> ok
             when
                 This :: wxAcceleratorEntry(),
                 Flags :: integer(),
                 KeyCode :: integer(),
                 Cmd :: integer().

Equivalent to set(This, Flags, KeyCode, Cmd, []).

set/5

-spec set(This, Flags, KeyCode, Cmd, [Option]) -> ok
             when
                 This :: wxAcceleratorEntry(),
                 Flags :: integer(),
                 KeyCode :: integer(),
                 Cmd :: integer(),
                 Option :: {item, wxMenuItem:wxMenuItem()}.

Sets the accelerator entry parameters.