wxAcceleratorEntry (wx v2.0.1)

An object used by an application wishing to create an accelerator table (see wxAcceleratorTable).

See: wxAcceleratorTable, wxWindow:setAcceleratorTable/2

wxWidgets docs: wxAcceleratorEntry

Types

-type wxAcceleratorEntry() :: term().

Specs

wxAcceleratorEntry() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxAcceleratorEntry()) -> ok.

Destroys the object.

getCommand/1

Specs

getCommand(This) -> integer() when This :: wxAcceleratorEntry().

Returns the command identifier for the accelerator table entry.

getFlags/1

Specs

getFlags(This) -> integer() when This :: wxAcceleratorEntry().

Returns the flags for the accelerator table entry.

getKeyCode/1

Specs

getKeyCode(This) -> integer() when This :: wxAcceleratorEntry().

Returns the keycode for the accelerator table entry.

new/0

Specs

new() -> wxAcceleratorEntry().

new/1

Specs

new([Option]) -> wxAcceleratorEntry()
       when
           Option ::
               {flags, integer()} |
               {keyCode, integer()} |
               {cmd, integer()} |
               {item, wxMenuItem:wxMenuItem()};
   (Entry) -> wxAcceleratorEntry() when Entry :: wxAcceleratorEntry().

Copy ctor.

set/4

Specs

set(This, Flags, KeyCode, Cmd) -> ok
       when
           This :: wxAcceleratorEntry(),
           Flags :: integer(),
           KeyCode :: integer(),
           Cmd :: integer().

set/5

Specs

set(This, Flags, KeyCode, Cmd, [Option]) -> ok
       when
           This :: wxAcceleratorEntry(),
           Flags :: integer(),
           KeyCode :: integer(),
           Cmd :: integer(),
           Option :: {item, wxMenuItem:wxMenuItem()}.

Sets the accelerator entry parameters.