wxAcceleratorTable (wx v2.0.1)

An accelerator table allows the application to specify a table of keyboard shortcuts for menu or button commands.

The object ?wxNullAcceleratorTable is defined to be a table with no data, and is the initial accelerator table for a window.

Example:

Remark: An accelerator takes precedence over normal processing and can be a convenient way to program some event handling. For example, you can use an accelerator table to enable a dialog with a multi-line text control to accept CTRL-Enter as meaning 'OK'.

Predefined objects (include wx.hrl): ?wxNullAcceleratorTable

See: wxAcceleratorEntry, wxWindow:setAcceleratorTable/2

wxWidgets docs: wxAcceleratorTable

Types

-type wxAcceleratorTable() :: term().

Specs

wxAcceleratorTable() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxAcceleratorTable()) -> ok.

Destroys the wxAcceleratorTable object.

See overview_refcount_destruct for more info.

isOk/1

Specs

isOk(This) -> boolean() when This :: wxAcceleratorTable().

Returns true if the accelerator table is valid.

new/0

Specs

new() -> wxAcceleratorTable().

Default ctor.

new/2

Specs

new(N, Entries) -> wxAcceleratorTable()
       when N :: integer(), Entries :: [wxAcceleratorEntry:wxAcceleratorEntry()].

Initializes the accelerator table from an array of wxAcceleratorEntry.

ok/1

Specs

ok(This) -> boolean() when This :: wxAcceleratorTable().

See: isOk/1.