wxAuiDockArt (wx v2.4.3)

wxAuiDockArt is part of the wxAUI class framework.

See also overview_aui.

wxAuiDockArt is the art provider: provides all drawing functionality to the wxAui dock manager. This allows the dock manager to have a pluggable look-and-feel.

By default, a wxAuiManager uses an instance of this class called wxAuiDefaultDockArt (not implemented in wx) which provides bitmap art and a colour scheme that is adapted to the major platforms' look. You can either derive from that class to alter its behaviour or write a completely new dock art class. Call wxAuiManager:setArtProvider/2 to force wxAUI to use your new dock art provider.

See:

wxAuiManager

wxAuiPaneInfo

wxWidgets docs: wxAuiDockArt

Types

wxAuiDockArt()

-type wxAuiDockArt() :: wx:wx_object().

Functions

getColour(This, Id)

-spec getColour(This, Id) -> wx:wx_colour4() when This :: wxAuiDockArt(), Id :: integer().

Get the colour of a certain setting.

id can be one of the colour values of wxAuiPaneDockArtSetting.

getFont(This, Id)

-spec getFont(This, Id) -> wxFont:wxFont() when This :: wxAuiDockArt(), Id :: integer().

Get a font setting.

getMetric(This, Id)

-spec getMetric(This, Id) -> integer() when This :: wxAuiDockArt(), Id :: integer().

Get the value of a certain setting.

id can be one of the size values of wxAuiPaneDockArtSetting.

setColour(This, Id, Colour)

-spec setColour(This, Id, Colour) -> ok
                   when This :: wxAuiDockArt(), Id :: integer(), Colour :: wx:wx_colour().

Set a certain setting with the value colour.

id can be one of the colour values of wxAuiPaneDockArtSetting.

setFont(This, Id, Font)

-spec setFont(This, Id, Font) -> ok
                 when This :: wxAuiDockArt(), Id :: integer(), Font :: wxFont:wxFont().

Set a font setting.

setMetric(This, Id, New_val)

-spec setMetric(This, Id, New_val) -> ok
                   when This :: wxAuiDockArt(), Id :: integer(), New_val :: integer().

Set a certain setting with the value new\_val.

id can be one of the size values of wxAuiPaneDockArtSetting.