wxAuiDockArt (wx v2.0.1)

wxAuiDockArt is part of the wxAUI class framework. See also overview_aui.

wxAuiDockArt is the art provider: provides all drawing functionality to the wxAui dock manager. This allows the dock manager to have a pluggable look-and-feel.

By default, a wxAuiManager uses an instance of this class called wxAuiDefaultDockArt (not implemented in wx) which provides bitmap art and a colour scheme that is adapted to the major platforms' look. You can either derive from that class to alter its behaviour or write a completely new dock art class. Call wxAuiManager:setArtProvider/2 to force wxAUI to use your new dock art provider.

See: wxAuiManager, wxAuiPaneInfo

wxWidgets docs: wxAuiDockArt

Types

-type wxAuiDockArt() :: term().

Specs

wxAuiDockArt() :: wx:wx_object().

Functions

getColour/2

Specs

getColour(This, Id) -> wx:wx_colour4() when This :: wxAuiDockArt(), Id :: integer().

Get the colour of a certain setting.

id can be one of the colour values of wxAuiPaneDockArtSetting.

getFont/2

Specs

getFont(This, Id) -> wxFont:wxFont() when This :: wxAuiDockArt(), Id :: integer().

Get a font setting.

getMetric/2

Specs

getMetric(This, Id) -> integer() when This :: wxAuiDockArt(), Id :: integer().

Get the value of a certain setting.

id can be one of the size values of wxAuiPaneDockArtSetting.

setColour/3

Specs

setColour(This, Id, Colour) -> ok
             when This :: wxAuiDockArt(), Id :: integer(), Colour :: wx:wx_colour().

Set a certain setting with the value colour.

id can be one of the colour values of wxAuiPaneDockArtSetting.

setFont/3

Specs

setFont(This, Id, Font) -> ok
           when This :: wxAuiDockArt(), Id :: integer(), Font :: wxFont:wxFont().

Set a font setting.

setMetric/3

Specs

setMetric(This, Id, New_val) -> ok
             when This :: wxAuiDockArt(), Id :: integer(), New_val :: integer().

Set a certain setting with the value new_val.

id can be one of the size values of wxAuiPaneDockArtSetting.