wxAuiManager (wx v2.0.1)

wxAuiManager is the central class of the wxAUI class framework.

wxAuiManager manages the panes associated with it for a particular wxFrame, using a pane's wxAuiPaneInfo information to determine each pane's docking and floating behaviour.

wxAuiManager uses wxWidgets' sizer mechanism to plan the layout of each frame. It uses a replaceable dock art class to do all drawing, so all drawing is localized in one area, and may be customized depending on an application's specific needs.

wxAuiManager works as follows: the programmer adds panes to the class, or makes changes to existing pane properties (dock position, floating state, show state, etc.). To apply these changes, wxAuiManager's update/1 function is called. This batch processing can be used to avoid flicker, by modifying more than one pane at a time, and then "committing" all of the changes at once by calling update/1.

Panes can be added quite easily:

Later on, the positions can be modified easily. The following will float an existing pane in a tool window:

Layers, Rows and Directions, Positions

Inside wxAUI, the docking layout is figured out by checking several pane parameters. Four of these are important for determining where a pane will end up:

Styles

This class supports the following styles:

See: Overview aui, wxAuiNotebook, wxAuiDockArt, wxAuiPaneInfo

This class is derived (and can use functions) from: wxEvtHandler

wxWidgets docs: wxAuiManager

Events

Event types emitted from this class: aui_pane_button, aui_pane_close, aui_pane_maximize, aui_pane_restore, aui_pane_activated, aui_render

Types

-type wxAuiManager() :: term().

Specs

wxAuiManager() :: wx:wx_object().

Functions

addPane/2

Specs

addPane(This, Window) -> boolean() when This :: wxAuiManager(), Window :: wxWindow:wxWindow().

addPane/3

Specs

addPane(This, Window, [Option]) -> boolean()
           when
               This :: wxAuiManager(),
               Window :: wxWindow:wxWindow(),
               Option :: {direction, integer()} | {caption, unicode:chardata()};
       (This, Window, Pane_info) -> boolean()
           when
               This :: wxAuiManager(),
               Window :: wxWindow:wxWindow(),
               Pane_info :: wxAuiPaneInfo:wxAuiPaneInfo().

addPane/4 tells the frame manager to start managing a child window.

There are several versions of this function. The first version allows the full spectrum of pane parameter possibilities. The second version is used for simpler user interfaces which do not require as much configuration. The last version allows a drop position to be specified, which will determine where the pane will be added.

addPane/4

Specs

addPane(This, Window, Pane_info, Drop_pos) -> boolean()
           when
               This :: wxAuiManager(),
               Window :: wxWindow:wxWindow(),
               Pane_info :: wxAuiPaneInfo:wxAuiPaneInfo(),
               Drop_pos :: {X :: integer(), Y :: integer()}.

destroy/1

Specs

destroy(This :: wxAuiManager()) -> ok.

Dtor.

detachPane/2

Specs

detachPane(This, Window) -> boolean() when This :: wxAuiManager(), Window :: wxWindow:wxWindow().

Tells the wxAuiManager to stop managing the pane specified by window.

The window, if in a floated frame, is reparented to the frame managed by wxAuiManager.

getAllPanes/1

Specs

getAllPanes(This) -> [wxAuiPaneInfo:wxAuiPaneInfo()] when This :: wxAuiManager().

Returns an array of all panes managed by the frame manager.

getArtProvider/1

Specs

getArtProvider(This) -> wxAuiDockArt:wxAuiDockArt() when This :: wxAuiManager().

Returns the current art provider being used.

See: wxAuiDockArt

getDockSizeConstraint/1

Specs

getDockSizeConstraint(This) -> {Widthpct :: number(), Heightpct :: number()}
                         when This :: wxAuiManager().

Returns the current dock constraint values.

See setDockSizeConstraint/3 for more information.

getFlags/1

Specs

getFlags(This) -> integer() when This :: wxAuiManager().

Returns the current ?wxAuiManagerOption's flags.

getManagedWindow/1

Specs

getManagedWindow(This) -> wxWindow:wxWindow() when This :: wxAuiManager().

Returns the frame currently being managed by wxAuiManager.

getManager/1

Specs

getManager(Window) -> wxAuiManager() when Window :: wxWindow:wxWindow().

Calling this method will return the wxAuiManager for a given window.

The window parameter should specify any child window or sub-child window of the frame or window managed by wxAuiManager.

The window parameter need not be managed by the manager itself, nor does it even need to be a child or sub-child of a managed window. It must however be inside the window hierarchy underneath the managed window.

getPane/2

Specs

getPane(This, Name) -> wxAuiPaneInfo:wxAuiPaneInfo()
           when This :: wxAuiManager(), Name :: unicode:chardata();
       (This, Window) -> wxAuiPaneInfo:wxAuiPaneInfo()
           when This :: wxAuiManager(), Window :: wxWindow:wxWindow().

getPane/2 is used to lookup a wxAuiPaneInfo object either by window pointer or by pane name, which acts as a unique id for a window pane.

The returned wxAuiPaneInfo object may then be modified to change a pane's look, state or position. After one or more modifications to wxAuiPaneInfo, update/1 should be called to commit the changes to the user interface. If the lookup failed (meaning the pane could not be found in the manager), a call to the returned wxAuiPaneInfo's IsOk() method will return false.

hideHint/1

Specs

hideHint(This) -> ok when This :: wxAuiManager().

hideHint/1 hides any docking hint that may be visible.

insertPane/3

Specs

insertPane(This, Window, Insert_location) -> boolean()
              when
                  This :: wxAuiManager(),
                  Window :: wxWindow:wxWindow(),
                  Insert_location :: wxAuiPaneInfo:wxAuiPaneInfo().

insertPane/4

Specs

insertPane(This, Window, Insert_location, [Option]) -> boolean()
              when
                  This :: wxAuiManager(),
                  Window :: wxWindow:wxWindow(),
                  Insert_location :: wxAuiPaneInfo:wxAuiPaneInfo(),
                  Option :: {insert_level, integer()}.

This method is used to insert either a previously unmanaged pane window into the frame manager, or to insert a currently managed pane somewhere else.

insertPane/4 will push all panes, rows, or docks aside and insert the window into the position specified by insert_location.

Because insert_location can specify either a pane, dock row, or dock layer, the insert_level parameter is used to disambiguate this. The parameter insert_level can take a value of wxAUI_INSERT_PANE, wxAUI_INSERT_ROW or wxAUI_INSERT_DOCK.

loadPaneInfo/3

Specs

loadPaneInfo(This, Pane_part, Pane) -> ok
                when
                    This :: wxAuiManager(),
                    Pane_part :: unicode:chardata(),
                    Pane :: wxAuiPaneInfo:wxAuiPaneInfo().

loadPaneInfo/3 is similar to LoadPerspective, with the exception that it only loads information about a single pane.

This method writes the serialized data into the passed pane. Pointers to UI elements are not modified.

Note: This operation also changes the name in the pane information!

See: loadPerspective/3

See: savePaneInfo/2

See: savePerspective/1

loadPerspective/2

Specs

loadPerspective(This, Perspective) -> boolean()
                   when This :: wxAuiManager(), Perspective :: unicode:chardata().

loadPerspective/3

Specs

loadPerspective(This, Perspective, [Option]) -> boolean()
                   when
                       This :: wxAuiManager(),
                       Perspective :: unicode:chardata(),
                       Option :: {update, boolean()}.

Loads a saved perspective.

A perspective is the layout state of an AUI managed window.

All currently existing panes that have an object in "perspective" with the same name ("equivalent") will receive the layout parameters of the object in "perspective". Existing panes that do not have an equivalent in "perspective" remain unchanged, objects in "perspective" having no equivalent in the manager are ignored.

See: loadPaneInfo/3

See: loadPerspective/3

See: savePerspective/1

new/0

Specs

new() -> wxAuiManager().

new/1

Specs

new([Option]) -> wxAuiManager()
       when Option :: {managed_wnd, wxWindow:wxWindow()} | {flags, integer()}.

Constructor.

savePaneInfo/2

Specs

savePaneInfo(This, Pane) -> unicode:charlist()
                when This :: wxAuiManager(), Pane :: wxAuiPaneInfo:wxAuiPaneInfo().

savePaneInfo/2 is similar to SavePerspective, with the exception that it only saves information about a single pane.

Return: The serialized layout parameters of the pane are returned within the string. Information about the pointers to UI elements stored in the pane are not serialized.

See: loadPaneInfo/3

See: loadPerspective/3

See: savePerspective/1

savePerspective/1

Specs

savePerspective(This) -> unicode:charlist() when This :: wxAuiManager().

Saves the entire user interface layout into an encoded wxString (not implemented in wx), which can then be stored by the application (probably using wxConfig).

See: loadPerspective/3

See: loadPaneInfo/3

See: savePaneInfo/2

setArtProvider/2

Specs

setArtProvider(This, Art_provider) -> ok
                  when This :: wxAuiManager(), Art_provider :: wxAuiDockArt:wxAuiDockArt().

Instructs wxAuiManager to use art provider specified by parameter art_provider for all drawing calls.

This allows pluggable look-and-feel features. The previous art provider object, if any, will be deleted by wxAuiManager.

See: wxAuiDockArt

setDockSizeConstraint/3

Specs

setDockSizeConstraint(This, Widthpct, Heightpct) -> ok
                         when This :: wxAuiManager(), Widthpct :: number(), Heightpct :: number().

When a user creates a new dock by dragging a window into a docked position, often times the large size of the window will create a dock that is unwieldy large.

wxAuiManager by default limits the size of any new dock to 1/3 of the window size. For horizontal docks, this would be 1/3 of the window height. For vertical docks, 1/3 of the width.

Calling this function will adjust this constraint value. The numbers must be between 0.0 and 1.0. For instance, calling SetDockSizeContraint with 0.5, 0.5 will cause new docks to be limited to half of the size of the entire managed window.

setFlags/2

Specs

setFlags(This, Flags) -> ok when This :: wxAuiManager(), Flags :: integer().

This method is used to specify ?wxAuiManagerOption's flags.

flags specifies options which allow the frame management behaviour to be modified.

setManagedWindow/2

Specs

setManagedWindow(This, Managed_wnd) -> ok
                    when This :: wxAuiManager(), Managed_wnd :: wxWindow:wxWindow().

Called to specify the frame or window which is to be managed by wxAuiManager.

Frame management is not restricted to just frames. Child windows or custom controls are also allowed.

showHint/2

Specs

showHint(This, Rect) -> ok
            when
                This :: wxAuiManager(),
                Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()}.

This function is used by controls to explicitly show a hint window at the specified rectangle.

It is rarely called, and is mostly used by controls implementing custom pane drag/drop behaviour. The specified rectangle should be in screen coordinates.

unInit/1

Specs

unInit(This) -> ok when This :: wxAuiManager().

Dissociate the managed window from the manager.

This function may be called before the managed frame or window is destroyed, but, since wxWidgets 3.1.4, it's unnecessary to call it explicitly, as it will be called automatically when this window is destroyed, as well as when the manager itself is.

update/1

Specs

update(This) -> ok when This :: wxAuiManager().

This method is called after any number of changes are made to any of the managed panes.

update/1 must be invoked after addPane/4 or insertPane/4 are called in order to "realize" or "commit" the changes. In addition, any number of changes may be made to wxAuiPaneInfo structures (retrieved with getPane/2), but to realize the changes, update/1 must be called. This construction allows pane flicker to be avoided by updating the whole layout at one time.