wxAuiManagerEvent (wx v2.0.1)

Event used to indicate various actions taken with wxAuiManager.

See wxAuiManager for available event types.

See: wxAuiManager, wxAuiPaneInfo

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxAuiManagerEvent

Events

Use wxEvtHandler:connect/3 with wxAuiManagerEventType to subscribe to events of this type.

Types

-type wxAuiManager() :: term().

Specs

wxAuiManager() :: #wxAuiManager{}.

-type wxAuiManagerEvent() :: term().

Specs

wxAuiManagerEvent() :: wx:wx_object().

-type wxAuiManagerEventType() :: term().

Specs

wxAuiManagerEventType() ::
    aui_pane_button | aui_pane_close | aui_pane_maximize | aui_pane_restore | aui_pane_activated |
    aui_render | aui_find_manager.

Functions

canVeto/1

Specs

canVeto(This) -> boolean() when This :: wxAuiManagerEvent().

Return: true if this event can be vetoed.

See: veto/2

getButton/1

Specs

getButton(This) -> integer() when This :: wxAuiManagerEvent().

Return: The ID of the button that was clicked.

getDC/1

Specs

getDC(This) -> wxDC:wxDC() when This :: wxAuiManagerEvent().

getManager/1

Specs

getManager(This) -> wxAuiManager:wxAuiManager() when This :: wxAuiManagerEvent().

Return: The wxAuiManager this event is associated with.

getPane/1

Specs

getPane(This) -> wxAuiPaneInfo:wxAuiPaneInfo() when This :: wxAuiManagerEvent().

Return: The pane this event is associated with.

getVeto/1

Specs

getVeto(This) -> boolean() when This :: wxAuiManagerEvent().

Return: true if this event was vetoed.

See: veto/2

setButton/2

Specs

setButton(This, Button) -> ok when This :: wxAuiManagerEvent(), Button :: integer().

Sets the ID of the button clicked that triggered this event.

setCanVeto/2

Specs

setCanVeto(This, Can_veto) -> ok when This :: wxAuiManagerEvent(), Can_veto :: boolean().

Sets whether or not this event can be vetoed.

setDC/2

Specs

setDC(This, Pdc) -> ok when This :: wxAuiManagerEvent(), Pdc :: wxDC:wxDC().

setManager/2

Specs

setManager(This, Manager) -> ok
              when This :: wxAuiManagerEvent(), Manager :: wxAuiManager:wxAuiManager().

Sets the wxAuiManager this event is associated with.

setPane/2

Specs

setPane(This, Pane) -> ok when This :: wxAuiManagerEvent(), Pane :: wxAuiPaneInfo:wxAuiPaneInfo().

Sets the pane this event is associated with.

veto/1

Specs

veto(This) -> ok when This :: wxAuiManagerEvent().

veto/2

Specs

veto(This, [Option]) -> ok when This :: wxAuiManagerEvent(), Option :: {veto, boolean()}.

Cancels the action indicated by this event if canVeto/1 is true.