wxAuiNotebookEvent (wx v2.0.1)

This class is used by the events generated by wxAuiNotebook.

See: wxAuiNotebook, wxBookCtrlEvent

This class is derived (and can use functions) from: wxBookCtrlEvent wxNotifyEvent wxCommandEvent wxEvent

wxWidgets docs: wxAuiNotebookEvent

Events

Use wxEvtHandler:connect/3 with wxAuiNotebookEventType to subscribe to events of this type.

Types

-type wxAuiNotebook() :: term().

Specs

wxAuiNotebook() :: #wxAuiNotebook{}.

-type wxAuiNotebookEvent() :: term().

Specs

wxAuiNotebookEvent() :: wx:wx_object().

-type wxAuiNotebookEventType() :: term().

Specs

wxAuiNotebookEventType() ::
    command_auinotebook_page_close | command_auinotebook_page_changed |
    command_auinotebook_page_changing | command_auinotebook_button |
    command_auinotebook_begin_drag | command_auinotebook_end_drag |
    command_auinotebook_drag_motion | command_auinotebook_allow_dnd |
    command_auinotebook_tab_middle_down | command_auinotebook_tab_middle_up |
    command_auinotebook_tab_right_down | command_auinotebook_tab_right_up |
    command_auinotebook_page_closed | command_auinotebook_drag_done |
    command_auinotebook_bg_dclick.

Functions

getDragSource/1

Specs

getDragSource(This) -> wxAuiNotebook:wxAuiNotebook() when This :: wxAuiNotebookEvent().

getOldSelection/1

Specs

getOldSelection(This) -> integer() when This :: wxAuiNotebookEvent().

Returns the page that was selected before the change, wxNOT_FOUND if none was selected.

getSelection/1

Specs

getSelection(This) -> integer() when This :: wxAuiNotebookEvent().

Returns the currently selected page, or wxNOT_FOUND if none was selected.

Note: under Windows, getSelection/1 will return the same value as getOldSelection/1 when called from the EVT_BOOKCTRL_PAGE_CHANGING handler and not the page which is going to be selected.

setDragSource/2

Specs

setDragSource(This, S) -> ok when This :: wxAuiNotebookEvent(), S :: wxAuiNotebook:wxAuiNotebook().

setOldSelection/2

Specs

setOldSelection(This, Page) -> ok when This :: wxAuiNotebookEvent(), Page :: integer().

Sets the id of the page selected before the change.

setSelection/2

Specs

setSelection(This, Page) -> ok when This :: wxAuiNotebookEvent(), Page :: integer().

Sets the selection member variable.