wxAuiPaneInfo (wx v2.0.1)

wxAuiPaneInfo is part of the wxAUI class framework. See also overview_aui.

wxAuiPaneInfo specifies all the parameters for a pane. These parameters specify where the pane is on the screen, whether it is docked or floating, or hidden. In addition, these parameters specify the pane's docked position, floating position, preferred size, minimum size, caption text among many other parameters.

See: wxAuiManager, wxAuiDockArt

wxWidgets docs: wxAuiPaneInfo

Types

-type wxAuiPaneInfo() :: term().

Specs

wxAuiPaneInfo() :: wx:wx_object().

Functions

bestSize/2

Specs

bestSize(This, Size) -> wxAuiPaneInfo()
            when This :: wxAuiPaneInfo(), Size :: {W :: integer(), H :: integer()}.

bestSize/3 sets the ideal size for the pane.

The docking manager will attempt to use this size as much as possible when docking or floating the pane.

bestSize/3

Specs

bestSize(This, X, Y) -> wxAuiPaneInfo()
            when This :: wxAuiPaneInfo(), X :: integer(), Y :: integer().

bottom/1

Specs

bottom(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

bottom/1 sets the pane dock position to the bottom side of the frame.

This is the same thing as calling Direction(wxAUI_DOCK_BOTTOM).

bottomDockable/1

Specs

bottomDockable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

bottomDockable/2

Specs

bottomDockable(This, [Option]) -> wxAuiPaneInfo()
                  when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

bottomDockable/2 indicates whether a pane can be docked at the bottom of the frame.

caption/2

Specs

caption(This, C) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), C :: unicode:chardata().

caption/2 sets the caption of the pane.

captionVisible/1

Specs

captionVisible(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

captionVisible/2

Specs

captionVisible(This, [Option]) -> wxAuiPaneInfo()
                  when This :: wxAuiPaneInfo(), Option :: {visible, boolean()}.

CaptionVisible indicates that a pane caption should be visible.

If false, no pane caption is drawn.

centre/1

Specs

centre(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

Center() (not implemented in wx) sets the pane dock position to the left side of the frame.

The centre pane is the space in the middle after all border panes (left, top, right, bottom) are subtracted from the layout. This is the same thing as calling Direction(wxAUI_DOCK_CENTRE).

centrePane/1

Specs

centrePane(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

centrePane/1 specifies that the pane should adopt the default center pane settings.

Centre panes usually do not have caption bars. This function provides an easy way of preparing a pane to be displayed in the center dock position.

closeButton/1

Specs

closeButton(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

closeButton/2

Specs

closeButton(This, [Option]) -> wxAuiPaneInfo()
               when This :: wxAuiPaneInfo(), Option :: {visible, boolean()}.

closeButton/2 indicates that a close button should be drawn for the pane.

defaultPane/1

Specs

defaultPane(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

defaultPane/1 specifies that the pane should adopt the default pane settings.

destroy/1

Specs

destroy(This :: wxAuiPaneInfo()) -> ok.

Destroys the object.

destroyOnClose/1

Specs

destroyOnClose(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

destroyOnClose/2

Specs

destroyOnClose(This, [Option]) -> wxAuiPaneInfo()
                  when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

destroyOnClose/2 indicates whether a pane should be destroyed when it is closed.

Normally a pane is simply hidden when the close button is clicked. Setting DestroyOnClose to true will cause the window to be destroyed when the user clicks the pane's close button.

direction/2

Specs

direction(This, Direction) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), Direction :: integer().

direction/2 determines the direction of the docked pane.

It is functionally the same as calling left/1, right/1, top/1 or bottom/1, except that docking direction may be specified programmatically via the parameter.

dock/1

Specs

dock(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

dock/1 indicates that a pane should be docked.

It is the opposite of float/1.

dockable/1

Specs

dockable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

dockable/2

Specs

dockable(This, [Option]) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

dockable/2 specifies whether a frame can be docked or not.

It is the same as specifying TopDockable(b).BottomDockable(b).LeftDockable(b).RightDockable(b).

fixed/1

Specs

fixed(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

fixed/1 forces a pane to be fixed size so that it cannot be resized.

After calling fixed/1, isFixed/1 will return true.

float/1

Specs

float(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

float/1 indicates that a pane should be floated.

It is the opposite of dock/1.

floatable/1

Specs

floatable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

floatable/2

Specs

floatable(This, [Option]) -> wxAuiPaneInfo()
             when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

floatable/2 sets whether the user will be able to undock a pane and turn it into a floating window.

floatingPosition/2

Specs

floatingPosition(This, Pos) -> wxAuiPaneInfo()
                    when This :: wxAuiPaneInfo(), Pos :: {X :: integer(), Y :: integer()}.

floatingPosition/3 sets the position of the floating pane.

floatingPosition/3

Specs

floatingPosition(This, X, Y) -> wxAuiPaneInfo()
                    when This :: wxAuiPaneInfo(), X :: integer(), Y :: integer().

floatingSize/2

Specs

floatingSize(This, Size) -> wxAuiPaneInfo()
                when This :: wxAuiPaneInfo(), Size :: {W :: integer(), H :: integer()}.

floatingSize/3 sets the size of the floating pane.

floatingSize/3

Specs

floatingSize(This, X, Y) -> wxAuiPaneInfo()
                when This :: wxAuiPaneInfo(), X :: integer(), Y :: integer().

getDirection/1

Specs

getDirection(This) -> integer() when This :: wxAuiPaneInfo().

getFloatingPosition/1

Specs

getFloatingPosition(This) -> {X :: integer(), Y :: integer()} when This :: wxAuiPaneInfo().

getFloatingSize/1

Specs

getFloatingSize(This) -> {W :: integer(), H :: integer()} when This :: wxAuiPaneInfo().

getFrame/1

Specs

getFrame(This) -> wxFrame:wxFrame() when This :: wxAuiPaneInfo().

getLayer/1

Specs

getLayer(This) -> integer() when This :: wxAuiPaneInfo().

getPosition/1

Specs

getPosition(This) -> integer() when This :: wxAuiPaneInfo().

getRow/1

Specs

getRow(This) -> integer() when This :: wxAuiPaneInfo().

getWindow/1

Specs

getWindow(This) -> wxWindow:wxWindow() when This :: wxAuiPaneInfo().

gripper/1

Specs

gripper(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

gripper/2

Specs

gripper(This, [Option]) -> wxAuiPaneInfo()
           when This :: wxAuiPaneInfo(), Option :: {visible, boolean()}.

gripper/2 indicates that a gripper should be drawn for the pane.

gripperTop/1

Specs

gripperTop(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

gripperTop/2

Specs

gripperTop(This, [Option]) -> wxAuiPaneInfo()
              when This :: wxAuiPaneInfo(), Option :: {attop, boolean()}.

gripperTop/2 indicates that a gripper should be drawn at the top of the pane.

hasBorder/1

Specs

hasBorder(This) -> boolean() when This :: wxAuiPaneInfo().

hasBorder/1 returns true if the pane displays a border.

hasCaption/1

Specs

hasCaption(This) -> boolean() when This :: wxAuiPaneInfo().

hasCaption/1 returns true if the pane displays a caption.

hasCloseButton/1

Specs

hasCloseButton(This) -> boolean() when This :: wxAuiPaneInfo().

hasCloseButton/1 returns true if the pane displays a button to close the pane.

hasFlag/2

Specs

hasFlag(This, Flag) -> boolean() when This :: wxAuiPaneInfo(), Flag :: integer().

hasFlag/2 returns true if the property specified by flag is active for the pane.

hasGripper/1

Specs

hasGripper(This) -> boolean() when This :: wxAuiPaneInfo().

hasGripper/1 returns true if the pane displays a gripper.

hasGripperTop/1

Specs

hasGripperTop(This) -> boolean() when This :: wxAuiPaneInfo().

hasGripper/1 returns true if the pane displays a gripper at the top.

hasMaximizeButton/1

Specs

hasMaximizeButton(This) -> boolean() when This :: wxAuiPaneInfo().

hasMaximizeButton/1 returns true if the pane displays a button to maximize the pane.

hasMinimizeButton/1

Specs

hasMinimizeButton(This) -> boolean() when This :: wxAuiPaneInfo().

hasMinimizeButton/1 returns true if the pane displays a button to minimize the pane.

hasPinButton/1

Specs

hasPinButton(This) -> boolean() when This :: wxAuiPaneInfo().

hasPinButton/1 returns true if the pane displays a button to float the pane.

hide/1

Specs

hide(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

hide/1 indicates that a pane should be hidden.

isBottomDockable/1

Specs

isBottomDockable(This) -> boolean() when This :: wxAuiPaneInfo().

isBottomDockable/1 returns true if the pane can be docked at the bottom of the managed frame.

See: IsDockable() (not implemented in wx)

isDocked/1

Specs

isDocked(This) -> boolean() when This :: wxAuiPaneInfo().

isDocked/1 returns true if the pane is currently docked.

isFixed/1

Specs

isFixed(This) -> boolean() when This :: wxAuiPaneInfo().

isFixed/1 returns true if the pane cannot be resized.

isFloatable/1

Specs

isFloatable(This) -> boolean() when This :: wxAuiPaneInfo().

isFloatable/1 returns true if the pane can be undocked and displayed as a floating window.

isFloating/1

Specs

isFloating(This) -> boolean() when This :: wxAuiPaneInfo().

isFloating/1 returns true if the pane is floating.

isLeftDockable/1

Specs

isLeftDockable(This) -> boolean() when This :: wxAuiPaneInfo().

isLeftDockable/1 returns true if the pane can be docked on the left of the managed frame.

See: IsDockable() (not implemented in wx)

isMovable/1

Specs

isMovable(This) -> boolean() when This :: wxAuiPaneInfo().

IsMoveable() returns true if the docked frame can be undocked or moved to another dock position.

isOk/1

Specs

isOk(This) -> boolean() when This :: wxAuiPaneInfo().

isOk/1 returns true if the wxAuiPaneInfo structure is valid.

A pane structure is valid if it has an associated window.

isResizable/1

Specs

isResizable(This) -> boolean() when This :: wxAuiPaneInfo().

isResizable/1 returns true if the pane can be resized.

isRightDockable/1

Specs

isRightDockable(This) -> boolean() when This :: wxAuiPaneInfo().

isRightDockable/1 returns true if the pane can be docked on the right of the managed frame.

See: IsDockable() (not implemented in wx)

isShown/1

Specs

isShown(This) -> boolean() when This :: wxAuiPaneInfo().

isShown/1 returns true if the pane is currently shown.

isToolbar/1

Specs

isToolbar(This) -> boolean() when This :: wxAuiPaneInfo().

isToolbar/1 returns true if the pane contains a toolbar.

isTopDockable/1

Specs

isTopDockable(This) -> boolean() when This :: wxAuiPaneInfo().

isTopDockable/1 returns true if the pane can be docked at the top of the managed frame.

See: IsDockable() (not implemented in wx)

layer/2

Specs

layer(This, Layer) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), Layer :: integer().

layer/2 determines the layer of the docked pane.

The dock layer is similar to an onion, the inner-most layer being layer 0. Each shell moving in the outward direction has a higher layer number. This allows for more complex docking layout formation.

left/1

Specs

left(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

left/1 sets the pane dock position to the left side of the frame.

This is the same thing as calling Direction(wxAUI_DOCK_LEFT).

leftDockable/1

Specs

leftDockable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

leftDockable/2

Specs

leftDockable(This, [Option]) -> wxAuiPaneInfo()
                when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

leftDockable/2 indicates whether a pane can be docked on the left of the frame.

maxSize/2

Specs

maxSize(This, Size) -> wxAuiPaneInfo()
           when This :: wxAuiPaneInfo(), Size :: {W :: integer(), H :: integer()}.

maxSize/3 sets the maximum size of the pane.

maxSize/3

Specs

maxSize(This, X, Y) -> wxAuiPaneInfo()
           when This :: wxAuiPaneInfo(), X :: integer(), Y :: integer().

maximizeButton/1

Specs

maximizeButton(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

maximizeButton/2

Specs

maximizeButton(This, [Option]) -> wxAuiPaneInfo()
                  when This :: wxAuiPaneInfo(), Option :: {visible, boolean()}.

maximizeButton/2 indicates that a maximize button should be drawn for the pane.

minSize/2

Specs

minSize(This, Size) -> wxAuiPaneInfo()
           when This :: wxAuiPaneInfo(), Size :: {W :: integer(), H :: integer()}.

minSize/3 sets the minimum size of the pane.

Please note that this is only partially supported as of this writing.

minSize/3

Specs

minSize(This, X, Y) -> wxAuiPaneInfo()
           when This :: wxAuiPaneInfo(), X :: integer(), Y :: integer().

minimizeButton/1

Specs

minimizeButton(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

minimizeButton/2

Specs

minimizeButton(This, [Option]) -> wxAuiPaneInfo()
                  when This :: wxAuiPaneInfo(), Option :: {visible, boolean()}.

minimizeButton/2 indicates that a minimize button should be drawn for the pane.

movable/1

Specs

movable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

movable/2

Specs

movable(This, [Option]) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

Movable indicates whether a frame can be moved.

name/2

Specs

name(This, N) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), N :: unicode:chardata().

name/2 sets the name of the pane so it can be referenced in lookup functions.

If a name is not specified by the user, a random name is assigned to the pane when it is added to the manager.

new/0

Specs

new() -> wxAuiPaneInfo().

new/1

Specs

new(C) -> wxAuiPaneInfo() when C :: wxAuiPaneInfo().

Copy constructor.

paneBorder/1

Specs

paneBorder(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

paneBorder/2

Specs

paneBorder(This, [Option]) -> wxAuiPaneInfo()
              when This :: wxAuiPaneInfo(), Option :: {visible, boolean()}.

PaneBorder indicates that a border should be drawn for the pane.

pinButton/1

Specs

pinButton(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

pinButton/2

Specs

pinButton(This, [Option]) -> wxAuiPaneInfo()
             when This :: wxAuiPaneInfo(), Option :: {visible, boolean()}.

pinButton/2 indicates that a pin button should be drawn for the pane.

position/2

Specs

position(This, Pos) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), Pos :: integer().

position/2 determines the position of the docked pane.

resizable/1

Specs

resizable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

resizable/2

Specs

resizable(This, [Option]) -> wxAuiPaneInfo()
             when This :: wxAuiPaneInfo(), Option :: {resizable, boolean()}.

resizable/2 allows a pane to be resized if the parameter is true, and forces it to be a fixed size if the parameter is false.

This is simply an antonym for fixed/1.

right/1

Specs

right(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

right/1 sets the pane dock position to the right side of the frame.

This is the same thing as calling Direction(wxAUI_DOCK_RIGHT).

rightDockable/1

Specs

rightDockable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

rightDockable/2

Specs

rightDockable(This, [Option]) -> wxAuiPaneInfo()
                 when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

rightDockable/2 indicates whether a pane can be docked on the right of the frame.

row/2

Specs

row(This, Row) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), Row :: integer().

row/2 determines the row of the docked pane.

safeSet/2

Specs

safeSet(This, Source) -> ok when This :: wxAuiPaneInfo(), Source :: wxAuiPaneInfo().

Write the safe parts of a PaneInfo object "source" into "this".

"Safe parts" are all non-UI elements (e.g. all layout determining parameters like the size, position etc.). "Unsafe parts" (pointers to button, frame and window) are not modified by this write operation.

Remark: This method is used when loading perspectives.

setFlag/3

Specs

setFlag(This, Flag, Option_state) -> wxAuiPaneInfo()
           when This :: wxAuiPaneInfo(), Flag :: integer(), Option_state :: boolean().

setFlag/3 turns the property given by flag on or off with the option_state parameter.

show/1

Specs

show(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

show/2

Specs

show(This, [Option]) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), Option :: {show, boolean()}.

show/2 indicates that a pane should be shown.

toolbarPane/1

Specs

toolbarPane(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

toolbarPane/1 specifies that the pane should adopt the default toolbar pane settings.

top/1

Specs

top(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

top/1 sets the pane dock position to the top of the frame.

This is the same thing as calling Direction(wxAUI_DOCK_TOP).

topDockable/1

Specs

topDockable(This) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo().

topDockable/2

Specs

topDockable(This, [Option]) -> wxAuiPaneInfo()
               when This :: wxAuiPaneInfo(), Option :: {b, boolean()}.

topDockable/2 indicates whether a pane can be docked at the top of the frame.

window/2

Specs

window(This, W) -> wxAuiPaneInfo() when This :: wxAuiPaneInfo(), W :: wxWindow:wxWindow().

window/2 assigns the window pointer that the wxAuiPaneInfo should use.

This normally does not need to be specified, as the window pointer is automatically assigned to the wxAuiPaneInfo structure as soon as it is added to the manager.