wxBitmapDataObject (wx v2.4.3.1)

wxBitmapDataObject is a specialization of wxDataObject for bitmap data.

It can be used without change to paste data into the wxClipboard or a wxDropSource (not implemented in wx). A user may wish to derive a new class from this class for providing a bitmap on-demand in order to minimize memory consumption when offering data in several formats, such as a bitmap and GIF.

This class may be used as is, but getBitmap/1 may be overridden to increase efficiency.

See:

Overview dnd

wxDataObject

wxFileDataObject

wxTextDataObject

wxDataObject

This class is derived, and can use functions, from:

wxDataObject

wxWidgets docs: wxBitmapDataObject

Types

wxBitmapDataObject()

-type wxBitmapDataObject() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxBitmapDataObject()) -> ok.

Destroys the object

getBitmap(This)

-spec getBitmap(This) -> wxBitmap:wxBitmap() when This :: wxBitmapDataObject().

Returns the bitmap associated with the data object.

You may wish to override this method when offering data on-demand, but this is not required by wxWidgets' internals. Use this method to get data in bitmap form from the wxClipboard.

new()

-spec new() -> wxBitmapDataObject().

Equivalent to new([]).

new/1

-spec new([Option]) -> wxBitmapDataObject() when Option :: {bitmap, wxBitmap:wxBitmap()};
         (Bitmap) -> wxBitmapDataObject() when Bitmap :: wxBitmap:wxBitmap().

Constructor, optionally passing a bitmap (otherwise use setBitmap/2 later).

setBitmap(This, Bitmap)

-spec setBitmap(This, Bitmap) -> ok when This :: wxBitmapDataObject(), Bitmap :: wxBitmap:wxBitmap().

Sets the bitmap associated with the data object.

This method is called when the data object receives data. Usually there will be no reason to override this function.