wxBoxSizer (wx v2.4.2)

Functions for wxBoxSizer class

The basic idea behind a box sizer is that windows will most often be laid out in rather simple basic geometry, typically in a row or a column or several hierarchies of either.

For more information, please see overview_sizer_box.

See: wxSizer, Overview sizer

This class is derived (and can use functions) from: wxSizer

wxWidgets docs: wxBoxSizer

Types

wxBoxSizer()

-type wxBoxSizer() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxBoxSizer()) -> ok.

Destroys the object.

getOrientation(This)

-spec getOrientation(This) -> integer() when This :: wxBoxSizer().

Returns the orientation of the box sizer, either wxVERTICAL or wxHORIZONTAL.

new(Orient)

-spec new(Orient) -> wxBoxSizer() when Orient :: integer().

Constructor for a wxBoxSizer.

orient may be either of wxVERTICAL or wxHORIZONTAL for creating either a column sizer or a row sizer.