wxCalendarDateAttr (wx v2.0.1)

wxCalendarDateAttr is a custom attributes for a calendar date. The objects of this class are used with wxCalendarCtrl.

See: wxCalendarCtrl

wxWidgets docs: wxCalendarDateAttr

Types

-type wxCalendarDateAttr() :: term().

Specs

wxCalendarDateAttr() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxCalendarDateAttr()) -> ok.

Destroys the object.

getBackgroundColour/1

Specs

getBackgroundColour(This) -> wx:wx_colour4() when This :: wxCalendarDateAttr().

Returns the background colour set for the calendar date.

getBorder/1

Specs

getBorder(This) -> wx:wx_enum() when This :: wxCalendarDateAttr().

Returns the border set for the calendar date.

getBorderColour/1

Specs

getBorderColour(This) -> wx:wx_colour4() when This :: wxCalendarDateAttr().

Returns the border colour set for the calendar date.

getFont/1

Specs

getFont(This) -> wxFont:wxFont() when This :: wxCalendarDateAttr().

Returns the font set for the calendar date.

getTextColour/1

Specs

getTextColour(This) -> wx:wx_colour4() when This :: wxCalendarDateAttr().

Returns the text colour set for the calendar date.

hasBackgroundColour/1

Specs

hasBackgroundColour(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default text background colour is set.

hasBorder/1

Specs

hasBorder(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default (i.e. any) border is set.

hasBorderColour/1

Specs

hasBorderColour(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default border colour is set.

hasFont/1

Specs

hasFont(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default font is set.

hasTextColour/1

Specs

hasTextColour(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if a non-default text foreground colour is set.

isHoliday/1

Specs

isHoliday(This) -> boolean() when This :: wxCalendarDateAttr().

Returns true if this calendar day is displayed as a holiday.

new/0

Specs

new() -> wxCalendarDateAttr().

new/1

Specs

new(Border) -> wxCalendarDateAttr() when Border :: wx:wx_enum();
   ([Option]) -> wxCalendarDateAttr()
       when
           Option ::
               {colText, wx:wx_colour()} |
               {colBack, wx:wx_colour()} |
               {colBorder, wx:wx_colour()} |
               {font, wxFont:wxFont()} |
               {border, wx:wx_enum()}.

Constructor for specifying all wxCalendarDateAttr properties.

new/2

Specs

new(Border, [Option]) -> wxCalendarDateAttr()
       when Border :: wx:wx_enum(), Option :: {colBorder, wx:wx_colour()}.

Constructor using default properties except the given border.

setBackgroundColour/2

Specs

setBackgroundColour(This, ColBack) -> ok
                       when This :: wxCalendarDateAttr(), ColBack :: wx:wx_colour().

Sets the text background colour to use.

setBorder/2

Specs

setBorder(This, Border) -> ok when This :: wxCalendarDateAttr(), Border :: wx:wx_enum().

Sets the border to use.

setBorderColour/2

Specs

setBorderColour(This, Col) -> ok when This :: wxCalendarDateAttr(), Col :: wx:wx_colour().

Sets the border colour to use.

setFont/2

Specs

setFont(This, Font) -> ok when This :: wxCalendarDateAttr(), Font :: wxFont:wxFont().

Sets the font to use.

setHoliday/2

Specs

setHoliday(This, Holiday) -> ok when This :: wxCalendarDateAttr(), Holiday :: boolean().

If holiday is true, this calendar day will be displayed as a holiday.

setTextColour/2

Specs

setTextColour(This, ColText) -> ok when This :: wxCalendarDateAttr(), ColText :: wx:wx_colour().

Sets the text (foreground) colour to use.