wxCheckBox (wx v2.0.1)

A checkbox is a labelled box which by default is either on (checkmark is visible) or off (no checkmark). Optionally (when the wxCHK_3STATE style flag is set) it can have a third state, called the mixed or undetermined state. Often this is used as a "Does Not Apply" state.

Styles

This class supports the following styles:

See: wxRadioButton, wxCommandEvent

This class is derived (and can use functions) from: wxControl wxWindow wxEvtHandler

wxWidgets docs: wxCheckBox

Events

Event types emitted from this class: command_checkbox_clicked

Types

-type wxCheckBox() :: term().

Specs

wxCheckBox() :: wx:wx_object().

Functions

create/4

Specs

create(This, Parent, Id, Label) -> boolean()
          when
              This :: wxCheckBox(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: unicode:chardata().

create/5

Specs

create(This, Parent, Id, Label, [Option]) -> boolean()
          when
              This :: wxCheckBox(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: unicode:chardata(),
              Option ::
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()} |
                  {validator, wx:wx_object()}.

Creates the checkbox for two-step construction.

See new/4 for details.

destroy/1

Specs

destroy(This :: wxCheckBox()) -> ok.

Destructor, destroying the checkbox.

get3StateValue/1

Specs

get3StateValue(This) -> wx:wx_enum() when This :: wxCheckBox().

Gets the state of a 3-state checkbox.

Asserts when the function is used with a 2-state checkbox.

getValue/1

Specs

getValue(This) -> boolean() when This :: wxCheckBox().

Gets the state of a 2-state checkbox.

Return: Returns true if it is checked, false otherwise.

is3State/1

Specs

is3State(This) -> boolean() when This :: wxCheckBox().

Returns whether or not the checkbox is a 3-state checkbox.

Return: true if this checkbox is a 3-state checkbox, false if it's a 2-state checkbox.

is3rdStateAllowedForUser/1

Specs

is3rdStateAllowedForUser(This) -> boolean() when This :: wxCheckBox().

Returns whether or not the user can set the checkbox to the third state.

Return: true if the user can set the third state of this checkbox, false if it can only be set programmatically or if it's a 2-state checkbox.

isChecked/1

Specs

isChecked(This) -> boolean() when This :: wxCheckBox().

This is just a maybe more readable synonym for getValue/1: just as the latter, it returns true if the checkbox is checked and false otherwise.

new/0

Specs

new() -> wxCheckBox().

Default constructor.

See: create/5, wxValidator (not implemented in wx)

new/3

Specs

new(Parent, Id, Label) -> wxCheckBox()
       when Parent :: wxWindow:wxWindow(), Id :: integer(), Label :: unicode:chardata().

new/4

Specs

new(Parent, Id, Label, [Option]) -> wxCheckBox()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Label :: unicode:chardata(),
           Option ::
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()} |
               {validator, wx:wx_object()}.

Constructor, creating and showing a checkbox.

See: create/5, wxValidator (not implemented in wx)

set3StateValue/2

Specs

set3StateValue(This, State) -> ok when This :: wxCheckBox(), State :: wx:wx_enum().

Sets the checkbox to the given state.

This does not cause a wxEVT_CHECKBOX event to get emitted.

Asserts when the checkbox is a 2-state checkbox and setting the state to wxCHK_UNDETERMINED.

setValue/2

Specs

setValue(This, State) -> ok when This :: wxCheckBox(), State :: boolean().

Sets the checkbox to the given state.

This does not cause a wxEVT_CHECKBOX event to get emitted.