wxColourData (wx v2.0.1)

This class holds a variety of information related to colour dialogs.

See: wx_color(), wxColourDialog, Overview cmndlg

wxWidgets docs: wxColourData

Types

-type wxColourData() :: term().

Specs

wxColourData() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxColourData()) -> ok.

Destructor.

getChooseFull/1

Specs

getChooseFull(This) -> boolean() when This :: wxColourData().

Under Windows, determines whether the Windows colour dialog will display the full dialog with custom colour selection controls.

Has no meaning under other platforms.

The default value is true.

getColour/1

Specs

getColour(This) -> wx:wx_colour4() when This :: wxColourData().

Gets the current colour associated with the colour dialog.

The default colour is black.

getCustomColour/2

Specs

getCustomColour(This, I) -> wx:wx_colour4() when This :: wxColourData(), I :: integer().

Returns custom colours associated with the colour dialog.

new/0

Specs

new() -> wxColourData().

Constructor.

Initializes the custom colours to wxNullColour, the data colour setting to black, and the choose full setting to true.

setChooseFull/2

Specs

setChooseFull(This, Flag) -> ok when This :: wxColourData(), Flag :: boolean().

Under Windows, tells the Windows colour dialog to display the full dialog with custom colour selection controls.

Under other platforms, has no effect.

The default value is true.

setColour/2

Specs

setColour(This, Colour) -> ok when This :: wxColourData(), Colour :: wx:wx_colour().

Sets the default colour for the colour dialog.

The default colour is black.

setCustomColour/3

Specs

setCustomColour(This, I, Colour) -> ok
                   when This :: wxColourData(), I :: integer(), Colour :: wx:wx_colour().

Sets custom colours for the colour dialog.