wxColourPickerCtrl (wx v2.0.1)

This control allows the user to select a colour. The generic implementation is a button which brings up a wxColourDialog when clicked. Native implementation may differ but this is usually a (small) widget which give access to the colour-chooser dialog. It is only available if wxUSE_COLOURPICKERCTRL is set to 1 (the default).

Styles

This class supports the following styles:

See: wxColourDialog, wxColourPickerEvent

This class is derived (and can use functions) from: wxPickerBase wxControl wxWindow wxEvtHandler

wxWidgets docs: wxColourPickerCtrl

Events

Event types emitted from this class: command_colourpicker_changed

Types

-type wxColourPickerCtrl() :: term().

Specs

wxColourPickerCtrl() :: wx:wx_object().

Functions

create/3

Specs

create(This, Parent, Id) -> boolean()
          when This :: wxColourPickerCtrl(), Parent :: wxWindow:wxWindow(), Id :: integer().

create/4

Specs

create(This, Parent, Id, [Option]) -> boolean()
          when
              This :: wxColourPickerCtrl(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Option ::
                  {col, wx:wx_colour()} |
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()} |
                  {validator, wx:wx_object()}.

Creates a colour picker with the given arguments.

Return: true if the control was successfully created or false if creation failed.

destroy/1

Specs

destroy(This :: wxColourPickerCtrl()) -> ok.

Destroys the object.

getColour/1

Specs

getColour(This) -> wx:wx_colour4() when This :: wxColourPickerCtrl().

Returns the currently selected colour.

new/0

Specs

new() -> wxColourPickerCtrl().

new/2

Specs

new(Parent, Id) -> wxColourPickerCtrl() when Parent :: wxWindow:wxWindow(), Id :: integer().

new/3

Specs

new(Parent, Id, [Option]) -> wxColourPickerCtrl()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Option ::
               {col, wx:wx_colour()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()} |
               {validator, wx:wx_object()}.

Initializes the object and calls create/4 with all the parameters.

setColour/2

Specs

setColour(This, Colname) -> ok when This :: wxColourPickerCtrl(), Colname :: unicode:chardata();
         (This, Col) -> ok when This :: wxColourPickerCtrl(), Col :: wx:wx_colour().

Sets the currently selected colour.

See wxColour::Set() (not implemented in wx).