wxDirDialog (wx v2.0.1)

This class represents the directory chooser dialog.

Styles

This class supports the following styles:

Note: This flag cannot be used with the wxDD_MULTIPLE style.

Remark: MacOS 10.11+ does not display a title bar on the dialog. Use setMessage/2 to change the string displayed to the user at the top of the dialog after creation. The wxTopLevelWindow:setTitle/2 method is provided for compatibility with pre-10.11 MacOS versions that do still support displaying the title bar.

See: Overview cmndlg, wxFileDialog

This class is derived (and can use functions) from: wxDialog wxTopLevelWindow wxWindow wxEvtHandler

wxWidgets docs: wxDirDialog

Types

-type wxDirDialog() :: term().

Specs

wxDirDialog() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxDirDialog()) -> ok.

Destructor.

getMessage/1

Specs

getMessage(This) -> unicode:charlist() when This :: wxDirDialog().

Returns the message that will be displayed on the dialog.

getPath/1

Specs

getPath(This) -> unicode:charlist() when This :: wxDirDialog().

Returns the default or user-selected path.

Note: This function can't be used with dialogs which have the wxDD_MULTIPLE style, use GetPaths() (not implemented in wx) instead.

new/1

Specs

new(Parent) -> wxDirDialog() when Parent :: wxWindow:wxWindow().

new/2

Specs

new(Parent, [Option]) -> wxDirDialog()
       when
           Parent :: wxWindow:wxWindow(),
           Option ::
               {title, unicode:chardata()} |
               {defaultPath, unicode:chardata()} |
               {style, integer()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {sz, {W :: integer(), H :: integer()}}.

Constructor.

Use wxDialog:showModal/1 to show the dialog.

setMessage/2

Specs

setMessage(This, Message) -> ok when This :: wxDirDialog(), Message :: unicode:chardata().

Sets the message that will be displayed on the dialog.

setPath/2

Specs

setPath(This, Path) -> ok when This :: wxDirDialog(), Path :: unicode:chardata().

Sets the default path.