wxDisplay (wx v2.4.2)

Functions for wxDisplay class

Determines the sizes and locations of displays connected to the system.

wxWidgets docs: wxDisplay

Types

wxDisplay()

-type wxDisplay() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxDisplay()) -> ok.

Destructor.

getClientArea(This)

-spec getClientArea(This) -> {X :: integer(), Y :: integer(), W :: integer(), H :: integer()}
                 when This :: wxDisplay().

Returns the client area of the display.

The client area is the part of the display available for the normal (non full screen) windows, usually it is the same as getGeometry/1 but it could be less if there is a taskbar (or equivalent) on this display.

getCount()

-spec getCount() -> integer().

Returns the number of connected displays.

getFromPoint(Pt)

-spec getFromPoint(Pt) -> integer() when Pt :: {X :: integer(), Y :: integer()}.

Returns the index of the display on which the given point lies, or wxNOT_FOUND if the point is not on any connected display.

getFromWindow(Win)

-spec getFromWindow(Win) -> integer() when Win :: wxWindow:wxWindow().

Returns the index of the display on which the given window lies.

If the window is on more than one display it gets the display that overlaps the window the most.

Returns wxNOT_FOUND if the window is not on any connected display.

getGeometry(This)

-spec getGeometry(This) -> {X :: integer(), Y :: integer(), W :: integer(), H :: integer()}
               when This :: wxDisplay().

Returns the bounding rectangle of the display whose index was passed to the constructor.

See: getClientArea/1, wx_misc:displaySize/0

getName(This)

-spec getName(This) -> unicode:charlist() when This :: wxDisplay().

Returns the display's name.

The returned value is currently an empty string under all platforms except MSW.

getPPI(This)

-spec getPPI(This) -> {W :: integer(), H :: integer()} when This :: wxDisplay().

Returns display resolution in pixels per inch.

Horizontal and vertical resolution are returned in x and y components of the {Width,Height} object respectively.

If the resolution information is not available, returns.

Since: 3.1.2

isOk(This)

-spec isOk(This) -> boolean() when This :: wxDisplay().

Returns true if the object was initialized successfully.

isPrimary(This)

-spec isPrimary(This) -> boolean() when This :: wxDisplay().

Returns true if the display is the primary display.

The primary display is the one whose index is 0.

new()

-spec new() -> wxDisplay().

Default constructor creating wxDisplay object representing the primary display.

new/1

-spec new(Index) -> wxDisplay() when Index :: integer();
   (Window) -> wxDisplay() when Window :: wxWindow:wxWindow().

Constructor creating the display object associated with the given window.

This is the most convenient way of finding the display on which the given window is shown while falling back to the default display if it is not shown at all or positioned outside of any display.

See: getFromWindow/1

Since: 3.1.2