wxFindReplaceData (wx v2.4.2)

Functions for wxFindReplaceData class

wxFindReplaceData holds the data for wxFindReplaceDialog.

It is used to initialize the dialog with the default values and will keep the last values from the dialog when it is closed. It is also updated each time a wxFindDialogEvent (not implemented in wx) is generated so instead of using the wxFindDialogEvent (not implemented in wx) methods you can also directly query this object.

Note that all SetXXX() methods may only be called before showing the dialog and calling them has no effect later.

wxWidgets docs: wxFindReplaceData

Types

wxFindReplaceData()

-type wxFindReplaceData() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxFindReplaceData()) -> ok.

Destroys the object.

getFindString(This)

-spec getFindString(This) -> unicode:charlist() when This :: wxFindReplaceData().

Get the string to find.

getFlags(This)

-spec getFlags(This) -> integer() when This :: wxFindReplaceData().

Get the combination of wxFindReplaceFlags values.

getReplaceString(This)

-spec getReplaceString(This) -> unicode:charlist() when This :: wxFindReplaceData().

Get the replacement string.

new()

-spec new() -> wxFindReplaceData().

new(Options)

-spec new([Option]) -> wxFindReplaceData() when Option :: {flags, integer()}.

Constructor initializes the flags to default value (0).

setFindString(This, Str)

-spec setFindString(This, Str) -> ok when This :: wxFindReplaceData(), Str :: unicode:chardata().

Set the string to find (used as initial value by the dialog).

setFlags(This, Flags)

-spec setFlags(This, Flags) -> ok when This :: wxFindReplaceData(), Flags :: integer().

Set the flags to use to initialize the controls of the dialog.

setReplaceString(This, Str)

-spec setReplaceString(This, Str) -> ok when This :: wxFindReplaceData(), Str :: unicode:chardata().

Set the replacement string (used as initial value by the dialog).