wxFindReplaceData (wx v2.0.1)

wxFindReplaceData holds the data for wxFindReplaceDialog.

It is used to initialize the dialog with the default values and will keep the last values from the dialog when it is closed. It is also updated each time a wxFindDialogEvent (not implemented in wx) is generated so instead of using the wxFindDialogEvent (not implemented in wx) methods you can also directly query this object.

Note that all SetXXX() methods may only be called before showing the dialog and calling them has no effect later.

wxWidgets docs: wxFindReplaceData

Types

-type wxFindReplaceData() :: term().

Specs

wxFindReplaceData() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxFindReplaceData()) -> ok.

Destroys the object.

getFindString/1

Specs

getFindString(This) -> unicode:charlist() when This :: wxFindReplaceData().

Get the string to find.

getFlags/1

Specs

getFlags(This) -> integer() when This :: wxFindReplaceData().

Get the combination of wxFindReplaceFlags values.

getReplaceString/1

Specs

getReplaceString(This) -> unicode:charlist() when This :: wxFindReplaceData().

Get the replacement string.

new/0

Specs

new() -> wxFindReplaceData().

new/1

Specs

new([Option]) -> wxFindReplaceData() when Option :: {flags, integer()}.

Constructor initializes the flags to default value (0).

setFindString/2

Specs

setFindString(This, Str) -> ok when This :: wxFindReplaceData(), Str :: unicode:chardata().

Set the string to find (used as initial value by the dialog).

setFlags/2

Specs

setFlags(This, Flags) -> ok when This :: wxFindReplaceData(), Flags :: integer().

Set the flags to use to initialize the controls of the dialog.

setReplaceString/2

Specs

setReplaceString(This, Str) -> ok when This :: wxFindReplaceData(), Str :: unicode:chardata().

Set the replacement string (used as initial value by the dialog).