wxFindReplaceDialog (wx v2.0.1)

wxFindReplaceDialog is a standard modeless dialog which is used to allow the user to search for some text (and possibly replace it with something else).

The actual searching is supposed to be done in the owner window which is the parent of this dialog. Note that it means that unlike for the other standard dialogs this one must have a parent window. Also note that there is no way to use this dialog in a modal way; it is always, by design and implementation, modeless.

Please see the page_samples_dialogs sample for an example of using it.

This class is derived (and can use functions) from: wxDialog wxTopLevelWindow wxWindow wxEvtHandler

wxWidgets docs: wxFindReplaceDialog

Types

-type wxFindReplaceDialog() :: term().

Specs

wxFindReplaceDialog() :: wx:wx_object().

Functions

create/4

Specs

create(This, Parent, Data, Title) -> boolean()
          when
              This :: wxFindReplaceDialog(),
              Parent :: wxWindow:wxWindow(),
              Data :: wxFindReplaceData:wxFindReplaceData(),
              Title :: unicode:chardata().

create/5

Specs

create(This, Parent, Data, Title, [Option]) -> boolean()
          when
              This :: wxFindReplaceDialog(),
              Parent :: wxWindow:wxWindow(),
              Data :: wxFindReplaceData:wxFindReplaceData(),
              Title :: unicode:chardata(),
              Option :: {style, integer()}.

Creates the dialog; use wxWindow:show/2 to show it on screen.

The parent and data parameters must be non-NULL.

destroy/1

Specs

destroy(This :: wxFindReplaceDialog()) -> ok.

Destructor.

getData/1

Specs

getData(This) -> wxFindReplaceData:wxFindReplaceData() when This :: wxFindReplaceDialog().

Get the wxFindReplaceData object used by this dialog.

new/0

Specs

new() -> wxFindReplaceDialog().

new/3

Specs

new(Parent, Data, Title) -> wxFindReplaceDialog()
       when
           Parent :: wxWindow:wxWindow(),
           Data :: wxFindReplaceData:wxFindReplaceData(),
           Title :: unicode:chardata().

new/4

Specs

new(Parent, Data, Title, [Option]) -> wxFindReplaceDialog()
       when
           Parent :: wxWindow:wxWindow(),
           Data :: wxFindReplaceData:wxFindReplaceData(),
           Title :: unicode:chardata(),
           Option :: {style, integer()}.

After using default constructor create/5 must be called.

The parent and data parameters must be non-NULL.