wxFlexGridSizer (wx v2.0.1)

A flex grid sizer is a sizer which lays out its children in a two-dimensional table with all table fields in one row having the same height and all fields in one column having the same width, but all rows or all columns are not necessarily the same height or width as in the wxGridSizer.

Since wxWidgets 2.5.0, wxFlexGridSizer can also size items equally in one direction but unequally ("flexibly") in the other. If the sizer is only flexible in one direction (this can be changed using setFlexibleDirection/2), it needs to be decided how the sizer should grow in the other ("non-flexible") direction in order to fill the available space. The setNonFlexibleGrowMode/2 method serves this purpose.

See: wxSizer, Overview sizer

This class is derived (and can use functions) from: wxGridSizer wxSizer

wxWidgets docs: wxFlexGridSizer

Types

-type wxFlexGridSizer() :: term().

Specs

wxFlexGridSizer() :: wx:wx_object().

Functions

addGrowableCol/2

Specs

addGrowableCol(This, Idx) -> ok when This :: wxFlexGridSizer(), Idx :: integer().

addGrowableCol/3

Specs

addGrowableCol(This, Idx, [Option]) -> ok
                  when
                      This :: wxFlexGridSizer(),
                      Idx :: integer(),
                      Option :: {proportion, integer()}.

Specifies that column idx (starting from zero) should be grown if there is extra space available to the sizer.

The proportion parameter has the same meaning as the stretch factor for the sizers (see wxBoxSizer) except that if all proportions are 0, then all columns are resized equally (instead of not being resized at all).

Notice that the column must not be already growable, if you need to change the proportion you must call removeGrowableCol/2 first and then make it growable (with a different proportion) again. You can use IsColGrowable() (not implemented in wx) to check whether a column is already growable.

addGrowableRow/2

Specs

addGrowableRow(This, Idx) -> ok when This :: wxFlexGridSizer(), Idx :: integer().

addGrowableRow/3

Specs

addGrowableRow(This, Idx, [Option]) -> ok
                  when
                      This :: wxFlexGridSizer(),
                      Idx :: integer(),
                      Option :: {proportion, integer()}.

Specifies that row idx (starting from zero) should be grown if there is extra space available to the sizer.

This is identical to addGrowableCol/3 except that it works with rows and not columns.

destroy/1

Specs

destroy(This :: wxFlexGridSizer()) -> ok.

Destroys the object.

getFlexibleDirection/1

Specs

getFlexibleDirection(This) -> integer() when This :: wxFlexGridSizer().

Returns a ?wxOrientation value that specifies whether the sizer flexibly resizes its columns, rows, or both (default).

Return: One of the following values:

See: setFlexibleDirection/2

getNonFlexibleGrowMode/1

Specs

getNonFlexibleGrowMode(This) -> wx:wx_enum() when This :: wxFlexGridSizer().

Returns the value that specifies how the sizer grows in the "non-flexible" direction if there is one.

The behaviour of the elements in the flexible direction (i.e. both rows and columns by default, or rows only if getFlexibleDirection/1 is wxVERTICAL or columns only if it is wxHORIZONTAL) is always governed by their proportion as specified in the call to addGrowableRow/3 or addGrowableCol/3. What happens in the other direction depends on the value of returned by this function as described below.

Return: One of the following values:

See: setFlexibleDirection/2, setNonFlexibleGrowMode/2

new/1

Specs

new(Cols) -> wxFlexGridSizer() when Cols :: integer().

new/2

Specs

new(Cols, [Option]) -> wxFlexGridSizer()
       when Cols :: integer(), Option :: {gap, {W :: integer(), H :: integer()}}.

new/3

Specs

new(Cols, Vgap, Hgap) -> wxFlexGridSizer()
       when Cols :: integer(), Vgap :: integer(), Hgap :: integer();
   (Rows, Cols, Gap) -> wxFlexGridSizer()
       when Rows :: integer(), Cols :: integer(), Gap :: {W :: integer(), H :: integer()}.

new/4

Specs

new(Rows, Cols, Vgap, Hgap) -> wxFlexGridSizer()
       when Rows :: integer(), Cols :: integer(), Vgap :: integer(), Hgap :: integer().

removeGrowableCol/2

Specs

removeGrowableCol(This, Idx) -> ok when This :: wxFlexGridSizer(), Idx :: integer().

Specifies that the idx column index is no longer growable.

removeGrowableRow/2

Specs

removeGrowableRow(This, Idx) -> ok when This :: wxFlexGridSizer(), Idx :: integer().

Specifies that the idx row index is no longer growable.

setFlexibleDirection/2

Specs

setFlexibleDirection(This, Direction) -> ok when This :: wxFlexGridSizer(), Direction :: integer().

Specifies whether the sizer should flexibly resize its columns, rows, or both.

Argument direction can be wxVERTICAL, wxHORIZONTAL or wxBOTH (which is the default value). Any other value is ignored.

See getFlexibleDirection/1 for the explanation of these values. Note that this method does not trigger relayout.

setNonFlexibleGrowMode/2

Specs

setNonFlexibleGrowMode(This, Mode) -> ok when This :: wxFlexGridSizer(), Mode :: wx:wx_enum().

Specifies how the sizer should grow in the non-flexible direction if there is one (so setFlexibleDirection/2 must have been called previously).

Argument mode can be one of those documented in getNonFlexibleGrowMode/1, please see there for their explanation. Note that this method does not trigger relayout.