wxFontData (wx v2.4.3)

This class holds a variety of information related to font dialogs.

See:

Overview cmndlg

wxFont

wxFontDialog

wxWidgets docs: wxFontData

Types

wxFontData()

-type wxFontData() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxFontData()) -> ok.

Destroys the object

enableEffects(This, Enable)

-spec enableEffects(This, Enable) -> ok when This :: wxFontData(), Enable :: boolean().

Enables or disables "effects" under Windows or generic only.

This refers to the controls for manipulating colour, strikeout and underline properties.

The default value is true.

getAllowSymbols(This)

-spec getAllowSymbols(This) -> boolean() when This :: wxFontData().

Under Windows, returns a flag determining whether symbol fonts can be selected.

Has no effect on other platforms.

The default value is true.

getChosenFont(This)

-spec getChosenFont(This) -> wxFont:wxFont() when This :: wxFontData().

Gets the font chosen by the user if the user pressed OK (wxFontDialog::ShowModal() (not implemented in wx) returned wxID_OK).

getColour(This)

-spec getColour(This) -> wx:wx_colour4() when This :: wxFontData().

Gets the colour associated with the font dialog.

The default value is black.

getEnableEffects(This)

-spec getEnableEffects(This) -> boolean() when This :: wxFontData().

Determines whether "effects" are enabled under Windows.

This refers to the controls for manipulating colour, strikeout and underline properties.

The default value is true.

getInitialFont(This)

-spec getInitialFont(This) -> wxFont:wxFont() when This :: wxFontData().

Gets the font that will be initially used by the font dialog.

This should have previously been set by the application.

getShowHelp(This)

-spec getShowHelp(This) -> boolean() when This :: wxFontData().

Returns true if the Help button will be shown (Windows only).

The default value is false.

new()

-spec new() -> wxFontData().

Constructor.

Initializes fontColour to black, showHelp to false, allowSymbols to true, enableEffects to true, minSize to 0 and maxSize to 0.

new(Data)

-spec new(Data) -> wxFontData() when Data :: wxFontData().

Copy Constructor.

setAllowSymbols(This, AllowSymbols)

-spec setAllowSymbols(This, AllowSymbols) -> ok when This :: wxFontData(), AllowSymbols :: boolean().

Under Windows, determines whether symbol fonts can be selected.

Has no effect on other platforms.

The default value is true.

setChosenFont(This, Font)

-spec setChosenFont(This, Font) -> ok when This :: wxFontData(), Font :: wxFont:wxFont().

Sets the font that will be returned to the user (for internal use only).

setColour(This, Colour)

-spec setColour(This, Colour) -> ok when This :: wxFontData(), Colour :: wx:wx_colour().

Sets the colour that will be used for the font foreground colour.

The default colour is black.

setInitialFont(This, Font)

-spec setInitialFont(This, Font) -> ok when This :: wxFontData(), Font :: wxFont:wxFont().

Sets the font that will be initially used by the font dialog.

setRange(This, Min, Max)

-spec setRange(This, Min, Max) -> ok when This :: wxFontData(), Min :: integer(), Max :: integer().

Sets the valid range for the font point size (Windows only).

The default is 0, 0 (unrestricted range).

setShowHelp(This, ShowHelp)

-spec setShowHelp(This, ShowHelp) -> ok when This :: wxFontData(), ShowHelp :: boolean().

Determines whether the Help button will be displayed in the font dialog (Windows only).

The default value is false.