wxFontData (wx v2.5.4)

This class holds a variety of information related to font dialogs.

See:

Overview cmndlg

wxFont

wxFontDialog

wxWidgets docs: wxFontData

Types

wxFontData()

-type wxFontData() :: wx:wx_object().

Functions

destroy(This :: wxFontData()) -> ok

Destroys the object

enableEffects(This :: wxFontData(), Enable :: boolean()) -> ok

Enables or disables "effects" under Windows or generic only.

This refers to the controls for manipulating colour, strikeout and underline properties.

The default value is true.

getAllowSymbols(This :: wxFontData()) -> boolean()

Under Windows, returns a flag determining whether symbol fonts can be selected.

Has no effect on other platforms.

The default value is true.

getChosenFont(This :: wxFontData()) -> wxFont:wxFont()

Gets the font chosen by the user if the user pressed OK (wxFontDialog::ShowModal() (not implemented in wx) returned wxID_OK).

getColour(This :: wxFontData()) -> wx:wx_colour4()

Gets the colour associated with the font dialog.

The default value is black.

getEnableEffects(This :: wxFontData()) -> boolean()

Determines whether "effects" are enabled under Windows.

This refers to the controls for manipulating colour, strikeout and underline properties.

The default value is true.

getInitialFont(This :: wxFontData()) -> wxFont:wxFont()

Gets the font that will be initially used by the font dialog.

This should have previously been set by the application.

getShowHelp(This :: wxFontData()) -> boolean()

Returns true if the Help button will be shown (Windows only).

The default value is false.

new() -> wxFontData()

Constructor.

Initializes fontColour to black, showHelp to false, allowSymbols to true, enableEffects to true, minSize to 0 and maxSize to 0.

new(Data :: wxFontData()) -> wxFontData()

Copy Constructor.

setAllowSymbols(This :: wxFontData(), AllowSymbols :: boolean()) -> ok

Under Windows, determines whether symbol fonts can be selected.

Has no effect on other platforms.

The default value is true.

setChosenFont(This :: wxFontData(), Font :: wxFont:wxFont()) -> ok

Sets the font that will be returned to the user (for internal use only).

setColour(This :: wxFontData(), Colour :: wx:wx_colour()) -> ok

Sets the colour that will be used for the font foreground colour.

The default colour is black.

setInitialFont(This :: wxFontData(), Font :: wxFont:wxFont()) -> ok

Sets the font that will be initially used by the font dialog.

setRange(This :: wxFontData(), Min :: integer(), Max :: integer()) -> ok

Types

Min = Max = integer()

Sets the valid range for the font point size (Windows only).

The default is 0, 0 (unrestricted range).

setShowHelp(This :: wxFontData(), ShowHelp :: boolean()) -> ok

Determines whether the Help button will be displayed in the font dialog (Windows only).

The default value is false.