wxFontData (wx v2.0.1)

This class holds a variety of information related to font dialogs.

See: Overview cmndlg, wxFont, wxFontDialog

wxWidgets docs: wxFontData

Types

-type wxFontData() :: term().

Specs

wxFontData() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxFontData()) -> ok.

Destroys the object.

enableEffects/2

Specs

enableEffects(This, Enable) -> ok when This :: wxFontData(), Enable :: boolean().

Enables or disables "effects" under Windows or generic only.

This refers to the controls for manipulating colour, strikeout and underline properties.

The default value is true.

getAllowSymbols/1

Specs

getAllowSymbols(This) -> boolean() when This :: wxFontData().

Under Windows, returns a flag determining whether symbol fonts can be selected.

Has no effect on other platforms.

The default value is true.

getChosenFont/1

Specs

getChosenFont(This) -> wxFont:wxFont() when This :: wxFontData().

Gets the font chosen by the user if the user pressed OK (wxFontDialog::ShowModal() (not implemented in wx) returned wxID_OK).

getColour/1

Specs

getColour(This) -> wx:wx_colour4() when This :: wxFontData().

Gets the colour associated with the font dialog.

The default value is black.

getEnableEffects/1

Specs

getEnableEffects(This) -> boolean() when This :: wxFontData().

Determines whether "effects" are enabled under Windows.

This refers to the controls for manipulating colour, strikeout and underline properties.

The default value is true.

getInitialFont/1

Specs

getInitialFont(This) -> wxFont:wxFont() when This :: wxFontData().

Gets the font that will be initially used by the font dialog.

This should have previously been set by the application.

getShowHelp/1

Specs

getShowHelp(This) -> boolean() when This :: wxFontData().

Returns true if the Help button will be shown (Windows only).

The default value is false.

new/0

Specs

new() -> wxFontData().

Constructor.

Initializes fontColour to black, showHelp to false, allowSymbols to true, enableEffects to true, minSize to 0 and maxSize to 0.

new/1

Specs

new(Data) -> wxFontData() when Data :: wxFontData().

Copy Constructor.

setAllowSymbols/2

Specs

setAllowSymbols(This, AllowSymbols) -> ok when This :: wxFontData(), AllowSymbols :: boolean().

Under Windows, determines whether symbol fonts can be selected.

Has no effect on other platforms.

The default value is true.

setChosenFont/2

Specs

setChosenFont(This, Font) -> ok when This :: wxFontData(), Font :: wxFont:wxFont().

Sets the font that will be returned to the user (for internal use only).

setColour/2

Specs

setColour(This, Colour) -> ok when This :: wxFontData(), Colour :: wx:wx_colour().

Sets the colour that will be used for the font foreground colour.

The default colour is black.

setInitialFont/2

Specs

setInitialFont(This, Font) -> ok when This :: wxFontData(), Font :: wxFont:wxFont().

Sets the font that will be initially used by the font dialog.

setRange/3

Specs

setRange(This, Min, Max) -> ok when This :: wxFontData(), Min :: integer(), Max :: integer().

Sets the valid range for the font point size (Windows only).

The default is 0, 0 (unrestricted range).

setShowHelp/2

Specs

setShowHelp(This, ShowHelp) -> ok when This :: wxFontData(), ShowHelp :: boolean().

Determines whether the Help button will be displayed in the font dialog (Windows only).

The default value is false.