wxGCDC (wx v2.0.1)

wxGCDC is a device context that draws on a wxGraphicsContext.

wxGCDC does its best to implement wxDC API, but the following features are not (fully) implemented because wxGraphicsContext doesn't support them:

See: wxDC, wxGraphicsContext

This class is derived (and can use functions) from: wxDC

wxWidgets docs: wxGCDC

Types

-type wxGCDC() :: term().

Specs

wxGCDC() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxGCDC()) -> ok.

getGraphicsContext/1

Specs

getGraphicsContext(This) -> wxGraphicsContext:wxGraphicsContext() when This :: wxGCDC().

Retrieves associated wxGraphicsContext.

new/0

Specs

new() -> wxGCDC().

new/1

Specs

new(WindowDC) -> wxGCDC()
       when
           WindowDC ::
               wxWindowDC:wxWindowDC() |
               wxMemoryDC:wxMemoryDC() |
               wxGraphicsContext:wxGraphicsContext().

Constructs a wxGCDC from a wxWindowDC.

setGraphicsContext/2

Specs

setGraphicsContext(This, Context) -> ok
                      when This :: wxGCDC(), Context :: wxGraphicsContext:wxGraphicsContext().

Set the graphics context to be used for this wxGCDC.

Note that this object takes ownership of context and will delete it when it is destroyed or when setGraphicsContext/2 is called again.

Also, unlike the constructor taking wxGraphicsContext, this method will reapply the current font, pen and brush, so that this object continues to use them, if they had been changed before (which is never the case when constructing wxGCDC directly from wxGraphicsContext).