wxGenericDirCtrl (wx v2.0.1)

This control can be used to place a directory listing (with optional files) on an arbitrary window.

The control contains a wxTreeCtrl window representing the directory hierarchy, and optionally, a wxChoice window containing a list of filters.

Styles

This class supports the following styles:

This class is derived (and can use functions) from: wxControl wxWindow wxEvtHandler

wxWidgets docs: wxGenericDirCtrl

Events

Event types emitted from this class: dirctrl_selectionchanged, dirctrl_fileactivated

Types

-type wxGenericDirCtrl() :: term().

Specs

wxGenericDirCtrl() :: wx:wx_object().

Functions

collapseTree/1

Specs

collapseTree(This) -> ok when This :: wxGenericDirCtrl().

Collapses the entire tree.

create/2

Specs

create(This, Parent) -> boolean() when This :: wxGenericDirCtrl(), Parent :: wxWindow:wxWindow().

create/3

Specs

create(This, Parent, [Option]) -> boolean()
          when
              This :: wxGenericDirCtrl(),
              Parent :: wxWindow:wxWindow(),
              Option ::
                  {id, integer()} |
                  {dir, unicode:chardata()} |
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()} |
                  {filter, unicode:chardata()} |
                  {defaultFilter, integer()}.

Create function for two-step construction.

See new/2 for details.

destroy/1

Specs

destroy(This :: wxGenericDirCtrl()) -> ok.

Destructor.

expandPath/2

Specs

expandPath(This, Path) -> boolean() when This :: wxGenericDirCtrl(), Path :: unicode:chardata().

Tries to expand as much of the given path as possible, so that the filename or directory is visible in the tree control.

getDefaultPath/1

Specs

getDefaultPath(This) -> unicode:charlist() when This :: wxGenericDirCtrl().

Gets the default path.

getFilePath/1

Specs

getFilePath(This) -> unicode:charlist() when This :: wxGenericDirCtrl().

Gets selected filename path only (else empty string).

This function doesn't count a directory as a selection.

getFilter/1

Specs

getFilter(This) -> unicode:charlist() when This :: wxGenericDirCtrl().

Returns the filter string.

getFilterIndex/1

Specs

getFilterIndex(This) -> integer() when This :: wxGenericDirCtrl().

Returns the current filter index (zero-based).

getPath/1

Specs

getPath(This) -> unicode:charlist() when This :: wxGenericDirCtrl().

Gets the currently-selected directory or filename.

getPath/2

Specs

getPath(This, ItemId) -> unicode:charlist() when This :: wxGenericDirCtrl(), ItemId :: integer().

Gets the path corresponding to the given tree control item.

Since: 2.9.5

getRootId/1

Specs

getRootId(This) -> integer() when This :: wxGenericDirCtrl().

Returns the root id for the tree control.

getTreeCtrl/1

Specs

getTreeCtrl(This) -> wxTreeCtrl:wxTreeCtrl() when This :: wxGenericDirCtrl().

Returns a pointer to the tree control.

init/1

Specs

init(This) -> ok when This :: wxGenericDirCtrl().

Initializes variables.

new/0

Specs

new() -> wxGenericDirCtrl().

Default constructor.

new/1

Specs

new(Parent) -> wxGenericDirCtrl() when Parent :: wxWindow:wxWindow().

new/2

Specs

new(Parent, [Option]) -> wxGenericDirCtrl()
       when
           Parent :: wxWindow:wxWindow(),
           Option ::
               {id, integer()} |
               {dir, unicode:chardata()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()} |
               {filter, unicode:chardata()} |
               {defaultFilter, integer()}.

Main constructor.

reCreateTree/1

Specs

reCreateTree(This) -> ok when This :: wxGenericDirCtrl().

Collapse and expand the tree, thus re-creating it from scratch.

May be used to update the displayed directory content.

setDefaultPath/2

Specs

setDefaultPath(This, Path) -> ok when This :: wxGenericDirCtrl(), Path :: unicode:chardata().

Sets the default path.

setFilter/2

Specs

setFilter(This, Filter) -> ok when This :: wxGenericDirCtrl(), Filter :: unicode:chardata().

Sets the filter string.

setFilterIndex/2

Specs

setFilterIndex(This, N) -> ok when This :: wxGenericDirCtrl(), N :: integer().

Sets the current filter index (zero-based).

setPath/2

Specs

setPath(This, Path) -> ok when This :: wxGenericDirCtrl(), Path :: unicode:chardata().

Sets the current path.