wxGraphicsFont (wx v2.0.1)

A wxGraphicsFont is a native representation of a font. The contents are specific and private to the respective renderer. Instances are ref counted and can therefore be assigned as usual. The only way to get a valid instance is via wxGraphicsContext:createFont/4 or wxGraphicsRenderer:createFont/4.

This class is derived (and can use functions) from: wxGraphicsObject

wxWidgets docs: wxGraphicsFont

Types

-type wxGraphicsFont() :: term().

Specs

wxGraphicsFont() :: wx:wx_object().