wxGraphicsMatrix (wx v2.0.1)

A wxGraphicsMatrix is a native representation of an affine matrix. The contents are specific and private to the respective renderer. Instances are ref counted and can therefore be assigned as usual. The only way to get a valid instance is via wxGraphicsContext:createMatrix/2 or wxGraphicsRenderer:createMatrix/2.

This class is derived (and can use functions) from: wxGraphicsObject

wxWidgets docs: wxGraphicsMatrix

Types

-type wxGraphicsMatrix() :: term().

Specs

wxGraphicsMatrix() :: wx:wx_object().

Functions

concat/2

Specs

concat(This, T) -> ok when This :: wxGraphicsMatrix(), T :: wxGraphicsMatrix().

Concatenates the matrix passed with the current matrix.

The effect of the resulting transformation is to first apply the transformation in t to the coordinates and then apply the transformation in the current matrix to the coordinates.

get/1

Specs

get(This) -> Result
       when
           Result ::
               {A :: number(),
                B :: number(),
                C :: number(),
                D :: number(),
                Tx :: number(),
                Ty :: number()},
           This :: wxGraphicsMatrix().

Returns the component values of the matrix via the argument pointers.

invert/1

Specs

invert(This) -> ok when This :: wxGraphicsMatrix().

Inverts the matrix.

isEqual/2

Specs

isEqual(This, T) -> boolean() when This :: wxGraphicsMatrix(), T :: wxGraphicsMatrix().

Returns true if the elements of the transformation matrix are equal.

isIdentity/1

Specs

isIdentity(This) -> boolean() when This :: wxGraphicsMatrix().

Return true if this is the identity matrix.

rotate/2

Specs

rotate(This, Angle) -> ok when This :: wxGraphicsMatrix(), Angle :: number().

Rotates this matrix clockwise (in radians).

scale/3

Specs

scale(This, XScale, YScale) -> ok
         when This :: wxGraphicsMatrix(), XScale :: number(), YScale :: number().

Scales this matrix.

set/1

Specs

set(This) -> ok when This :: wxGraphicsMatrix().

set/2

Specs

set(This, [Option]) -> ok
       when
           This :: wxGraphicsMatrix(),
           Option ::
               {a, number()} |
               {b, number()} |
               {c, number()} |
               {d, number()} |
               {tx, number()} |
               {ty, number()}.

Sets the matrix to the respective values (default values are the identity matrix).

transformDistance/1

Specs

transformDistance(This) -> {Dx :: number(), Dy :: number()} when This :: wxGraphicsMatrix().

Applies this matrix to a distance (ie.

performs all transforms except translations).

transformPoint/1

Specs

transformPoint(This) -> {X :: number(), Y :: number()} when This :: wxGraphicsMatrix().

Applies this matrix to a point.

translate/3

Specs

translate(This, Dx, Dy) -> ok when This :: wxGraphicsMatrix(), Dx :: number(), Dy :: number().

Translates this matrix.