wxGridCellAttr (wx v2.0.1)

This class can be used to alter the cells' appearance in the grid by changing their attributes from the defaults. An object of this class may be returned by wxGridTableBase::GetAttr() (not implemented in wx).

Note that objects of this class are reference-counted and it's recommended to use wxGridCellAttrPtr smart pointer class when working with them to avoid memory leaks.

wxWidgets docs: wxGridCellAttr

Types

-type wxGridCellAttr() :: term().

Specs

wxGridCellAttr() :: wx:wx_object().

Functions

getAlignment/1

Specs

getAlignment(This) -> {HAlign :: integer(), VAlign :: integer()} when This :: wxGridCellAttr().

Get the alignment to use for the cell with the given attribute.

If this attribute doesn't specify any alignment, the default attribute alignment is used (which can be changed using wxGrid:setDefaultCellAlignment/3 but is left and top by default).

Notice that hAlign and vAlign values are always overwritten by this function, use GetNonDefaultAlignment() (not implemented in wx) if this is not desirable.

getBackgroundColour/1

Specs

getBackgroundColour(This) -> wx:wx_colour4() when This :: wxGridCellAttr().

Returns the background colour.

getEditor/4

Specs

getEditor(This, Grid, Row, Col) -> wxGridCellEditor:wxGridCellEditor()
             when
                 This :: wxGridCellAttr(),
                 Grid :: wxGrid:wxGrid(),
                 Row :: integer(),
                 Col :: integer().

Returns the cell editor.

The caller is responsible for calling DecRef() (not implemented in wx) on the returned pointer, use GetEditorPtr() (not implemented in wx) to do it automatically.

getFont/1

Specs

getFont(This) -> wxFont:wxFont() when This :: wxGridCellAttr().

Returns the font.

getRenderer/4

Specs

getRenderer(This, Grid, Row, Col) -> wxGridCellRenderer:wxGridCellRenderer()
               when
                   This :: wxGridCellAttr(),
                   Grid :: wxGrid:wxGrid(),
                   Row :: integer(),
                   Col :: integer().

Returns the cell renderer.

The caller is responsible for calling DecRef() (not implemented in wx) on the returned pointer, use GetRendererPtr() (not implemented in wx) to do it automatically.

getTextColour/1

Specs

getTextColour(This) -> wx:wx_colour4() when This :: wxGridCellAttr().

Returns the text colour.

hasAlignment/1

Specs

hasAlignment(This) -> boolean() when This :: wxGridCellAttr().

Returns true if this attribute has a valid alignment set.

hasBackgroundColour/1

Specs

hasBackgroundColour(This) -> boolean() when This :: wxGridCellAttr().

Returns true if this attribute has a valid background colour set.

hasEditor/1

Specs

hasEditor(This) -> boolean() when This :: wxGridCellAttr().

Returns true if this attribute has a valid cell editor set.

hasFont/1

Specs

hasFont(This) -> boolean() when This :: wxGridCellAttr().

Returns true if this attribute has a valid font set.

hasRenderer/1

Specs

hasRenderer(This) -> boolean() when This :: wxGridCellAttr().

Returns true if this attribute has a valid cell renderer set.

hasTextColour/1

Specs

hasTextColour(This) -> boolean() when This :: wxGridCellAttr().

Returns true if this attribute has a valid text colour set.

isReadOnly/1

Specs

isReadOnly(This) -> boolean() when This :: wxGridCellAttr().

Returns true if this cell is set as read-only.

setAlignment/3

Specs

setAlignment(This, HAlign, VAlign) -> ok
                when This :: wxGridCellAttr(), HAlign :: integer(), VAlign :: integer().

Sets the alignment.

hAlign can be one of wxALIGN_LEFT, wxALIGN_CENTRE or wxALIGN_RIGHT and vAlign can be one of wxALIGN_TOP, wxALIGN_CENTRE or wxALIGN_BOTTOM.

setBackgroundColour/2

Specs

setBackgroundColour(This, ColBack) -> ok when This :: wxGridCellAttr(), ColBack :: wx:wx_colour().

Sets the background colour.

setDefAttr/2

Specs

setDefAttr(This, DefAttr) -> ok when This :: wxGridCellAttr(), DefAttr :: wxGridCellAttr().

setEditor/2

Specs

setEditor(This, Editor) -> ok
             when This :: wxGridCellAttr(), Editor :: wxGridCellEditor:wxGridCellEditor().

Sets the editor to be used with the cells with this attribute.

setFont/2

Specs

setFont(This, Font) -> ok when This :: wxGridCellAttr(), Font :: wxFont:wxFont().

Sets the font.

setReadOnly/1

Specs

setReadOnly(This) -> ok when This :: wxGridCellAttr().

setReadOnly/2

Specs

setReadOnly(This, [Option]) -> ok when This :: wxGridCellAttr(), Option :: {isReadOnly, boolean()}.

Sets the cell as read-only.

setRenderer/2

Specs

setRenderer(This, Renderer) -> ok
               when This :: wxGridCellAttr(), Renderer :: wxGridCellRenderer:wxGridCellRenderer().

Sets the renderer to be used for cells with this attribute.

Takes ownership of the pointer.

setTextColour/2

Specs

setTextColour(This, ColText) -> ok when This :: wxGridCellAttr(), ColText :: wx:wx_colour().

Sets the text colour.