wxGridCellBoolEditor (wx v2.0.1)

Grid cell editor for boolean data.

See: wxGridCellEditor, wxGridCellAutoWrapStringEditor (not implemented in wx), wxGridCellChoiceEditor, wxGridCellEnumEditor (not implemented in wx), wxGridCellFloatEditor, wxGridCellNumberEditor, wxGridCellTextEditor, wxGridCellDateEditor (not implemented in wx)

This class is derived (and can use functions) from: wxGridCellEditor

wxWidgets docs: wxGridCellBoolEditor

Types

-type wxGridCellBoolEditor() :: term().

Specs

wxGridCellBoolEditor() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxGridCellBoolEditor()) -> ok.

Destroys the object.

isTrueValue/1

Specs

isTrueValue(Value) -> boolean() when Value :: unicode:chardata().

Returns true if the given value is equal to the string representation of the truth value we currently use (see useStringValues/1).

new/0

Specs

new() -> wxGridCellBoolEditor().

Default constructor.

useStringValues/0

Specs

useStringValues() -> ok.

useStringValues/1

Specs

useStringValues([Option]) -> ok
                   when
                       Option ::
                           {valueTrue, unicode:chardata()} | {valueFalse, unicode:chardata()}.

This method allows you to customize the values returned by wxGridCellNumberEditor:getValue/1 for the cell using this editor.

By default, the default values of the arguments are used, i.e. "1" is returned if the cell is checked and an empty string otherwise.