wxGridCellChoiceEditor (wx v2.0.1)

Grid cell editor for string data providing the user a choice from a list of strings.

See: wxGridCellEditor, wxGridCellAutoWrapStringEditor (not implemented in wx), wxGridCellBoolEditor, wxGridCellEnumEditor (not implemented in wx), wxGridCellFloatEditor, wxGridCellNumberEditor, wxGridCellTextEditor, wxGridCellDateEditor (not implemented in wx)

This class is derived (and can use functions) from: wxGridCellEditor

wxWidgets docs: wxGridCellChoiceEditor

Types

-type wxGridCellChoiceEditor() :: term().

Specs

wxGridCellChoiceEditor() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxGridCellChoiceEditor()) -> ok.

Destroys the object.

new/1

Specs

new(Choices) -> wxGridCellChoiceEditor() when Choices :: [unicode:chardata()].

new/2

Specs

new(Choices, [Option]) -> wxGridCellChoiceEditor()
       when Choices :: [unicode:chardata()], Option :: {allowOthers, boolean()}.

Choice cell renderer ctor.

setParameters/2

Specs

setParameters(This, Params) -> ok
                 when This :: wxGridCellChoiceEditor(), Params :: unicode:chardata().

Parameters string format is "item1[,item2[...,itemN]]".

This method can be called before the editor is used for the first time, or later, in which case it replaces the previously specified strings with the new ones.