wxGridCellEditor (wx v2.0.1)

This class is responsible for providing and manipulating the in-place edit controls for the grid. Instances of wxGridCellEditor (actually, instances of derived classes since it is an abstract class) can be associated with the cell attributes for individual cells, rows, columns, or even for the entire grid.

Normally wxGridCellEditor shows some UI control allowing the user to edit the cell, but starting with wxWidgets 3.1.4 it's also possible to define "activatable" cell editors, that change the value of the cell directly when it's activated (typically by pressing Space key or clicking on it), see TryActivate() (not implemented in wx) method. Note that when implementing an editor which is always activatable, i.e. never shows any in-place editor, it is more convenient to derive its class from wxGridCellActivatableEditor (not implemented in wx) than from wxGridCellEditor itself.

See: wxGridCellAutoWrapStringEditor (not implemented in wx), wxGridCellBoolEditor, wxGridCellChoiceEditor, wxGridCellEnumEditor (not implemented in wx), wxGridCellFloatEditor, wxGridCellNumberEditor, wxGridCellTextEditor, wxGridCellDateEditor (not implemented in wx)

wxWidgets docs: wxGridCellEditor

Types

-type wxGridCellEditor() :: term().

Specs

wxGridCellEditor() :: wx:wx_object().

Functions

create/4

Specs

create(This, Parent, Id, EvtHandler) -> ok
          when
              This :: wxGridCellEditor(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              EvtHandler :: wxEvtHandler:wxEvtHandler().

Creates the actual edit control.

handleReturn/2

Specs

handleReturn(This, Event) -> ok when This :: wxGridCellEditor(), Event :: wxKeyEvent:wxKeyEvent().

Some types of controls on some platforms may need some help with the Return key.

isCreated/1

Specs

isCreated(This) -> boolean() when This :: wxGridCellEditor().

Returns true if the edit control has been created.

reset/1

Specs

reset(This) -> ok when This :: wxGridCellEditor().

Reset the value in the control back to its starting value.

setSize/2

Specs

setSize(This, Rect) -> ok
           when
               This :: wxGridCellEditor(),
               Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()}.

Size and position the edit control.

show/2

Specs

show(This, Show) -> ok when This :: wxGridCellEditor(), Show :: boolean().

show/3

Specs

show(This, Show, [Option]) -> ok
        when
            This :: wxGridCellEditor(),
            Show :: boolean(),
            Option :: {attr, wxGridCellAttr:wxGridCellAttr()}.

Show or hide the edit control, use the specified attributes to set colours/fonts for it.

startingClick/1

Specs

startingClick(This) -> ok when This :: wxGridCellEditor().

If the editor is enabled by clicking on the cell, this method will be called.

startingKey/2

Specs

startingKey(This, Event) -> ok when This :: wxGridCellEditor(), Event :: wxKeyEvent:wxKeyEvent().

If the editor is enabled by pressing keys on the grid, this will be called to let the editor do something about that first key if desired.