wxGridCellNumberEditor (wx v2.4.2)

Functions for wxGridCellNumberEditor class

Grid cell editor for numeric integer data.

See: wxGridCellEditor, wxGridCellAutoWrapStringEditor (not implemented in wx), wxGridCellBoolEditor, wxGridCellChoiceEditor, wxGridCellEnumEditor (not implemented in wx), wxGridCellFloatEditor, wxGridCellTextEditor, wxGridCellDateEditor (not implemented in wx)

This class is derived (and can use functions) from: wxGridCellTextEditor wxGridCellEditor

wxWidgets docs: wxGridCellNumberEditor

Types

wxGridCellNumberEditor()

-type wxGridCellNumberEditor() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxGridCellNumberEditor()) -> ok.

Destroys the object.

getValue(This)

-spec getValue(This) -> unicode:charlist() when This :: wxGridCellNumberEditor().

Returns the value currently in the editor control.

new()

-spec new() -> wxGridCellNumberEditor().

new(Options)

-spec new([Option]) -> wxGridCellNumberEditor() when Option :: {min, integer()} | {max, integer()}.

Allows you to specify the range for acceptable data.

Values equal to -1 for both min and max indicate that no range checking should be done.

setParameters(This, Params)

-spec setParameters(This, Params) -> ok
                 when This :: wxGridCellNumberEditor(), Params :: unicode:chardata().

Parameters string format is "min,max".