wxGridSizer (wx v2.0.1)

A grid sizer is a sizer which lays out its children in a two-dimensional table with all table fields having the same size, i.e. the width of each field is the width of the widest child, the height of each field is the height of the tallest child.

See: wxSizer, Overview sizer

This class is derived (and can use functions) from: wxSizer

wxWidgets docs: wxGridSizer

Types

-type wxGridSizer() :: term().

Specs

wxGridSizer() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxGridSizer()) -> ok.

Destroys the object.

getCols/1

Specs

getCols(This) -> integer() when This :: wxGridSizer().

Returns the number of columns that has been specified for the sizer.

Returns zero if the sizer is automatically adjusting the number of columns depending on number of its children. To get the effective number of columns or rows being currently used, see GetEffectiveColsCount() (not implemented in wx)

getHGap/1

Specs

getHGap(This) -> integer() when This :: wxGridSizer().

Returns the horizontal gap (in pixels) between cells in the sizer.

getRows/1

Specs

getRows(This) -> integer() when This :: wxGridSizer().

Returns the number of rows that has been specified for the sizer.

Returns zero if the sizer is automatically adjusting the number of rows depending on number of its children. To get the effective number of columns or rows being currently used, see GetEffectiveRowsCount() (not implemented in wx).

getVGap/1

Specs

getVGap(This) -> integer() when This :: wxGridSizer().

Returns the vertical gap (in pixels) between the cells in the sizer.

new/1

Specs

new(Cols) -> wxGridSizer() when Cols :: integer().

new/2

Specs

new(Cols, [Option]) -> wxGridSizer()
       when Cols :: integer(), Option :: {gap, {W :: integer(), H :: integer()}}.

new/3

Specs

new(Cols, Vgap, Hgap) -> wxGridSizer()
       when Cols :: integer(), Vgap :: integer(), Hgap :: integer();
   (Rows, Cols, Gap) -> wxGridSizer()
       when Rows :: integer(), Cols :: integer(), Gap :: {W :: integer(), H :: integer()}.

new/4

Specs

new(Rows, Cols, Vgap, Hgap) -> wxGridSizer()
       when Rows :: integer(), Cols :: integer(), Vgap :: integer(), Hgap :: integer().

setCols/2

Specs

setCols(This, Cols) -> ok when This :: wxGridSizer(), Cols :: integer().

Sets the number of columns in the sizer.

setHGap/2

Specs

setHGap(This, Gap) -> ok when This :: wxGridSizer(), Gap :: integer().

Sets the horizontal gap (in pixels) between cells in the sizer.

setRows/2

Specs

setRows(This, Rows) -> ok when This :: wxGridSizer(), Rows :: integer().

Sets the number of rows in the sizer.

setVGap/2

Specs

setVGap(This, Gap) -> ok when This :: wxGridSizer(), Gap :: integer().

Sets the vertical gap (in pixels) between the cells in the sizer.