wxHtmlEasyPrinting (wx v2.0.1)

This class provides very simple interface to printing architecture. It allows you to print HTML documents using only a few commands.

Note: Do not create this class on the stack only. You should create an instance on app startup and use this instance for all printing operations. The reason is that this class stores various settings in it.

wxWidgets docs: wxHtmlEasyPrinting

Types

-type wxHtmlEasyPrinting() :: term().

Specs

wxHtmlEasyPrinting() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxHtmlEasyPrinting()) -> ok.

Destroys the object.

getPageSetupData/1

Specs

getPageSetupData(This) -> wxPageSetupDialogData:wxPageSetupDialogData()
                    when This :: wxHtmlEasyPrinting().

Returns a pointer to wxPageSetupDialogData instance used by this class.

You can set its parameters (via SetXXXX methods).

getPrintData/1

Specs

getPrintData(This) -> wxPrintData:wxPrintData() when This :: wxHtmlEasyPrinting().

Returns pointer to wxPrintData instance used by this class.

You can set its parameters (via SetXXXX methods).

new/0

Specs

new() -> wxHtmlEasyPrinting().

new/1

Specs

new([Option]) -> wxHtmlEasyPrinting()
       when Option :: {name, unicode:chardata()} | {parentWindow, wxWindow:wxWindow()}.

Constructor.

pageSetup/1

Specs

pageSetup(This) -> ok when This :: wxHtmlEasyPrinting().

Display page setup dialog and allows the user to modify settings.

previewFile/2

Specs

previewFile(This, Htmlfile) -> boolean()
               when This :: wxHtmlEasyPrinting(), Htmlfile :: unicode:chardata().

Preview HTML file.

Returns false in case of error - call wxPrinter:getLastError/0 to get detailed information about the kind of the error.

previewText/2

Specs

previewText(This, Htmltext) -> boolean()
               when This :: wxHtmlEasyPrinting(), Htmltext :: unicode:chardata().

previewText/3

Specs

previewText(This, Htmltext, [Option]) -> boolean()
               when
                   This :: wxHtmlEasyPrinting(),
                   Htmltext :: unicode:chardata(),
                   Option :: {basepath, unicode:chardata()}.

Preview HTML text (not file!).

Returns false in case of error - call wxPrinter:getLastError/0 to get detailed information about the kind of the error.

printFile/2

Specs

printFile(This, Htmlfile) -> boolean()
             when This :: wxHtmlEasyPrinting(), Htmlfile :: unicode:chardata().

Print HTML file.

Returns false in case of error - call wxPrinter:getLastError/0 to get detailed information about the kind of the error.

printText/2

Specs

printText(This, Htmltext) -> boolean()
             when This :: wxHtmlEasyPrinting(), Htmltext :: unicode:chardata().

printText/3

Specs

printText(This, Htmltext, [Option]) -> boolean()
             when
                 This :: wxHtmlEasyPrinting(),
                 Htmltext :: unicode:chardata(),
                 Option :: {basepath, unicode:chardata()}.

Print HTML text (not file!).

Returns false in case of error - call wxPrinter:getLastError/0 to get detailed information about the kind of the error.

setFonts/3

Specs

setFonts(This, Normal_face, Fixed_face) -> ok
            when
                This :: wxHtmlEasyPrinting(),
                Normal_face :: unicode:chardata(),
                Fixed_face :: unicode:chardata().

setFonts/4

Specs

setFonts(This, Normal_face, Fixed_face, [Option]) -> ok
            when
                This :: wxHtmlEasyPrinting(),
                Normal_face :: unicode:chardata(),
                Fixed_face :: unicode:chardata(),
                Option :: {sizes, [integer()]}.

Sets fonts.

See wxHtmlDCRenderer::SetFonts (not implemented in wx) for detailed description.

setFooter/2

Specs

setFooter(This, Footer) -> ok when This :: wxHtmlEasyPrinting(), Footer :: unicode:chardata().

setFooter/3

Specs

setFooter(This, Footer, [Option]) -> ok
             when
                 This :: wxHtmlEasyPrinting(),
                 Footer :: unicode:chardata(),
                 Option :: {pg, integer()}.

Set page footer.

The following macros can be used inside it: @DATE@ is replaced by the current date in default format @PAGENUM@ is replaced by page number @PAGESCNT@ is replaced by total number of pages @TIME@ is replaced by the current time in default format @TITLE@ is replaced with the title of the document

setHeader/2

Specs

setHeader(This, Header) -> ok when This :: wxHtmlEasyPrinting(), Header :: unicode:chardata().

setHeader/3

Specs

setHeader(This, Header, [Option]) -> ok
             when
                 This :: wxHtmlEasyPrinting(),
                 Header :: unicode:chardata(),
                 Option :: {pg, integer()}.

Set page header.

The following macros can be used inside it: