wxImageList (wx v2.0.1)

A wxImageList contains a list of images, which are stored in an unspecified form. Images can have masks for transparent drawing, and can be made from a variety of sources including bitmaps and icons.

wxImageList is used principally in conjunction with wxTreeCtrl and wxListCtrl classes.

See: wxTreeCtrl, wxListCtrl

wxWidgets docs: wxImageList

Types

-type wxImageList() :: term().

Specs

wxImageList() :: wx:wx_object().

Functions

add/2

Specs

add(This, Icon) -> integer()
       when This :: wxImageList(), Icon :: wxIcon:wxIcon() | wxBitmap:wxBitmap().

Adds a new image using an icon.

Return: The new zero-based image index.

Remark: The original bitmap or icon is not affected by the add/3 operation, and can be deleted afterwards. If the bitmap is wider than the images in the list, then the bitmap will automatically be split into smaller images, each matching the dimensions of the image list. This does not apply when adding icons.

Only for:wxmsw,wxosx

add/3

Specs

add(This, Bitmap, MaskColour) -> integer()
       when This :: wxImageList(), Bitmap :: wxBitmap:wxBitmap(), MaskColour :: wx:wx_colour();
   (This, Bitmap, Mask) -> integer()
       when This :: wxImageList(), Bitmap :: wxBitmap:wxBitmap(), Mask :: wxBitmap:wxBitmap().

Adds a new image or images using a bitmap and optional mask bitmap.

Return: The new zero-based image index.

Remark: The original bitmap or icon is not affected by the add/3 operation, and can be deleted afterwards. If the bitmap is wider than the images in the list, then the bitmap will automatically be split into smaller images, each matching the dimensions of the image list. This does not apply when adding icons.

create/3

Specs

create(This, Width, Height) -> boolean()
          when This :: wxImageList(), Width :: integer(), Height :: integer().

create/4

Specs

create(This, Width, Height, [Option]) -> boolean()
          when
              This :: wxImageList(),
              Width :: integer(),
              Height :: integer(),
              Option :: {mask, boolean()} | {initialCount, integer()}.

Initializes the list.

See new/3 for details.

destroy/1

Specs

destroy(This :: wxImageList()) -> ok.

Destroys the object.

draw/5

Specs

draw(This, Index, Dc, X, Y) -> boolean()
        when
            This :: wxImageList(),
            Index :: integer(),
            Dc :: wxDC:wxDC(),
            X :: integer(),
            Y :: integer().

draw/6

Specs

draw(This, Index, Dc, X, Y, [Option]) -> boolean()
        when
            This :: wxImageList(),
            Index :: integer(),
            Dc :: wxDC:wxDC(),
            X :: integer(),
            Y :: integer(),
            Option :: {flags, integer()} | {solidBackground, boolean()}.

Draws a specified image onto a device context.

getBitmap/2

Specs

getBitmap(This, Index) -> wxBitmap:wxBitmap() when This :: wxImageList(), Index :: integer().

Returns the bitmap corresponding to the given index.

getIcon/2

Specs

getIcon(This, Index) -> wxIcon:wxIcon() when This :: wxImageList(), Index :: integer().

Returns the icon corresponding to the given index.

getImageCount/1

Specs

getImageCount(This) -> integer() when This :: wxImageList().

Returns the number of images in the list.

getSize/2

Specs

getSize(This, Index) -> Result
           when
               Result :: {Res :: boolean(), Width :: integer(), Height :: integer()},
               This :: wxImageList(),
               Index :: integer().

Retrieves the size of the images in the list.

Currently, the index parameter is ignored as all images in the list have the same size.

Return: true if the function succeeded, false if it failed (for example, if the image list was not yet initialized).

new/0

Specs

new() -> wxImageList().

Default ctor.

new/2

Specs

new(Width, Height) -> wxImageList() when Width :: integer(), Height :: integer().

new/3

Specs

new(Width, Height, [Option]) -> wxImageList()
       when
           Width :: integer(),
           Height :: integer(),
           Option :: {mask, boolean()} | {initialCount, integer()}.

Constructor specifying the image size, whether image masks should be created, and the initial size of the list.

See: create/4

remove/2

Specs

remove(This, Index) -> boolean() when This :: wxImageList(), Index :: integer().

Removes the image at the given position.

removeAll/1

Specs

removeAll(This) -> boolean() when This :: wxImageList().

Removes all the images in the list.

replace/3

Specs

replace(This, Index, Icon) -> boolean()
           when
               This :: wxImageList(),
               Index :: integer(),
               Icon :: wxIcon:wxIcon() | wxBitmap:wxBitmap().

Replaces the existing image with the new image.

Return: true if the replacement was successful, false otherwise.

Remark: The original bitmap or icon is not affected by the replace/4 operation, and can be deleted afterwards.

Only for:wxmsw,wxosx

replace/4

Specs

replace(This, Index, Bitmap, Mask) -> boolean()
           when
               This :: wxImageList(),
               Index :: integer(),
               Bitmap :: wxBitmap:wxBitmap(),
               Mask :: wxBitmap:wxBitmap().

Replaces the existing image with the new image.

Windows only.

Return: true if the replacement was successful, false otherwise.

Remark: The original bitmap or icon is not affected by the replace/4 operation, and can be deleted afterwards.