wxJoystickEvent (wx v2.0.1)

This event class contains information about joystick events, particularly events received by windows.

See: wxJoystick (not implemented in wx)

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxJoystickEvent

Events

Use wxEvtHandler:connect/3 with wxJoystickEventType to subscribe to events of this type.

Types

-type wxJoystick() :: term().

Specs

wxJoystick() :: #wxJoystick{}.

-type wxJoystickEvent() :: term().

Specs

wxJoystickEvent() :: wx:wx_object().

-type wxJoystickEventType() :: term().

Specs

wxJoystickEventType() :: joy_button_down | joy_button_up | joy_move | joy_zmove.

Functions

buttonDown/1

Specs

buttonDown(This) -> boolean() when This :: wxJoystickEvent().

buttonDown/2

Specs

buttonDown(This, [Option]) -> boolean() when This :: wxJoystickEvent(), Option :: {but, integer()}.

Returns true if the event was a down event from the specified button (or any button).

buttonIsDown/1

Specs

buttonIsDown(This) -> boolean() when This :: wxJoystickEvent().

buttonIsDown/2

Specs

buttonIsDown(This, [Option]) -> boolean()
                when This :: wxJoystickEvent(), Option :: {but, integer()}.

Returns true if the specified button (or any button) was in a down state.

buttonUp/1

Specs

buttonUp(This) -> boolean() when This :: wxJoystickEvent().

buttonUp/2

Specs

buttonUp(This, [Option]) -> boolean() when This :: wxJoystickEvent(), Option :: {but, integer()}.

Returns true if the event was an up event from the specified button (or any button).

getButtonChange/1

Specs

getButtonChange(This) -> integer() when This :: wxJoystickEvent().

Returns the identifier of the button changing state.

The return value is where n is the index of the button changing state, which can also be retrieved using GetButtonOrdinal() (not implemented in wx).

Note that for n equal to 1, 2, 3 or 4 there are predefined wxJOY_BUTTONn constants which can be used for more clarity, however these constants are not defined for the buttons beyond the first four.

getButtonState/1

Specs

getButtonState(This) -> integer() when This :: wxJoystickEvent().

Returns the down state of the buttons.

This is a wxJOY_BUTTONn identifier, where n is one of 1, 2, 3, 4.

getJoystick/1

Specs

getJoystick(This) -> integer() when This :: wxJoystickEvent().

Returns the identifier of the joystick generating the event - one of wxJOYSTICK1 and wxJOYSTICK2.

getPosition/1

Specs

getPosition(This) -> {X :: integer(), Y :: integer()} when This :: wxJoystickEvent().

Returns the x, y position of the joystick event.

These coordinates are valid for all the events except wxEVT_JOY_ZMOVE.

getZPosition/1

Specs

getZPosition(This) -> integer() when This :: wxJoystickEvent().

Returns the z position of the joystick event.

This method can only be used for wxEVT_JOY_ZMOVE events.

isButton/1

Specs

isButton(This) -> boolean() when This :: wxJoystickEvent().

Returns true if this was a button up or down event (not 'is any button down?').

isMove/1

Specs

isMove(This) -> boolean() when This :: wxJoystickEvent().

Returns true if this was an x, y move event.

isZMove/1

Specs

isZMove(This) -> boolean() when This :: wxJoystickEvent().

Returns true if this was a z move event.