wxListItem (wx v2.0.1)

This class stores information about a wxListCtrl item or column.

wxListItem is a class which contains information about:

The wxListItem object can also contain item-specific colour and font information: for this you need to call one of setTextColour/2, setBackgroundColour/2 or setFont/2 functions on it passing it the colour/font to use. If the colour/font is not specified, the default list control colour/font is used.

See: wxListCtrl

wxWidgets docs: wxListItem

Types

-type wxListItem() :: term().

Specs

wxListItem() :: wx:wx_object().

Functions

clear/1

Specs

clear(This) -> ok when This :: wxListItem().

Resets the item state to the default.

destroy/1

Specs

destroy(This :: wxListItem()) -> ok.

Destroys the object.

getAlign/1

Specs

getAlign(This) -> wx:wx_enum() when This :: wxListItem().

Returns the alignment for this item.

Can be one of wxLIST_FORMAT_LEFT, wxLIST_FORMAT_RIGHT or wxLIST_FORMAT_CENTRE.

getBackgroundColour/1

Specs

getBackgroundColour(This) -> wx:wx_colour4() when This :: wxListItem().

Returns the background colour for this item.

getColumn/1

Specs

getColumn(This) -> integer() when This :: wxListItem().

Returns the zero-based column; meaningful only in report mode.

getFont/1

Specs

getFont(This) -> wxFont:wxFont() when This :: wxListItem().

Returns the font used to display the item.

getId/1

Specs

getId(This) -> integer() when This :: wxListItem().

Returns the zero-based item position.

getImage/1

Specs

getImage(This) -> integer() when This :: wxListItem().

Returns the zero-based index of the image associated with the item into the image list.

getMask/1

Specs

getMask(This) -> integer() when This :: wxListItem().

Returns a bit mask indicating which fields of the structure are valid.

Can be any combination of the following values:

getState/1

Specs

getState(This) -> integer() when This :: wxListItem().

Returns a bit field representing the state of the item.

Can be any combination of:

getText/1

Specs

getText(This) -> unicode:charlist() when This :: wxListItem().

Returns the label/header text.

getTextColour/1

Specs

getTextColour(This) -> wx:wx_colour4() when This :: wxListItem().

Returns the text colour.

getWidth/1

Specs

getWidth(This) -> integer() when This :: wxListItem().

Meaningful only for column headers in report mode.

Returns the column width.

new/0

Specs

new() -> wxListItem().

Constructor.

new/1

Specs

new(Item) -> wxListItem() when Item :: wxListItem().

setAlign/2

Specs

setAlign(This, Align) -> ok when This :: wxListItem(), Align :: wx:wx_enum().

Sets the alignment for the item.

See also getAlign/1

setBackgroundColour/2

Specs

setBackgroundColour(This, ColBack) -> ok when This :: wxListItem(), ColBack :: wx:wx_colour().

Sets the background colour for the item.

setColumn/2

Specs

setColumn(This, Col) -> ok when This :: wxListItem(), Col :: integer().

Sets the zero-based column.

Meaningful only in report mode.

setFont/2

Specs

setFont(This, Font) -> ok when This :: wxListItem(), Font :: wxFont:wxFont().

Sets the font for the item.

setId/2

Specs

setId(This, Id) -> ok when This :: wxListItem(), Id :: integer().

Sets the zero-based item position.

setImage/2

Specs

setImage(This, Image) -> ok when This :: wxListItem(), Image :: integer().

Sets the zero-based index of the image associated with the item into the image list.

setMask/2

Specs

setMask(This, Mask) -> ok when This :: wxListItem(), Mask :: integer().

Sets the mask of valid fields.

See getMask/1.

setState/2

Specs

setState(This, State) -> ok when This :: wxListItem(), State :: integer().

Sets the item state flags (note that the valid state flags are influenced by the value of the state mask, see setStateMask/2).

See getState/1 for valid flag values.

setStateMask/2

Specs

setStateMask(This, StateMask) -> ok when This :: wxListItem(), StateMask :: integer().

Sets the bitmask that is used to determine which of the state flags are to be set.

See also setState/2.

setText/2

Specs

setText(This, Text) -> ok when This :: wxListItem(), Text :: unicode:chardata().

Sets the text label for the item.

setTextColour/2

Specs

setTextColour(This, ColText) -> ok when This :: wxListItem(), ColText :: wx:wx_colour().

Sets the text colour for the item.

setWidth/2

Specs

setWidth(This, Width) -> ok when This :: wxListItem(), Width :: integer().

Meaningful only for column headers in report mode.

Sets the column width.