wxListbook (wx v2.0.1)

wxListbook is a class similar to wxNotebook but which uses a wxListCtrl to show the labels instead of the tabs.

The underlying wxListCtrl displays page labels in a one-column report view by default. Calling wxBookCtrl::SetImageList will implicitly switch the control to use an icon view.

For usage documentation of this class, please refer to the base abstract class wxBookCtrl. You can also use the page_samples_notebook to see wxListbook in action.

Styles

This class supports the following styles:

See: ?wxBookCtrl, wxNotebook, Examples

This class is derived (and can use functions) from: wxBookCtrlBase wxControl wxWindow wxEvtHandler

wxWidgets docs: wxListbook

Events

Event types emitted from this class: listbook_page_changed, listbook_page_changing

Types

-type wxListbook() :: term().

Specs

wxListbook() :: wx:wx_object().

Functions

addPage/3

Specs

addPage(This, Page, Text) -> boolean()
           when This :: wxListbook(), Page :: wxWindow:wxWindow(), Text :: unicode:chardata().

addPage/4

Specs

addPage(This, Page, Text, [Option]) -> boolean()
           when
               This :: wxListbook(),
               Page :: wxWindow:wxWindow(),
               Text :: unicode:chardata(),
               Option :: {bSelect, boolean()} | {imageId, integer()}.

Adds a new page.

The page must have the book control itself as the parent and must not have been added to this control previously.

The call to this function will generate the page changing and page changed events if select is true, but not when inserting the very first page (as there is no previous page selection to switch from in this case and so it wouldn't make sense to e.g. veto such event).

Return: true if successful, false otherwise.

Remark: Do not delete the page, it will be deleted by the book control.

See: insertPage/5

advanceSelection/1

Specs

advanceSelection(This) -> ok when This :: wxListbook().

advanceSelection/2

Specs

advanceSelection(This, [Option]) -> ok when This :: wxListbook(), Option :: {forward, boolean()}.

Cycles through the tabs.

The call to this function generates the page changing events.

assignImageList/2

Specs

assignImageList(This, ImageList) -> ok
                   when This :: wxListbook(), ImageList :: wxImageList:wxImageList().

Sets the image list for the page control and takes ownership of the list.

See: wxImageList, setImageList/2

changeSelection/2

Specs

changeSelection(This, Page) -> integer() when This :: wxListbook(), Page :: integer().

Changes the selection to the given page, returning the previous selection.

This function behaves as setSelection/2 but does not generate the page changing events.

See overview_events_prog for more information.

create/3

Specs

create(This, Parent, Id) -> boolean()
          when This :: wxListbook(), Parent :: wxWindow:wxWindow(), Id :: integer().

create/4

Specs

create(This, Parent, Id, [Option]) -> boolean()
          when
              This :: wxListbook(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Option ::
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()}.

Create the list book control that has already been constructed with the default constructor.

deleteAllPages/1

Specs

deleteAllPages(This) -> boolean() when This :: wxListbook().

Deletes all pages.

destroy/1

Specs

destroy(This :: wxListbook()) -> ok.

Destroys the object.

getCurrentPage/1

Specs

getCurrentPage(This) -> wxWindow:wxWindow() when This :: wxListbook().

Returns the currently selected page or NULL.

getImageList/1

Specs

getImageList(This) -> wxImageList:wxImageList() when This :: wxListbook().

Returns the associated image list, may be NULL.

See: wxImageList, setImageList/2

getPage/2

Specs

getPage(This, Page) -> wxWindow:wxWindow() when This :: wxListbook(), Page :: integer().

Returns the window at the given page position.

getPageCount/1

Specs

getPageCount(This) -> integer() when This :: wxListbook().

Returns the number of pages in the control.

getPageImage/2

Specs

getPageImage(This, NPage) -> integer() when This :: wxListbook(), NPage :: integer().

Returns the image index for the given page.

getPageText/2

Specs

getPageText(This, NPage) -> unicode:charlist() when This :: wxListbook(), NPage :: integer().

Returns the string for the given page.

getSelection/1

Specs

getSelection(This) -> integer() when This :: wxListbook().

Returns the currently selected page, or wxNOT_FOUND if none was selected.

Note that this method may return either the previously or newly selected page when called from the EVT_BOOKCTRL_PAGE_CHANGED handler depending on the platform and so wxBookCtrlEvent:getSelection/1 should be used instead in this case.

hitTest/2

Specs

hitTest(This, Pt) -> Result
           when
               Result :: {Res :: integer(), Flags :: integer()},
               This :: wxListbook(),
               Pt :: {X :: integer(), Y :: integer()}.

Returns the index of the tab at the specified position or wxNOT_FOUND if none.

If flags parameter is non-NULL, the position of the point inside the tab is returned as well.

Return: Returns the zero-based tab index or wxNOT_FOUND if there is no tab at the specified position.

insertPage/4

Specs

insertPage(This, Index, Page, Text) -> boolean()
              when
                  This :: wxListbook(),
                  Index :: integer(),
                  Page :: wxWindow:wxWindow(),
                  Text :: unicode:chardata().

insertPage/5

Specs

insertPage(This, Index, Page, Text, [Option]) -> boolean()
              when
                  This :: wxListbook(),
                  Index :: integer(),
                  Page :: wxWindow:wxWindow(),
                  Text :: unicode:chardata(),
                  Option :: {bSelect, boolean()} | {imageId, integer()}.

Inserts a new page at the specified position.

Return: true if successful, false otherwise.

Remark: Do not delete the page, it will be deleted by the book control.

See: addPage/4

new/0

Specs

new() -> wxListbook().

Default ctor.

new/2

Specs

new(Parent, Id) -> wxListbook() when Parent :: wxWindow:wxWindow(), Id :: integer().

new/3

Specs

new(Parent, Id, [Option]) -> wxListbook()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Option ::
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Constructs a listbook control.

setImageList/2

Specs

setImageList(This, ImageList) -> ok
                when This :: wxListbook(), ImageList :: wxImageList:wxImageList().

Sets the image list to use.

It does not take ownership of the image list, you must delete it yourself.

See: wxImageList, assignImageList/2

setPageImage/3

Specs

setPageImage(This, Page, Image) -> boolean()
                when This :: wxListbook(), Page :: integer(), Image :: integer().

Sets the image index for the given page.

image is an index into the image list which was set with setImageList/2.

setPageSize/2

Specs

setPageSize(This, Size) -> ok when This :: wxListbook(), Size :: {W :: integer(), H :: integer()}.

Sets the width and height of the pages.

Note: This method is currently not implemented for wxGTK.

setPageText/3

Specs

setPageText(This, Page, Text) -> boolean()
               when This :: wxListbook(), Page :: integer(), Text :: unicode:chardata().

Sets the text for the given page.

setSelection/2

Specs

setSelection(This, Page) -> integer() when This :: wxListbook(), Page :: integer().

Sets the selection to the given page, returning the previous selection.

Notice that the call to this function generates the page changing events, use the changeSelection/2 function if you don't want these events to be generated.

See: getSelection/1