wxLocale (wx v2.0.1)

wxLocale class encapsulates all language-dependent settings and is a generalization of the C locale concept.

In wxWidgets this class manages current locale. It also initializes and activates wxTranslations (not implemented in wx) object that manages message catalogs.

For a list of the supported languages, please see ?wxLanguage enum values. These constants may be used to specify the language in init/3 and are returned by getSystemLanguage/0.

See: Overview i18n, Examples, wxXLocale (not implemented in wx), wxTranslations (not implemented in wx)

wxWidgets docs: wxLocale

Types

-type wxLocale() :: term().

Specs

wxLocale() :: wx:wx_object().

Functions

addCatalog/2

Specs

addCatalog(This, Domain) -> boolean() when This :: wxLocale(), Domain :: unicode:chardata().

Calls wxTranslations::AddCatalog(const wxString&).

addCatalog/3

Specs

addCatalog(This, Domain, MsgIdLanguage) -> boolean()
              when This :: wxLocale(), Domain :: unicode:chardata(), MsgIdLanguage :: wx:wx_enum().

Calls wxTranslations::AddCatalog(const wxString&, wxLanguage) (not implemented in wx).

addCatalog/4

Specs

addCatalog(This, Domain, MsgIdLanguage, MsgIdCharset) -> boolean()
              when
                  This :: wxLocale(),
                  Domain :: unicode:chardata(),
                  MsgIdLanguage :: wx:wx_enum(),
                  MsgIdCharset :: unicode:chardata().

Calls wxTranslations::AddCatalog(const wxString&, wxLanguage, const wxString&) (not implemented in wx).

addCatalogLookupPathPrefix/1

Specs

addCatalogLookupPathPrefix(Prefix) -> ok when Prefix :: unicode:chardata().

Calls wxFileTranslationsLoader::AddCatalogLookupPathPrefix() (not implemented in wx).

destroy/1

Specs

destroy(This :: wxLocale()) -> ok.

The destructor, like the constructor, also has global side effects: the previously set locale is restored and so the changes described in init/3 documentation are rolled back.

getCanonicalName/1

Specs

getCanonicalName(This) -> unicode:charlist() when This :: wxLocale().

Returns the canonical form of current locale name.

Canonical form is the one that is used on UNIX systems: it is a two- or five-letter string in xx or xx_YY format, where xx is ISO 639 code of language and YY is ISO 3166 code of the country. Examples are "en", "en_GB", "en_US" or "fr_FR". This form is internally used when looking up message catalogs. Compare getSysName/1.

getHeaderValue/2

Specs

getHeaderValue(This, Header) -> unicode:charlist()
                  when This :: wxLocale(), Header :: unicode:chardata().

getHeaderValue/3

Specs

getHeaderValue(This, Header, [Option]) -> unicode:charlist()
                  when
                      This :: wxLocale(),
                      Header :: unicode:chardata(),
                      Option :: {szDomain, unicode:chardata()}.

Calls wxTranslations::GetHeaderValue() (not implemented in wx).

getLanguage/1

Specs

getLanguage(This) -> integer() when This :: wxLocale().

Returns the ?wxLanguage constant of current language.

Note that you can call this function only if you used the form of init/3 that takes ?wxLanguage argument.

getLanguageName/1

Specs

getLanguageName(Lang) -> unicode:charlist() when Lang :: integer().

Returns English name of the given language or empty string if this language is unknown.

See GetLanguageInfo() (not implemented in wx) for a remark about special meaning of wxLANGUAGE_DEFAULT.

getLocale/1

Specs

getLocale(This) -> unicode:charlist() when This :: wxLocale().

Returns the locale name as passed to the constructor or init/3.

This is a full, human-readable name, e.g. "English" or "French".

getName/1

Specs

getName(This) -> unicode:charlist() when This :: wxLocale().

Returns the current short name for the locale (as given to the constructor or the init/3 function).

getString/2

Specs

getString(This, OrigString) -> unicode:charlist()
             when This :: wxLocale(), OrigString :: unicode:chardata().

getString/3

Specs

getString(This, OrigString, [Option]) -> unicode:charlist()
             when
                 This :: wxLocale(),
                 OrigString :: unicode:chardata(),
                 Option :: {szDomain, unicode:chardata()}.

Calls wxGetTranslation(const wxString&, const wxString&).

getString/4

Specs

getString(This, OrigString, OrigString2, N) -> unicode:charlist()
             when
                 This :: wxLocale(),
                 OrigString :: unicode:chardata(),
                 OrigString2 :: unicode:chardata(),
                 N :: integer().

getString/5

Specs

getString(This, OrigString, OrigString2, N, [Option]) -> unicode:charlist()
             when
                 This :: wxLocale(),
                 OrigString :: unicode:chardata(),
                 OrigString2 :: unicode:chardata(),
                 N :: integer(),
                 Option :: {szDomain, unicode:chardata()}.

Calls wxGetTranslation(const wxString&, const wxString&, unsigned, const wxString&).

getSysName/1

Specs

getSysName(This) -> unicode:charlist() when This :: wxLocale().

Returns current platform-specific locale name as passed to setlocale().

Compare getCanonicalName/1.

getSystemEncoding/0

Specs

getSystemEncoding() -> wx:wx_enum().

Tries to detect the user's default font encoding.

Returns ?wxFontEncoding() value or wxFONTENCODING_SYSTEM if it couldn't be determined.

getSystemEncodingName/0

Specs

getSystemEncodingName() -> unicode:charlist().

Tries to detect the name of the user's default font encoding.

This string isn't particularly useful for the application as its form is platform-dependent and so you should probably use getSystemEncoding/0 instead.

Returns a user-readable string value or an empty string if it couldn't be determined.

getSystemLanguage/0

Specs

getSystemLanguage() -> integer().

Tries to detect the user's default locale setting.

Returns the ?wxLanguage value or wxLANGUAGE_UNKNOWN if the language-guessing algorithm failed.

Note: This function works with locales and returns the user's default locale. This may be, and usually is, the same as their preferred UI language, but it's not the same thing. Use wxTranslation to obtain language information.

See: wxTranslations::GetBestTranslation() (not implemented in wx)

init/1

Specs

init(This) -> boolean() when This :: wxLocale().

init/2

Specs

init(This, [Option]) -> boolean()
        when This :: wxLocale(), Option :: {language, integer()} | {flags, integer()}.

Initializes the wxLocale instance.

The call of this function has several global side effects which you should understand: first of all, the application locale is changed - note that this will affect many of standard C library functions such as printf() or strftime(). Second, this wxLocale object becomes the new current global locale for the application and so all subsequent calls to ?wxGetTranslation() will try to translate the messages using the message catalogs for this locale.

Return: true on success or false if the given locale couldn't be set.

init/3

Specs

init(This, Name, [Option]) -> boolean()
        when
            This :: wxLocale(),
            Name :: unicode:chardata(),
            Option ::
                {shortName, unicode:chardata()} |
                {locale, unicode:chardata()} |
                {bLoadDefault, boolean()}.

Deprecated: This form is deprecated, use the other one unless you know what you are doing.

isLoaded/2

Specs

isLoaded(This, Domain) -> boolean() when This :: wxLocale(), Domain :: unicode:chardata().

Calls wxTranslations::IsLoaded() (not implemented in wx).

isOk/1

Specs

isOk(This) -> boolean() when This :: wxLocale().

Returns true if the locale could be set successfully.

new/0

Specs

new() -> wxLocale().

This is the default constructor and it does nothing to initialize the object: init/3 must be used to do that.

new/1

Specs

new(Language) -> wxLocale() when Language :: integer();
   (Name) -> wxLocale() when Name :: unicode:chardata().

new/2

Specs

new(Language, [Option]) -> wxLocale() when Language :: integer(), Option :: {flags, integer()};
   (Name, [Option]) -> wxLocale()
       when
           Name :: unicode:chardata(),
           Option ::
               {shortName, unicode:chardata()} |
               {locale, unicode:chardata()} |
               {bLoadDefault, boolean()}.

See init/3 for parameters description.

The call of this function has several global side effects which you should understand: first of all, the application locale is changed - note that this will affect many of standard C library functions such as printf() or strftime(). Second, this wxLocale object becomes the new current global locale for the application and so all subsequent calls to ?wxGetTranslation() will try to translate the messages using the message catalogs for this locale.