wxMask (wx v2.0.1)

This class encapsulates a monochrome mask bitmap, where the masked area is black and the unmasked area is white.

When associated with a bitmap and drawn in a device context, the unmasked area of the bitmap will be drawn, and the masked area will not be drawn.

Note: A mask can be associated also with a bitmap with an alpha channel but drawing such bitmaps under wxMSW may be slow so using them should be avoided if drawing performance is an important factor.

See: wxBitmap, wxDC:blit/6, wxMemoryDC

wxWidgets docs: wxMask

Types

-type wxMask() :: term().

Specs

wxMask() :: wx:wx_object().

Functions

create/2

Specs

create(This, Bitmap) -> boolean() when This :: wxMask(), Bitmap :: wxBitmap:wxBitmap().

Constructs a mask from a monochrome bitmap.

create/3

Specs

create(This, Bitmap, Index) -> boolean()
          when This :: wxMask(), Bitmap :: wxBitmap:wxBitmap(), Index :: integer();
      (This, Bitmap, Colour) -> boolean()
          when This :: wxMask(), Bitmap :: wxBitmap:wxBitmap(), Colour :: wx:wx_colour().

Constructs a mask from a bitmap and a colour that indicates the background.

destroy/1

Specs

destroy(This :: wxMask()) -> ok.

Destroys the wxMask object and the underlying bitmap data.

new/0

Specs

new() -> wxMask().

Default constructor.

new/1

Specs

new(Bitmap) -> wxMask() when Bitmap :: wxBitmap:wxBitmap().

Constructs a mask from a monochrome bitmap.

new/2

Specs

new(Bitmap, Index) -> wxMask() when Bitmap :: wxBitmap:wxBitmap(), Index :: integer();
   (Bitmap, Colour) -> wxMask() when Bitmap :: wxBitmap:wxBitmap(), Colour :: wx:wx_colour().

Constructs a mask from a bitmap and a colour that indicates the background.