wxMenuBar (wx v2.0.1)

A menu bar is a series of menus accessible from the top of a frame.

Remark: To respond to a menu selection, provide a handler for EVT_MENU, in the frame that contains the menu bar.

If you have a toolbar which uses the same identifiers as your EVT_MENU entries, events from the toolbar will also be processed by your EVT_MENU event handlers.

Tip: under Windows, if you discover that menu shortcuts (for example, Alt-F to show the file menu) are not working, check any EVT_CHAR events you are handling in child windows. If you are not calling event.Skip() for events that you don't process in these event handlers, menu shortcuts may cease to work.

See: wxMenu, Overview events

This class is derived (and can use functions) from: wxWindow wxEvtHandler

wxWidgets docs: wxMenuBar

Types

-type wxMenuBar() :: term().

Specs

wxMenuBar() :: wx:wx_object().

Functions

append/3

Specs

append(This, Menu, Title) -> boolean()
          when This :: wxMenuBar(), Menu :: wxMenu:wxMenu(), Title :: unicode:chardata().

Adds the item to the end of the menu bar.

Return: true on success, false if an error occurred.

See: insert/4

check/3

Specs

check(This, Id, Check) -> ok when This :: wxMenuBar(), Id :: integer(), Check :: boolean().

Checks or unchecks a menu item.

Remark: Only use this when the menu bar has been associated with a frame; otherwise, use the wxMenu equivalent call.

destroy/1

Specs

destroy(This :: wxMenuBar()) -> ok.

Destructor, destroying the menu bar and removing it from the parent frame (if any).

enable/3

Specs

enable(This, Id, Enable) -> ok when This :: wxMenuBar(), Id :: integer(), Enable :: boolean().

Enables or disables (greys out) a menu item.

Remark: Only use this when the menu bar has been associated with a frame; otherwise, use the wxMenu equivalent call.

enableTop/3

Specs

enableTop(This, Pos, Enable) -> ok when This :: wxMenuBar(), Pos :: integer(), Enable :: boolean().

Enables or disables a whole menu.

Remark: Only use this when the menu bar has been associated with a frame.

findItem/2

Specs

findItem(This, Id) -> wxMenuItem:wxMenuItem() when This :: wxMenuBar(), Id :: integer().

Finds the menu item object associated with the given menu item identifier.

Return: The found menu item object, or NULL if one was not found.

findMenu/2

Specs

findMenu(This, Title) -> integer() when This :: wxMenuBar(), Title :: unicode:chardata().

Returns the index of the menu with the given title or wxNOT_FOUND if no such menu exists in this menubar.

The title parameter may specify either the menu title (with accelerator characters, i.e. "&File") or just the menu label ("File") indifferently.

findMenuItem/3

Specs

findMenuItem(This, MenuString, ItemString) -> integer()
                when
                    This :: wxMenuBar(),
                    MenuString :: unicode:chardata(),
                    ItemString :: unicode:chardata().

Finds the menu item id for a menu name/menu item string pair.

Return: The menu item identifier, or wxNOT_FOUND if none was found.

Remark: Any special menu codes are stripped out of source and target strings before matching.

getAutoWindowMenu/0

Specs

getAutoWindowMenu() -> boolean().

getHelpString/2

Specs

getHelpString(This, Id) -> unicode:charlist() when This :: wxMenuBar(), Id :: integer().

Gets the help string associated with the menu item identifier.

Return: The help string, or the empty string if there was no help string or the menu item was not found.

See: setHelpString/3

getLabel/2

Specs

getLabel(This, Id) -> unicode:charlist() when This :: wxMenuBar(), Id :: integer().

Gets the label associated with a menu item.

Return: The menu item label, or the empty string if the item was not found.

Remark: Use only after the menubar has been associated with a frame.

getLabelTop/2

Specs

getLabelTop(This, Pos) -> unicode:charlist() when This :: wxMenuBar(), Pos :: integer().

See: getMenuLabel/2.

getMenu/2

Specs

getMenu(This, MenuIndex) -> wxMenu:wxMenu() when This :: wxMenuBar(), MenuIndex :: integer().

Returns the menu at menuIndex (zero-based).

getMenuCount/1

Specs

getMenuCount(This) -> integer() when This :: wxMenuBar().

Returns the number of menus in this menubar.

getMenuLabel/2

Specs

getMenuLabel(This, Pos) -> unicode:charlist() when This :: wxMenuBar(), Pos :: integer().

Returns the label of a top-level menu.

Note that the returned string includes the accelerator characters that have been specified in the menu title string during its construction.

Return: The menu label, or the empty string if the menu was not found.

Remark: Use only after the menubar has been associated with a frame.

See: getMenuLabelText/2, setMenuLabel/3

getMenuLabelText/2

Specs

getMenuLabelText(This, Pos) -> unicode:charlist() when This :: wxMenuBar(), Pos :: integer().

Returns the label of a top-level menu.

Note that the returned string does not include any accelerator characters that may have been specified in the menu title string during its construction.

Return: The menu label, or the empty string if the menu was not found.

Remark: Use only after the menubar has been associated with a frame.

See: getMenuLabel/2, setMenuLabel/3

insert/4

Specs

insert(This, Pos, Menu, Title) -> boolean()
          when
              This :: wxMenuBar(),
              Pos :: integer(),
              Menu :: wxMenu:wxMenu(),
              Title :: unicode:chardata().

Inserts the menu at the given position into the menu bar.

Inserting menu at position 0 will insert it in the very beginning of it, inserting at position getMenuCount/1 is the same as calling append/3.

Return: true on success, false if an error occurred.

See: append/3

isChecked/2

Specs

isChecked(This, Id) -> boolean() when This :: wxMenuBar(), Id :: integer().

Determines whether an item is checked.

Return: true if the item was found and is checked, false otherwise.

isEnabled/2

Specs

isEnabled(This, Id) -> boolean() when This :: wxMenuBar(), Id :: integer().

Determines whether an item is enabled.

Return: true if the item was found and is enabled, false otherwise.

new/0

Specs

new() -> wxMenuBar().

Construct an empty menu bar.

new/1

Specs

new(Style) -> wxMenuBar() when Style :: integer().

oSXGetAppleMenu/1

Specs

oSXGetAppleMenu(This) -> wxMenu:wxMenu() when This :: wxMenuBar().

Returns the Apple menu.

This is the leftmost menu with application's name as its title. You shouldn't remove any items from it, but it is safe to insert extra menu items or submenus into it.

Only for:wxosx

Since: 3.0.1

remove/2

Specs

remove(This, Pos) -> wxMenu:wxMenu() when This :: wxMenuBar(), Pos :: integer().

Removes the menu from the menu bar and returns the menu object - the caller is responsible for deleting it.

This function may be used together with insert/4 to change the menubar dynamically.

See: replace/4

replace/4

Specs

replace(This, Pos, Menu, Title) -> wxMenu:wxMenu()
           when
               This :: wxMenuBar(),
               Pos :: integer(),
               Menu :: wxMenu:wxMenu(),
               Title :: unicode:chardata().

Replaces the menu at the given position with another one.

Return: The menu which was previously at position pos. The caller is responsible for deleting it.

See: insert/4, remove/2

setAutoWindowMenu/1

Specs

setAutoWindowMenu(Enable) -> ok when Enable :: boolean().

setHelpString/3

Specs

setHelpString(This, Id, HelpString) -> ok
                 when This :: wxMenuBar(), Id :: integer(), HelpString :: unicode:chardata().

Sets the help string associated with a menu item.

See: getHelpString/2

setLabel/3

Specs

setLabel(This, Id, Label) -> ok
            when This :: wxMenuBar(), Id :: integer(), Label :: unicode:chardata().

Sets the label of a menu item.

Remark: Use only after the menubar has been associated with a frame.

See: getLabel/2

setLabelTop/3

Specs

setLabelTop(This, Pos, Label) -> ok
               when This :: wxMenuBar(), Pos :: integer(), Label :: unicode:chardata().

See: setMenuLabel/3.

setMenuLabel/3

Specs

setMenuLabel(This, Pos, Label) -> ok
                when This :: wxMenuBar(), Pos :: integer(), Label :: unicode:chardata().

Sets the label of a top-level menu.

Remark: Use only after the menubar has been associated with a frame.