wxMenuEvent (wx v2.0.1)

This class is used for a variety of menu-related events. Note that these do not include menu command events, which are handled using wxCommandEvent objects.

Events of this class are generated by both menus that are part of a wxMenuBar, attached to wxFrame, and popup menus shown by wxWindow:popupMenu/4. They are sent to the following objects until one of them handles the event: -# The menu object itself, as returned by GetMenu(), if any. -# The wxMenuBar to which this menu is attached, if any. -# The window associated with the menu, e.g. the one calling PopupMenu() for the popup menus. -# The top level parent of that window if it's different from the window itself.

This is similar to command events generated by the menu items, but, unlike them, wxMenuEvent are only sent to the window itself and its top level parent but not any intermediate windows in the hierarchy.

The default handler for wxEVT_MENU_HIGHLIGHT in wxFrame displays help text in the status bar, see wxFrame:setStatusBarPane/2.

See: wxCommandEvent, Overview events

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxMenuEvent

Events

Use wxEvtHandler:connect/3 with wxMenuEventType to subscribe to events of this type.

Types

-type wxMenu() :: term().

Specs

wxMenu() :: #wxMenu{}.

-type wxMenuEvent() :: term().

Specs

wxMenuEvent() :: wx:wx_object().

-type wxMenuEventType() :: term().

Specs

wxMenuEventType() :: menu_open | menu_close | menu_highlight.

Functions

getMenu/1

Specs

getMenu(This) -> wxMenu:wxMenu() when This :: wxMenuEvent().

Returns the menu which is being opened or closed, or the menu containing the highlighted item.

Note that the returned value can be NULL if the menu being opened doesn't have a corresponding wxMenu, e.g. this happens when opening the system menu in wxMSW port.

Remark: Since 3.1.3 this function can be used with OPEN, CLOSE and HIGHLIGHT events. Before 3.1.3, this method can only be used with the OPEN and CLOSE events.

getMenuId/1

Specs

getMenuId(This) -> integer() when This :: wxMenuEvent().

Returns the menu identifier associated with the event.

This method should be only used with the HIGHLIGHT events.

isPopup/1

Specs

isPopup(This) -> boolean() when This :: wxMenuEvent().

Returns true if the menu which is being opened or closed is a popup menu, false if it is a normal one.

This method should only be used with the OPEN and CLOSE events.