wxMouseEvent (wx v2.0.1)

This event class contains information about the events generated by the mouse: they include mouse buttons press and release events and mouse move events.

All mouse events involving the buttons use wxMOUSE_BTN_LEFT for the left mouse button, wxMOUSE_BTN_MIDDLE for the middle one and wxMOUSE_BTN_RIGHT for the right one. And if the system supports more buttons, the wxMOUSE_BTN_AUX1 and wxMOUSE_BTN_AUX2 events can also be generated. Note that not all mice have even a middle button so a portable application should avoid relying on the events from it (but the right button click can be emulated using the left mouse button with the control key under Mac platforms with a single button mouse).

For the wxEVT_ENTER_WINDOW and wxEVT_LEAVE_WINDOW events purposes, the mouse is considered to be inside the window if it is in the window client area and not inside one of its children. In other words, the parent window receives wxEVT_LEAVE_WINDOW event not only when the mouse leaves the window entirely but also when it enters one of its children.

The position associated with a mouse event is expressed in the window coordinates of the window which generated the event, you can use wxWindow:clientToScreen/3 to convert it to screen coordinates and possibly call wxWindow:screenToClient/2 next to convert it to window coordinates of another window.

Note: Note the difference between methods like leftDown/1 and the inherited leftIsDown/1: the former returns true when the event corresponds to the left mouse button click while the latter returns true if the left mouse button is currently being pressed. For example, when the user is dragging the mouse you can use leftIsDown/1 to test whether the left mouse button is (still) depressed. Also, by convention, if leftDown/1 returns true, leftIsDown/1 will also return true in wxWidgets whatever the underlying GUI behaviour is (which is platform-dependent). The same applies, of course, to other mouse buttons as well.

See: wxKeyEvent

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxMouseEvent

Events

Use wxEvtHandler:connect/3 with wxMouseEventType to subscribe to events of this type.

Types

-type wxMouse() :: term().

Specs

wxMouse() :: #wxMouse{}.

-type wxMouseEvent() :: term().

Specs

wxMouseEvent() :: wx:wx_object().

-type wxMouseEventType() :: term().

Specs

wxMouseEventType() ::
    left_down | left_up | middle_down | middle_up | right_down | right_up | motion |
    enter_window | leave_window | left_dclick | middle_dclick | right_dclick | mousewheel |
    aux1_down | aux1_up | aux1_dclick | aux2_down | aux2_up | aux2_dclick.

Functions

altDown/1

Specs

altDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the Alt key is pressed.

Notice that wxKeyEvent:getModifiers/1 should usually be used instead of this one.

button/2

Specs

button(This, But) -> boolean() when This :: wxMouseEvent(), But :: wx:wx_enum().

Returns true if the event was generated by the specified button.

See: wxMouseState::ButtoinIsDown()

buttonDClick/1

Specs

buttonDClick(This) -> boolean() when This :: wxMouseEvent().

buttonDClick/2

Specs

buttonDClick(This, [Option]) -> boolean()
                when This :: wxMouseEvent(), Option :: {but, wx:wx_enum()}.

If the argument is omitted, this returns true if the event was a mouse double click event.

Otherwise the argument specifies which double click event was generated (see button/2 for the possible values).

buttonDown/1

Specs

buttonDown(This) -> boolean() when This :: wxMouseEvent().

buttonDown/2

Specs

buttonDown(This, [Option]) -> boolean() when This :: wxMouseEvent(), Option :: {but, wx:wx_enum()}.

If the argument is omitted, this returns true if the event was a mouse button down event.

Otherwise the argument specifies which button-down event was generated (see button/2 for the possible values).

buttonUp/1

Specs

buttonUp(This) -> boolean() when This :: wxMouseEvent().

buttonUp/2

Specs

buttonUp(This, [Option]) -> boolean() when This :: wxMouseEvent(), Option :: {but, wx:wx_enum()}.

If the argument is omitted, this returns true if the event was a mouse button up event.

Otherwise the argument specifies which button-up event was generated (see button/2 for the possible values).

cmdDown/1

Specs

cmdDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the key used for command accelerators is pressed.

Same as controlDown/1. Deprecated.

Notice that wxKeyEvent:getModifiers/1 should usually be used instead of this one.

controlDown/1

Specs

controlDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the Control key or Apple/Command key under macOS is pressed.

This function doesn't distinguish between right and left control keys.

Notice that wxKeyEvent:getModifiers/1 should usually be used instead of this one.

dragging/1

Specs

dragging(This) -> boolean() when This :: wxMouseEvent().

Returns true if this was a dragging event (motion while a button is depressed).

See: moving/1

entering/1

Specs

entering(This) -> boolean() when This :: wxMouseEvent().

Returns true if the mouse was entering the window.

See: leaving/1

getButton/1

Specs

getButton(This) -> integer() when This :: wxMouseEvent().

Returns the mouse button which generated this event or wxMOUSE_BTN_NONE if no button is involved (for mouse move, enter or leave event, for example).

Otherwise wxMOUSE_BTN_LEFT is returned for the left button down, up and double click events, wxMOUSE_BTN_MIDDLE and wxMOUSE_BTN_RIGHT for the same events for the middle and the right buttons respectively.

getLinesPerAction/1

Specs

getLinesPerAction(This) -> integer() when This :: wxMouseEvent().

Returns the configured number of lines (or whatever) to be scrolled per wheel action.

Default value under most platforms is three.

See: GetColumnsPerAction() (not implemented in wx)

getLogicalPosition/2

Specs

getLogicalPosition(This, Dc) -> {X :: integer(), Y :: integer()}
                      when This :: wxMouseEvent(), Dc :: wxDC:wxDC().

Returns the logical mouse position in pixels (i.e. translated according to the translation set for the DC, which usually indicates that the window has been scrolled).

getPosition/1

Specs

getPosition(This) -> {X :: integer(), Y :: integer()} when This :: wxMouseEvent().

Returns the physical mouse position.

getWheelAxis/1

Specs

getWheelAxis(This) -> wx:wx_enum() when This :: wxMouseEvent().

Gets the axis the wheel operation concerns.

Usually the mouse wheel is used to scroll vertically so wxMOUSE_WHEEL_VERTICAL is returned but some mice (and most trackpads) also allow to use the wheel to scroll horizontally in which case wxMOUSE_WHEEL_HORIZONTAL is returned.

Notice that before wxWidgets 2.9.4 this method returned int.

getWheelDelta/1

Specs

getWheelDelta(This) -> integer() when This :: wxMouseEvent().

Get wheel delta, normally 120.

This is the threshold for action to be taken, and one such action (for example, scrolling one increment) should occur for each delta.

getWheelRotation/1

Specs

getWheelRotation(This) -> integer() when This :: wxMouseEvent().

Get wheel rotation, positive or negative indicates direction of rotation.

Current devices all send an event when rotation is at least +/-WheelDelta, but finer resolution devices can be created in the future.

Because of this you shouldn't assume that one event is equal to 1 line, but you should be able to either do partial line scrolling or wait until several events accumulate before scrolling.

getX/1

Specs

getX(This) -> integer() when This :: wxMouseEvent().

Returns X coordinate of the physical mouse event position.

getY/1

Specs

getY(This) -> integer() when This :: wxMouseEvent().

Returns Y coordinate of the physical mouse event position.

isButton/1

Specs

isButton(This) -> boolean() when This :: wxMouseEvent().

Returns true if the event was a mouse button event (not necessarily a button down event - that may be tested using buttonDown/2).

isPageScroll/1

Specs

isPageScroll(This) -> boolean() when This :: wxMouseEvent().

Returns true if the system has been setup to do page scrolling with the mouse wheel instead of line scrolling.

leaving/1

Specs

leaving(This) -> boolean() when This :: wxMouseEvent().

Returns true if the mouse was leaving the window.

See: entering/1

leftDClick/1

Specs

leftDClick(This) -> boolean() when This :: wxMouseEvent().

Returns true if the event was a left double click.

leftDown/1

Specs

leftDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the left mouse button changed to down.

leftIsDown/1

Specs

leftIsDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the left mouse button is currently down.

leftUp/1

Specs

leftUp(This) -> boolean() when This :: wxMouseEvent().

Returns true if the left mouse button changed to up.

metaDown/1

Specs

metaDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the Meta key was down at the time of the event.

middleDClick/1

Specs

middleDClick(This) -> boolean() when This :: wxMouseEvent().

Returns true if the event was a middle double click.

middleDown/1

Specs

middleDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the middle mouse button changed to down.

middleIsDown/1

Specs

middleIsDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the middle mouse button is currently down.

middleUp/1

Specs

middleUp(This) -> boolean() when This :: wxMouseEvent().

Returns true if the middle mouse button changed to up.

moving/1

Specs

moving(This) -> boolean() when This :: wxMouseEvent().

Returns true if this was a motion event and no mouse buttons were pressed.

If any mouse button is held pressed, then this method returns false and dragging/1 returns true.

rightDClick/1

Specs

rightDClick(This) -> boolean() when This :: wxMouseEvent().

Returns true if the event was a right double click.

rightDown/1

Specs

rightDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the right mouse button changed to down.

rightIsDown/1

Specs

rightIsDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the right mouse button is currently down.

rightUp/1

Specs

rightUp(This) -> boolean() when This :: wxMouseEvent().

Returns true if the right mouse button changed to up.

shiftDown/1

Specs

shiftDown(This) -> boolean() when This :: wxMouseEvent().

Returns true if the Shift key is pressed.

This function doesn't distinguish between right and left shift keys.

Notice that wxKeyEvent:getModifiers/1 should usually be used instead of this one.