wxMultiChoiceDialog (wx v2.4.2)

Functions for wxMultiChoiceDialog class

This class represents a dialog that shows a list of strings, and allows the user to select one or more.

Styles

This class supports the following styles:

See: Overview cmndlg, wxSingleChoiceDialog

This class is derived (and can use functions) from: wxDialog wxTopLevelWindow wxWindow wxEvtHandler

wxWidgets docs: wxMultiChoiceDialog

Types

wxMultiChoiceDialog()

-type wxMultiChoiceDialog() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxMultiChoiceDialog()) -> ok.

Destroys the object.

getSelections(This)

-spec getSelections(This) -> [integer()] when This :: wxMultiChoiceDialog().

Returns array with indexes of selected items.

new(Parent, Message, Caption, Choices)

-spec new(Parent, Message, Caption, Choices) -> wxMultiChoiceDialog()
       when
           Parent :: wxWindow:wxWindow(),
           Message :: unicode:chardata(),
           Caption :: unicode:chardata(),
           Choices :: [unicode:chardata()].

new/5

-spec new(Parent, Message, Caption, Choices, [Option]) -> wxMultiChoiceDialog()
       when
           Parent :: wxWindow:wxWindow(),
           Message :: unicode:chardata(),
           Caption :: unicode:chardata(),
           Choices :: [unicode:chardata()],
           Option :: {style, integer()} | {pos, {X :: integer(), Y :: integer()}}.

Constructor taking an array of wxString (not implemented in wx) choices.

Remark: Use wxDialog:showModal/1 to show the dialog.

setSelections(This, Selections)

-spec setSelections(This, Selections) -> ok
                 when This :: wxMultiChoiceDialog(), Selections :: [integer()].

Sets selected items from the array of selected items' indexes.