wxNavigationKeyEvent (wx v2.0.1)

This event class contains information about navigation events, generated by navigation keys such as tab and page down.

This event is mainly used by wxWidgets implementations. A wxNavigationKeyEvent handler is automatically provided by wxWidgets when you enable keyboard navigation inside a window by inheriting it from wxNavigationEnabled<>.

See: wxWindow:navigate/2, wxWindow::NavigateIn (not implemented in wx)

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxNavigationKeyEvent

Events

Use wxEvtHandler:connect/3 with wxNavigationKeyEventType to subscribe to events of this type.

Types

-type wxNavigationKey() :: term().

Specs

wxNavigationKey() :: #wxNavigationKey{}.

-type wxNavigationKeyEvent() :: term().

Specs

wxNavigationKeyEvent() :: wx:wx_object().

-type wxNavigationKeyEventType() :: term().

Specs

wxNavigationKeyEventType() :: navigation_key.

Functions

getCurrentFocus/1

Specs

getCurrentFocus(This) -> wxWindow:wxWindow() when This :: wxNavigationKeyEvent().

Returns the child that has the focus, or NULL.

getDirection/1

Specs

getDirection(This) -> boolean() when This :: wxNavigationKeyEvent().

Returns true if the navigation was in the forward direction.

isFromTab/1

Specs

isFromTab(This) -> boolean() when This :: wxNavigationKeyEvent().

Returns true if the navigation event was from a tab key.

This is required for proper navigation over radio buttons.

isWindowChange/1

Specs

isWindowChange(This) -> boolean() when This :: wxNavigationKeyEvent().

Returns true if the navigation event represents a window change (for example, from Ctrl-Page Down in a notebook).

setCurrentFocus/2

Specs

setCurrentFocus(This, CurrentFocus) -> ok
                   when This :: wxNavigationKeyEvent(), CurrentFocus :: wxWindow:wxWindow().

Sets the current focus window member.

setDirection/2

Specs

setDirection(This, Direction) -> ok when This :: wxNavigationKeyEvent(), Direction :: boolean().

Sets the direction to forward if direction is true, or backward if false.

setFromTab/2

Specs

setFromTab(This, FromTab) -> ok when This :: wxNavigationKeyEvent(), FromTab :: boolean().

Marks the navigation event as from a tab key.

setWindowChange/2

Specs

setWindowChange(This, WindowChange) -> ok
                   when This :: wxNavigationKeyEvent(), WindowChange :: boolean().

Marks the event as a window change event.