wxPageSetupDialogData (wx v2.0.1)

This class holds a variety of information related to wxPageSetupDialog.

It contains a wxPrintData member which is used to hold basic printer configuration data (as opposed to the user-interface configuration settings stored by wxPageSetupDialogData).

See: Overview printing, wxPageSetupDialog

wxWidgets docs: wxPageSetupDialogData

Types

-type wxPageSetupDialogData() :: term().

Specs

wxPageSetupDialogData() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxPageSetupDialogData()) -> ok.

Destructor.

enableHelp/2

Specs

enableHelp(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the "Help" button (Windows only).

enableMargins/2

Specs

enableMargins(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the margin controls (Windows only).

enableOrientation/2

Specs

enableOrientation(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the orientation control (Windows only).

enablePaper/2

Specs

enablePaper(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the paper size control (Windows only).

enablePrinter/2

Specs

enablePrinter(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Enables or disables the "Printer" button, which invokes a printer setup dialog.

getDefaultInfo/1

Specs

getDefaultInfo(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the dialog will simply return default printer information (such as orientation) instead of showing a dialog (Windows only).

getDefaultMinMargins/1

Specs

getDefaultMinMargins(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the page setup dialog will take its minimum margin values from the currently selected printer properties (Windows only).

getEnableHelp/1

Specs

getEnableHelp(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the printer setup button is enabled.

getEnableMargins/1

Specs

getEnableMargins(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the margin controls are enabled (Windows only).

getEnableOrientation/1

Specs

getEnableOrientation(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the orientation control is enabled (Windows only).

getEnablePaper/1

Specs

getEnablePaper(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the paper size control is enabled (Windows only).

getEnablePrinter/1

Specs

getEnablePrinter(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the printer setup button is enabled.

getMarginBottomRight/1

Specs

getMarginBottomRight(This) -> {X :: integer(), Y :: integer()}
                        when This :: wxPageSetupDialogData().

Returns the right (x) and bottom (y) margins in millimetres.

getMarginTopLeft/1

Specs

getMarginTopLeft(This) -> {X :: integer(), Y :: integer()} when This :: wxPageSetupDialogData().

Returns the left (x) and top (y) margins in millimetres.

getMinMarginBottomRight/1

Specs

getMinMarginBottomRight(This) -> {X :: integer(), Y :: integer()}
                           when This :: wxPageSetupDialogData().

Returns the right (x) and bottom (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

getMinMarginTopLeft/1

Specs

getMinMarginTopLeft(This) -> {X :: integer(), Y :: integer()} when This :: wxPageSetupDialogData().

Returns the left (x) and top (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

getPaperId/1

Specs

getPaperId(This) -> wx:wx_enum() when This :: wxPageSetupDialogData().

Returns the paper id (stored in the internal wxPrintData object).

See: wxPrintData:setPaperId/2

getPaperSize/1

Specs

getPaperSize(This) -> {W :: integer(), H :: integer()} when This :: wxPageSetupDialogData().

Returns the paper size in millimetres.

getPrintData/1

Specs

getPrintData(This) -> wxPrintData:wxPrintData() when This :: wxPageSetupDialogData().

isOk/1

Specs

isOk(This) -> boolean() when This :: wxPageSetupDialogData().

Returns true if the print data associated with the dialog data is valid.

This can return false on Windows if the current printer is not set, for example. On all other platforms, it returns true.

new/0

Specs

new() -> wxPageSetupDialogData().

Default constructor.

new/1

Specs

new(PrintData) -> wxPageSetupDialogData()
       when PrintData :: wxPrintData:wxPrintData() | wxPageSetupDialogData:wxPageSetupDialogData().

Construct an object from a print data object.

setDefaultInfo/2

Specs

setDefaultInfo(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Pass true if the dialog will simply return default printer information (such as orientation) instead of showing a dialog (Windows only).

setDefaultMinMargins/2

Specs

setDefaultMinMargins(This, Flag) -> ok when This :: wxPageSetupDialogData(), Flag :: boolean().

Pass true if the page setup dialog will take its minimum margin values from the currently selected printer properties (Windows only).

Units are in millimetres.

setMarginBottomRight/2

Specs

setMarginBottomRight(This, Pt) -> ok
                        when
                            This :: wxPageSetupDialogData(),
                            Pt :: {X :: integer(), Y :: integer()}.

Sets the right (x) and bottom (y) margins in millimetres.

setMarginTopLeft/2

Specs

setMarginTopLeft(This, Pt) -> ok
                    when This :: wxPageSetupDialogData(), Pt :: {X :: integer(), Y :: integer()}.

Sets the left (x) and top (y) margins in millimetres.

setMinMarginBottomRight/2

Specs

setMinMarginBottomRight(This, Pt) -> ok
                           when
                               This :: wxPageSetupDialogData(),
                               Pt :: {X :: integer(), Y :: integer()}.

Sets the right (x) and bottom (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

setMinMarginTopLeft/2

Specs

setMinMarginTopLeft(This, Pt) -> ok
                       when
                           This :: wxPageSetupDialogData(), Pt :: {X :: integer(), Y :: integer()}.

Sets the left (x) and top (y) minimum margins the user can enter (Windows only).

Units are in millimetres.

setPaperId/2

Specs

setPaperId(This, Id) -> ok when This :: wxPageSetupDialogData(), Id :: wx:wx_enum().

Sets the paper size id.

Calling this function overrides the explicit paper dimensions passed in setPaperSize/2.

See: wxPrintData:setPaperId/2

setPaperSize/2

Specs

setPaperSize(This, Size) -> ok
                when This :: wxPageSetupDialogData(), Size :: {W :: integer(), H :: integer()}.

Sets the paper size in millimetres.

If a corresponding paper id is found, it will be set in the internal wxPrintData object, otherwise the paper size overrides the paper id.

setPrintData/2

Specs

setPrintData(This, PrintData) -> ok
                when This :: wxPageSetupDialogData(), PrintData :: wxPrintData:wxPrintData().

Sets the print data associated with this object.