wxPaintEvent (wx v2.0.1)

A paint event is sent when a window's contents needs to be repainted.

The handler of this event must create a wxPaintDC object and use it for painting the window contents. For example:

Notice that you must not create other kinds of wxDC (e.g. wxClientDC or wxWindowDC) in EVT_PAINT handlers and also don't create wxPaintDC outside of this event handlers.

You can optimize painting by retrieving the rectangles that have been damaged and only repainting these. The rectangles are in terms of the client area, and are unscrolled, so you will need to do some calculations using the current view position to obtain logical, scrolled units. Here is an example of using the wxRegionIterator (not implemented in wx) class:

Remark: Please notice that in general it is impossible to change the drawing of a standard control (such as wxButton) and so you shouldn't attempt to handle paint events for them as even if it might work on some platforms, this is inherently not portable and won't work everywhere.

See: Overview events

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxPaintEvent

Events

Use wxEvtHandler:connect/3 with wxPaintEventType to subscribe to events of this type.

Types

-type wxPaint() :: term().

Specs

wxPaint() :: #wxPaint{}.

-type wxPaintEvent() :: term().

Specs

wxPaintEvent() :: wx:wx_object().

-type wxPaintEventType() :: term().

Specs

wxPaintEventType() :: paint.