wxPen (wx v2.0.1)

A pen is a drawing tool for drawing outlines. It is used for drawing lines and painting the outline of rectangles, ellipses, etc. It has a colour, a width and a style.

Note: On a monochrome display, wxWidgets shows all non-white pens as black.

Do not initialize objects on the stack before the program commences, since other required structures may not have been set up yet. Instead, define global pointers to objects and create them in wxApp::OnInit() (not implemented in wx) or when required.

An application may wish to dynamically create pens with different characteristics, and there is the consequent danger that a large number of duplicate pens will be created. Therefore an application may wish to get a pointer to a pen by using the global list of pens ?wxThePenList, and calling the member function wxPenList::FindOrCreatePen() (not implemented in wx). See wxPenList (not implemented in wx) for more info.

This class uses reference counting and copy-on-write internally so that assignments between two instances of this class are very cheap. You can therefore use actual objects instead of pointers without efficiency problems. If an instance of this class is changed it will create its own data internally so that other instances, which previously shared the data using the reference counting, are not affected.

Predefined objects (include wx.hrl):

See: wxPenList (not implemented in wx), wxDC, wxDC:setPen/2

wxWidgets docs: wxPen

Types

-type wxPen() :: term().

Specs

wxPen() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxPen()) -> ok.

Destructor.

See: reference-counted object destruction

Remark: Although all remaining pens are deleted when the application exits, the application should try to clean up all pens itself. This is because wxWidgets cannot know if a pointer to the pen object is stored in an application data structure, and there is a risk of double deletion.

getCap/1

Specs

getCap(This) -> wx:wx_enum() when This :: wxPen().

Returns the pen cap style, which may be one of wxCAP_ROUND, wxCAP_PROJECTING and wxCAP_BUTT.

The default is wxCAP_ROUND.

See: setCap/2

getColour/1

Specs

getColour(This) -> wx:wx_colour4() when This :: wxPen().

Returns a reference to the pen colour.

See: setColour/4

getJoin/1

Specs

getJoin(This) -> wx:wx_enum() when This :: wxPen().

Returns the pen join style, which may be one of wxJOIN_BEVEL, wxJOIN_ROUND and wxJOIN_MITER.

The default is wxJOIN_ROUND.

See: setJoin/2

getStyle/1

Specs

getStyle(This) -> wx:wx_enum() when This :: wxPen().

Returns the pen style.

See: new/2, setStyle/2

getWidth/1

Specs

getWidth(This) -> integer() when This :: wxPen().

Returns the pen width.

See: setWidth/2

isOk/1

Specs

isOk(This) -> boolean() when This :: wxPen().

Returns true if the pen is initialised.

Notice that an uninitialized pen object can't be queried for any pen properties and all calls to the accessor methods on it will result in an assert failure.

new/0

Specs

new() -> wxPen().

Default constructor.

The pen will be uninitialised, and isOk/1 will return false.

new/1

Specs

new(Colour) -> wxPen() when Colour :: wx:wx_colour();
   (Pen) -> wxPen() when Pen :: wxPen().

Copy constructor, uses overview_refcount.

new/2

Specs

new(Colour, [Option]) -> wxPen()
       when Colour :: wx:wx_colour(), Option :: {width, integer()} | {style, wx:wx_enum()}.

Constructs a pen from a colour object, pen width and style.

Remark: Different versions of Windows and different versions of other platforms support very different subsets of the styles above so handle with care.

See: setStyle/2, setColour/4, setWidth/2

setCap/2

Specs

setCap(This, CapStyle) -> ok when This :: wxPen(), CapStyle :: wx:wx_enum().

Sets the pen cap style, which may be one of wxCAP_ROUND, wxCAP_PROJECTING and wxCAP_BUTT.

The default is wxCAP_ROUND.

See: getCap/1

setColour/2

Specs

setColour(This, Colour) -> ok when This :: wxPen(), Colour :: wx:wx_colour().

The pen's colour is changed to the given colour.

See: getColour/1

setColour/4

Specs

setColour(This, Red, Green, Blue) -> ok
             when This :: wxPen(), Red :: integer(), Green :: integer(), Blue :: integer().

setJoin/2

Specs

setJoin(This, Join_style) -> ok when This :: wxPen(), Join_style :: wx:wx_enum().

Sets the pen join style, which may be one of wxJOIN_BEVEL, wxJOIN_ROUND and wxJOIN_MITER.

The default is wxJOIN_ROUND.

See: getJoin/1

setStyle/2

Specs

setStyle(This, Style) -> ok when This :: wxPen(), Style :: wx:wx_enum().

Set the pen style.

See: new/2

setWidth/2

Specs

setWidth(This, Width) -> ok when This :: wxPen(), Width :: integer().

Sets the pen width.

See: getWidth/1