wxPickerBase (wx v2.0.1)

Base abstract class for all pickers which support an auxiliary text control.

This class handles all positioning and sizing of the text control like a an horizontal wxBoxSizer would do, with the text control on the left of the picker button.

The proportion (see wxSizer documentation for more info about proportion values) of the picker control defaults to 1 when there isn't a text control associated (see wxPB_USE_TEXTCTRL style) and to 0 otherwise.

Styles

This class supports the following styles:

See: wxColourPickerCtrl

This class is derived (and can use functions) from: wxControl wxWindow wxEvtHandler

wxWidgets docs: wxPickerBase

Types

-type wxPickerBase() :: term().

Specs

wxPickerBase() :: wx:wx_object().

Functions

getInternalMargin/1

Specs

getInternalMargin(This) -> integer() when This :: wxPickerBase().

Returns the margin (in pixel) between the picker and the text control.

This function can be used only when hasTextCtrl/1 returns true.

getPickerCtrlProportion/1

Specs

getPickerCtrlProportion(This) -> integer() when This :: wxPickerBase().

Returns the proportion value of the picker.

getTextCtrl/1

Specs

getTextCtrl(This) -> wxTextCtrl:wxTextCtrl() when This :: wxPickerBase().

Returns a pointer to the text control handled by this window or NULL if the wxPB_USE_TEXTCTRL style was not specified when this control was created.

Remark: The contents of the text control could be an invalid representation of the entity which can be chosen through the picker (e.g. when the user enters an invalid colour syntax because of a typo). Thus you should never parse the content of the textctrl to get the user's input; rather use the derived-class getter (e.g. wxColourPickerCtrl:getColour/1, wxFilePickerCtrl:getPath/1, etc).

getTextCtrlProportion/1

Specs

getTextCtrlProportion(This) -> integer() when This :: wxPickerBase().

Returns the proportion value of the text control.

This function can be used only when hasTextCtrl/1 returns true.

hasTextCtrl/1

Specs

hasTextCtrl(This) -> boolean() when This :: wxPickerBase().

Returns true if this window has a valid text control (i.e. if the wxPB_USE_TEXTCTRL style was given when creating this control).

isPickerCtrlGrowable/1

Specs

isPickerCtrlGrowable(This) -> boolean() when This :: wxPickerBase().

Returns true if the picker control is growable.

isTextCtrlGrowable/1

Specs

isTextCtrlGrowable(This) -> boolean() when This :: wxPickerBase().

Returns true if the text control is growable.

This function can be used only when hasTextCtrl/1 returns true.

setInternalMargin/2

Specs

setInternalMargin(This, Margin) -> ok when This :: wxPickerBase(), Margin :: integer().

Sets the margin (in pixel) between the picker and the text control.

This function can be used only when hasTextCtrl/1 returns true.

setPickerCtrlGrowable/1

Specs

setPickerCtrlGrowable(This) -> ok when This :: wxPickerBase().

setPickerCtrlGrowable/2

Specs

setPickerCtrlGrowable(This, [Option]) -> ok
                         when This :: wxPickerBase(), Option :: {grow, boolean()}.

Sets the picker control as growable when grow is true.

setPickerCtrlProportion/2

Specs

setPickerCtrlProportion(This, Prop) -> ok when This :: wxPickerBase(), Prop :: integer().

Sets the proportion value of the picker.

Look at the detailed description of wxPickerBase for more info.

setTextCtrlGrowable/1

Specs

setTextCtrlGrowable(This) -> ok when This :: wxPickerBase().

setTextCtrlGrowable/2

Specs

setTextCtrlGrowable(This, [Option]) -> ok when This :: wxPickerBase(), Option :: {grow, boolean()}.

Sets the text control as growable when grow is true.

This function can be used only when hasTextCtrl/1 returns true.

setTextCtrlProportion/2

Specs

setTextCtrlProportion(This, Prop) -> ok when This :: wxPickerBase(), Prop :: integer().

Sets the proportion value of the text control.

Look at the detailed description of wxPickerBase for more info.

This function can be used only when hasTextCtrl/1 returns true.