wxPopupTransientWindow (wx v2.0.1)

A wxPopupWindow which disappears automatically when the user clicks mouse outside it or if it loses focus in any other way.

This window can be useful for implementing custom combobox-like controls for example.

See: wxPopupWindow

This class is derived (and can use functions) from: wxPopupWindow wxWindow wxEvtHandler

wxWidgets docs: wxPopupTransientWindow

Types

-type wxPopupTransientWindow() :: term().

Specs

wxPopupTransientWindow() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxPopupTransientWindow()) -> ok.

Destroys the object.

dismiss/1

Specs

dismiss(This) -> ok when This :: wxPopupTransientWindow().

Hide the window.

new/0

Specs

new() -> wxPopupTransientWindow().

Default constructor.

new/1

Specs

new(Parent) -> wxPopupTransientWindow() when Parent :: wxWindow:wxWindow().

new/2

Specs

new(Parent, [Option]) -> wxPopupTransientWindow()
       when Parent :: wxWindow:wxWindow(), Option :: {style, integer()}.

Constructor.

popup/1

Specs

popup(This) -> ok when This :: wxPopupTransientWindow().

popup/2

Specs

popup(This, [Option]) -> ok
         when This :: wxPopupTransientWindow(), Option :: {focus, wxWindow:wxWindow()}.

Popup the window (this will show it too).

If focus is non-NULL, it will be kept focused while this window is shown if supported by the current platform, otherwise the popup itself will receive focus. In any case, the popup will disappear automatically if it loses focus because of a user action.

See: dismiss/1