wxPreviewFrame (wx v2.0.1)

This class provides the default method of managing the print preview interface. Member functions may be overridden to replace functionality, or the class may be used without derivation.

See: wxPreviewCanvas, wxPreviewControlBar, wxPrintPreview

This class is derived (and can use functions) from: wxFrame wxTopLevelWindow wxWindow wxEvtHandler

wxWidgets docs: wxPreviewFrame

Types

-type wxPreviewFrame() :: term().

Specs

wxPreviewFrame() :: wx:wx_object().

Functions

createCanvas/1

Specs

createCanvas(This) -> ok when This :: wxPreviewFrame().

Creates a wxPreviewCanvas.

Override this function to allow a user-defined preview canvas object to be created.

createControlBar/1

Specs

createControlBar(This) -> ok when This :: wxPreviewFrame().

Creates a wxPreviewControlBar.

Override this function to allow a user-defined preview control bar object to be created.

destroy/1

Specs

destroy(This :: wxPreviewFrame()) -> ok.

Destructor.

initialize/1

Specs

initialize(This) -> ok when This :: wxPreviewFrame().

Initializes the frame elements and prepares for showing it.

Calling this method is equivalent to calling InitializeWithModality() (not implemented in wx) with wxPreviewFrame_AppModal argument, please see its documentation for more details.

Please notice that this function is virtual mostly for backwards compatibility only, there is no real need to override it as it's never called by wxWidgets itself.

new/2

Specs

new(Preview, Parent) -> wxPreviewFrame()
       when Preview :: wxPrintPreview:wxPrintPreview(), Parent :: wxWindow:wxWindow().

new/3

Specs

new(Preview, Parent, [Option]) -> wxPreviewFrame()
       when
           Preview :: wxPrintPreview:wxPrintPreview(),
           Parent :: wxWindow:wxWindow(),
           Option ::
               {title, unicode:chardata()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Constructor.

Pass a print preview object plus other normal frame arguments. The print preview object will be destroyed by the frame when it closes.

onCloseWindow/2

Specs

onCloseWindow(This, Event) -> ok
                 when This :: wxPreviewFrame(), Event :: wxCloseEvent:wxCloseEvent().

Enables any disabled frames in the application, and deletes the print preview object, implicitly deleting any printout objects associated with the print preview object.