wxPrintData (wx v2.4.2)

Functions for wxPrintData class

This class holds a variety of information related to printers and printer device contexts. This class is used to create a wxPrinterDC (not implemented in wx) and a wxPostScriptDC. It is also used as a data member of wxPrintDialogData and wxPageSetupDialogData, as part of the mechanism for transferring data between the print dialogs and the application.

See: Overview printing, wxPrintDialog, wxPageSetupDialog, wxPrintDialogData, wxPageSetupDialogData, Overview cmndlg, wxPrinterDC (not implemented in wx), wxPostScriptDC

wxWidgets docs: wxPrintData

Types

wxPrintData()

-type wxPrintData() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxPrintData()) -> ok.

Destructor.

getBin(This)

-spec getBin(This) -> wx:wx_enum() when This :: wxPrintData().

Returns the current bin (papersource).

By default, the system is left to select the bin (wxPRINTBIN_DEFAULT is returned).

See setBin/2 for the full list of bin values.

getCollate(This)

-spec getCollate(This) -> boolean() when This :: wxPrintData().

Returns true if collation is on.

getColour(This)

-spec getColour(This) -> boolean() when This :: wxPrintData().

Returns true if colour printing is on.

getDuplex(This)

-spec getDuplex(This) -> wx:wx_enum() when This :: wxPrintData().

Returns the duplex mode.

One of wxDUPLEX_SIMPLEX, wxDUPLEX_HORIZONTAL, wxDUPLEX_VERTICAL.

getNoCopies(This)

-spec getNoCopies(This) -> integer() when This :: wxPrintData().

Returns the number of copies requested by the user.

getOrientation(This)

-spec getOrientation(This) -> wx:wx_enum() when This :: wxPrintData().

Gets the orientation.

This can be wxLANDSCAPE or wxPORTRAIT.

getPaperId(This)

-spec getPaperId(This) -> wx:wx_enum() when This :: wxPrintData().

Returns the paper size id.

See: setPaperId/2

getPrinterName(This)

-spec getPrinterName(This) -> unicode:charlist() when This :: wxPrintData().

Returns the printer name.

If the printer name is the empty string, it indicates that the default printer should be used.

getQuality(This)

-spec getQuality(This) -> integer() when This :: wxPrintData().

Returns the current print quality.

This can be a positive integer, denoting the number of dots per inch, or one of the following identifiers:

On input you should pass one of these identifiers, but on return you may get back a positive integer indicating the current resolution setting.

isOk(This)

-spec isOk(This) -> boolean() when This :: wxPrintData().

Returns true if the print data is valid for using in print dialogs.

This can return false on Windows if the current printer is not set, for example. On all other platforms, it returns true.

new()

-spec new() -> wxPrintData().

Default constructor.

new(Data)

-spec new(Data) -> wxPrintData() when Data :: wxPrintData().

Copy constructor.

setBin(This, Flag)

-spec setBin(This, Flag) -> ok when This :: wxPrintData(), Flag :: wx:wx_enum().

Sets the current bin.

setCollate(This, Flag)

-spec setCollate(This, Flag) -> ok when This :: wxPrintData(), Flag :: boolean().

Sets collation to on or off.

setColour(This, Flag)

-spec setColour(This, Flag) -> ok when This :: wxPrintData(), Flag :: boolean().

Sets colour printing on or off.

setDuplex(This, Mode)

-spec setDuplex(This, Mode) -> ok when This :: wxPrintData(), Mode :: wx:wx_enum().

Returns the duplex mode.

One of wxDUPLEX_SIMPLEX, wxDUPLEX_HORIZONTAL, wxDUPLEX_VERTICAL.

setNoCopies(This, N)

-spec setNoCopies(This, N) -> ok when This :: wxPrintData(), N :: integer().

Sets the default number of copies to be printed out.

setOrientation(This, Orientation)

-spec setOrientation(This, Orientation) -> ok when This :: wxPrintData(), Orientation :: wx:wx_enum().

Sets the orientation.

This can be wxLANDSCAPE or wxPORTRAIT.

setPaperId(This, PaperId)

-spec setPaperId(This, PaperId) -> ok when This :: wxPrintData(), PaperId :: wx:wx_enum().

Sets the paper id.

This indicates the type of paper to be used. For a mapping between paper id, paper size and string name, see wxPrintPaperDatabase in "paper.h" (not yet documented).

See: SetPaperSize() (not implemented in wx)

setPrinterName(This, PrinterName)

-spec setPrinterName(This, PrinterName) -> ok
                  when This :: wxPrintData(), PrinterName :: unicode:chardata().

Sets the printer name.

This can be the empty string to indicate that the default printer should be used.

setQuality(This, Quality)

-spec setQuality(This, Quality) -> ok when This :: wxPrintData(), Quality :: integer().

Sets the desired print quality.

This can be a positive integer, denoting the number of dots per inch, or one of the following identifiers:

On input you should pass one of these identifiers, but on return you may get back a positive integer indicating the current resolution setting.