wxPrintDialog (wx v2.4.3)

This class represents the print and print setup common dialogs.

You may obtain a wxPrinterDC (not implemented in wx) device context from a successfully dismissed print dialog.

See:

Overview printing

Overview cmndlg

This class is derived, and can use functions, from:

wxDialog

wxTopLevelWindow

wxWindow

wxEvtHandler

wxWidgets docs: wxPrintDialog

Types

wxPrintDialog()

-type wxPrintDialog() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxPrintDialog()) -> ok.

Destroys the object

getPrintDC(This)

-spec getPrintDC(This) -> wxDC:wxDC() when This :: wxPrintDialog().

Returns the device context created by the print dialog, if any.

When this function has been called, the ownership of the device context is transferred to the application, so it must then be deleted explicitly.

getPrintDialogData(This)

-spec getPrintDialogData(This) -> wxPrintDialogData:wxPrintDialogData() when This :: wxPrintDialog().

Returns the print dialog data associated with the print dialog.

new(Parent)

-spec new(Parent) -> wxPrintDialog() when Parent :: wxWindow:wxWindow().

Equivalent to new(Parent, []).

new/2

-spec new(Parent, [Option]) -> wxPrintDialog()
             when Parent :: wxWindow:wxWindow(), Option :: {data, wxPrintDialogData:wxPrintDialogData()};
         (Parent, Data) -> wxPrintDialog()
             when Parent :: wxWindow:wxWindow(), Data :: wxPrintData:wxPrintData().