wxPrintDialogData (wx v2.5)

This class holds information related to the visual characteristics of wxPrintDialog.

It contains a wxPrintData object with underlying printing settings.

See:

Overview printing

wxPrintDialog

Overview cmndlg

wxWidgets docs: wxPrintDialogData

Types

wxPrintDialogData()

-type wxPrintDialogData() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxPrintDialogData()) -> ok.

Destroys the object

enableHelp(This, Flag)

-spec enableHelp(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Help" button.

enablePageNumbers(This, Flag)

-spec enablePageNumbers(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Page numbers" controls.

enablePrintToFile(This, Flag)

-spec enablePrintToFile(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Print to file" checkbox.

enableSelection(This, Flag)

-spec enableSelection(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Selection" radio button.

getAllPages(This)

-spec getAllPages(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user requested that all pages be printed.

getCollate(This)

-spec getCollate(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user requested that the document(s) be collated.

getFromPage(This)

-spec getFromPage(This) -> integer() when This :: wxPrintDialogData().

Returns the from page number, as entered by the user.

getMaxPage(This)

-spec getMaxPage(This) -> integer() when This :: wxPrintDialogData().

Returns the maximum page number.

getMinPage(This)

-spec getMinPage(This) -> integer() when This :: wxPrintDialogData().

Returns the minimum page number.

getNoCopies(This)

-spec getNoCopies(This) -> integer() when This :: wxPrintDialogData().

Returns the number of copies requested by the user.

getPrintData(This)

-spec getPrintData(This) -> wxPrintData:wxPrintData() when This :: wxPrintDialogData().

Returns a reference to the internal wxPrintData object.

getPrintToFile(This)

-spec getPrintToFile(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user has selected printing to a file.

getSelection(This)

-spec getSelection(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user requested that the selection be printed (where "selection" is a concept specific to the application).

getToPage(This)

-spec getToPage(This) -> integer() when This :: wxPrintDialogData().

Returns the "print to" page number, as entered by the user.

isOk(This)

-spec isOk(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the print data is valid for using in print dialogs.

This can return false on Windows if the current printer is not set, for example. On all other platforms, it returns true.

new()

-spec new() -> wxPrintDialogData().

Default constructor.

new(DialogData)

-spec new(DialogData) -> wxPrintDialogData()
             when DialogData :: wxPrintDialogData:wxPrintDialogData() | wxPrintData:wxPrintData().

Copy constructor.

setCollate(This, Flag)

-spec setCollate(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Sets the "Collate" checkbox to true or false.

setFromPage(This, Page)

-spec setFromPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the from page number.

setMaxPage(This, Page)

-spec setMaxPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the maximum page number.

setMinPage(This, Page)

-spec setMinPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the minimum page number.

setNoCopies(This, N)

-spec setNoCopies(This, N) -> ok when This :: wxPrintDialogData(), N :: integer().

Sets the default number of copies the user has requested to be printed out.

setPrintData(This, PrintData)

-spec setPrintData(This, PrintData) -> ok
                      when This :: wxPrintDialogData(), PrintData :: wxPrintData:wxPrintData().

Sets the internal wxPrintData.

setPrintToFile(This, Flag)

-spec setPrintToFile(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Sets the "Print to file" checkbox to true or false.

setSelection(This, Flag)

-spec setSelection(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Selects the "Selection" radio button.

The effect of printing the selection depends on how the application implements this command, if at all.

setToPage(This, Page)

-spec setToPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the "print to" page number.