wxPrintDialogData (wx v2.0.1)

This class holds information related to the visual characteristics of wxPrintDialog. It contains a wxPrintData object with underlying printing settings.

See: Overview printing, wxPrintDialog, Overview cmndlg

wxWidgets docs: wxPrintDialogData

Types

-type wxPrintDialogData() :: term().

Specs

wxPrintDialogData() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxPrintDialogData()) -> ok.

Destructor.

enableHelp/2

Specs

enableHelp(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Help" button.

enablePageNumbers/2

Specs

enablePageNumbers(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Page numbers" controls.

enablePrintToFile/2

Specs

enablePrintToFile(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Print to file" checkbox.

enableSelection/2

Specs

enableSelection(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Enables or disables the "Selection" radio button.

getAllPages/1

Specs

getAllPages(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user requested that all pages be printed.

getCollate/1

Specs

getCollate(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user requested that the document(s) be collated.

getFromPage/1

Specs

getFromPage(This) -> integer() when This :: wxPrintDialogData().

Returns the from page number, as entered by the user.

getMaxPage/1

Specs

getMaxPage(This) -> integer() when This :: wxPrintDialogData().

Returns the maximum page number.

getMinPage/1

Specs

getMinPage(This) -> integer() when This :: wxPrintDialogData().

Returns the minimum page number.

getNoCopies/1

Specs

getNoCopies(This) -> integer() when This :: wxPrintDialogData().

Returns the number of copies requested by the user.

getPrintData/1

Specs

getPrintData(This) -> wxPrintData:wxPrintData() when This :: wxPrintDialogData().

Returns a reference to the internal wxPrintData object.

getPrintToFile/1

Specs

getPrintToFile(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user has selected printing to a file.

getSelection/1

Specs

getSelection(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the user requested that the selection be printed (where "selection" is a concept specific to the application).

getToPage/1

Specs

getToPage(This) -> integer() when This :: wxPrintDialogData().

Returns the "print to" page number, as entered by the user.

isOk/1

Specs

isOk(This) -> boolean() when This :: wxPrintDialogData().

Returns true if the print data is valid for using in print dialogs.

This can return false on Windows if the current printer is not set, for example. On all other platforms, it returns true.

new/0

Specs

new() -> wxPrintDialogData().

Default constructor.

new/1

Specs

new(DialogData) -> wxPrintDialogData()
       when DialogData :: wxPrintDialogData:wxPrintDialogData() | wxPrintData:wxPrintData().

Copy constructor.

setCollate/2

Specs

setCollate(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Sets the "Collate" checkbox to true or false.

setFromPage/2

Specs

setFromPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the from page number.

setMaxPage/2

Specs

setMaxPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the maximum page number.

setMinPage/2

Specs

setMinPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the minimum page number.

setNoCopies/2

Specs

setNoCopies(This, N) -> ok when This :: wxPrintDialogData(), N :: integer().

Sets the default number of copies the user has requested to be printed out.

setPrintData/2

Specs

setPrintData(This, PrintData) -> ok
                when This :: wxPrintDialogData(), PrintData :: wxPrintData:wxPrintData().

Sets the internal wxPrintData.

setPrintToFile/2

Specs

setPrintToFile(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Sets the "Print to file" checkbox to true or false.

setSelection/2

Specs

setSelection(This, Flag) -> ok when This :: wxPrintDialogData(), Flag :: boolean().

Selects the "Selection" radio button.

The effect of printing the selection depends on how the application implements this command, if at all.

setToPage/2

Specs

setToPage(This, Page) -> ok when This :: wxPrintDialogData(), Page :: integer().

Sets the "print to" page number.