wxRadioBox (wx v2.0.1)

A radio box item is used to select one of number of mutually exclusive choices. It is displayed as a vertical column or horizontal row of labelled buttons.

Styles

This class supports the following styles:

See: Overview events, wxRadioButton, wxCheckBox

This class is derived (and can use functions) from: wxControl wxWindow wxEvtHandler

wxWidgets docs: wxRadioBox

Events

Event types emitted from this class: command_radiobox_selected

Types

-type wxRadioBox() :: term().

Specs

wxRadioBox() :: wx:wx_object().

Functions

create/7

Specs

create(This, Parent, Id, Label, Pos, Size, Choices) -> boolean()
          when
              This :: wxRadioBox(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: unicode:chardata(),
              Pos :: {X :: integer(), Y :: integer()},
              Size :: {W :: integer(), H :: integer()},
              Choices :: [unicode:chardata()].

create/8

Specs

create(This, Parent, Id, Label, Pos, Size, Choices, [Option]) -> boolean()
          when
              This :: wxRadioBox(),
              Parent :: wxWindow:wxWindow(),
              Id :: integer(),
              Label :: unicode:chardata(),
              Pos :: {X :: integer(), Y :: integer()},
              Size :: {W :: integer(), H :: integer()},
              Choices :: [unicode:chardata()],
              Option :: {majorDim, integer()} | {style, integer()} | {val, wx:wx_object()}.

Creates the radiobox for two-step construction.

See new/7 for further details.

destroy/1

Specs

destroy(This :: wxRadioBox()) -> ok.

Destructor, destroying the radiobox item.

enable/1

Specs

enable(This) -> boolean() when This :: wxRadioBox().

enable/2

Specs

enable(This, N) -> boolean() when This :: wxRadioBox(), N :: integer();
      (This, [Option]) -> boolean() when This :: wxRadioBox(), Option :: {enable, boolean()}.

Enables or disables the radiobox.

See: wxWindow:enable/2

enable/3

Specs

enable(This, N, [Option]) -> boolean()
          when This :: wxRadioBox(), N :: integer(), Option :: {enable, boolean()}.

Enables or disables an individual button in the radiobox.

See: wxWindow:enable/2

getColumnCount/1

Specs

getColumnCount(This) -> integer() when This :: wxRadioBox().

Returns the number of columns in the radiobox.

getItemFromPoint/2

Specs

getItemFromPoint(This, Pt) -> integer()
                    when This :: wxRadioBox(), Pt :: {X :: integer(), Y :: integer()}.

Returns a radio box item under the point, a zero-based item index, or wxNOT_FOUND if no item is under the point.

getItemHelpText/2

Specs

getItemHelpText(This, Item) -> unicode:charlist() when This :: wxRadioBox(), Item :: integer().

Returns the helptext associated with the specified item if any or wxEmptyString.

See: setItemHelpText/3

getItemToolTip/2

Specs

getItemToolTip(This, Item) -> wxToolTip:wxToolTip() when This :: wxRadioBox(), Item :: integer().

Returns the tooltip associated with the specified item if any or NULL.

See: setItemToolTip/3, wxWindow:getToolTip/1

getRowCount/1

Specs

getRowCount(This) -> integer() when This :: wxRadioBox().

Returns the number of rows in the radiobox.

getSelection/1

Specs

getSelection(This) -> integer() when This :: wxRadioBox().

Returns the index of the selected item or wxNOT_FOUND if no item is selected.

Return: The position of the current selection.

Remark: This method can be used with single selection list boxes only, you should use wxListBox:getSelections/1 for the list boxes with wxLB_MULTIPLE style.

See: setSelection/2, wxControlWithItems:getStringSelection/1

getString/2

Specs

getString(This, N) -> unicode:charlist() when This :: wxRadioBox(), N :: integer().

Returns the label of the item with the given index.

Return: The label of the item or an empty string if the position was invalid.

isItemEnabled/2

Specs

isItemEnabled(This, N) -> boolean() when This :: wxRadioBox(), N :: integer().

Returns true if the item is enabled or false if it was disabled using enable/3.

This function is currently only implemented in wxMSW, wxGTK, wxQT and wxUniversal and always returns true in the other ports.

isItemShown/2

Specs

isItemShown(This, N) -> boolean() when This :: wxRadioBox(), N :: integer().

Returns true if the item is currently shown or false if it was hidden using show/3.

Note that this function returns true for an item which hadn't been hidden even if the entire radiobox is not currently shown.

This function is currently only implemented in wxMSW, wxGTK, wxQT and wxUniversal and always returns true in the other ports.

new/6

Specs

new(Parent, Id, Label, Pos, Size, Choices) -> wxRadioBox()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Label :: unicode:chardata(),
           Pos :: {X :: integer(), Y :: integer()},
           Size :: {W :: integer(), H :: integer()},
           Choices :: [unicode:chardata()].

new/7

Specs

new(Parent, Id, Label, Pos, Size, Choices, [Option]) -> wxRadioBox()
       when
           Parent :: wxWindow:wxWindow(),
           Id :: integer(),
           Label :: unicode:chardata(),
           Pos :: {X :: integer(), Y :: integer()},
           Size :: {W :: integer(), H :: integer()},
           Choices :: [unicode:chardata()],
           Option :: {majorDim, integer()} | {style, integer()} | {val, wx:wx_object()}.

Constructor, creating and showing a radiobox.

See: create/8, wxValidator (not implemented in wx)

setItemHelpText/3

Specs

setItemHelpText(This, Item, Helptext) -> ok
                   when This :: wxRadioBox(), Item :: integer(), Helptext :: unicode:chardata().

Sets the helptext for an item.

Empty string erases any existing helptext.

See: getItemHelpText/2

setItemToolTip/3

Specs

setItemToolTip(This, Item, Text) -> ok
                  when This :: wxRadioBox(), Item :: integer(), Text :: unicode:chardata().

Sets the tooltip text for the specified item in the radio group.

This function is currently only implemented in wxMSW and wxGTK2 and does nothing in the other ports.

See: getItemToolTip/2, wxWindow:setToolTip/2

setSelection/2

Specs

setSelection(This, N) -> ok when This :: wxRadioBox(), N :: integer().

Sets the selection to the given item.

Notice that a radio box always has selection, so n must be valid here and passing wxNOT_FOUND is not allowed.

show/2

Specs

show(This, Item) -> boolean() when This :: wxRadioBox(), Item :: integer().

show/3

Specs

show(This, Item, [Option]) -> boolean()
        when This :: wxRadioBox(), Item :: integer(), Option :: {show, boolean()}.

Shows or hides individual buttons.

Return: true if the item has been shown or hidden or false if nothing was done because it already was in the requested state.

See: show/3