wxRegion (wx v2.0.1)

A wxRegion represents a simple or complex region on a device context or window.

This class uses reference counting and copy-on-write internally so that assignments between two instances of this class are very cheap. You can therefore use actual objects instead of pointers without efficiency problems. If an instance of this class is changed it will create its own data internally so that other instances, which previously shared the data using the reference counting, are not affected.

Predefined objects (include wx.hrl):

See: wxRegionIterator (not implemented in wx)

wxWidgets docs: wxRegion

Types

-type wxRegion() :: term().

Specs

wxRegion() :: wx:wx_object().

Functions

'Xor'/2

Specs

'Xor'(This, Rect) -> boolean()
         when
             This :: wxRegion(),
             Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()};
     (This, Region) -> boolean() when This :: wxRegion(), Region :: wxRegion().

Finds the Xor of this region and another region.

This method can be used even if this region is invalid and has the natural behaviour in this case, i.e. makes this region equal to the given region.

Return: true if successful, false otherwise.

Remark: This operation creates a region that combines all of this region and the second region, except for any overlapping areas. The result is stored in this region.

'Xor'/5

Specs

'Xor'(This, X, Y, Width, Height) -> boolean()
         when
             This :: wxRegion(),
             X :: integer(),
             Y :: integer(),
             Width :: integer(),
             Height :: integer().

Finds the Xor of this region and another, rectangular region, specified using position and size.

This method can be used even if this region is invalid and has the natural behaviour in this case, i.e. makes this region equal to the given rectangle.

Return: true if successful, false otherwise.

Remark: This operation creates a region that combines all of this region and the second region, except for any overlapping areas. The result is stored in this region.

clear/1

Specs

clear(This) -> ok when This :: wxRegion().

Clears the current region.

The object becomes invalid, or null, after being cleared.

contains/2

Specs

contains(This, Pt) -> wx:wx_enum() when This :: wxRegion(), Pt :: {X :: integer(), Y :: integer()};
        (This, Rect) -> wx:wx_enum()
            when
                This :: wxRegion(),
                Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()}.

Returns a value indicating whether the given rectangle is contained within the region.

This method always returns wxOutRegion for an invalid region but may, nevertheless, be safely called in this case.

Return: One of ?wxOutRegion, ?wxPartRegion or ?wxInRegion.

Note: On Windows, only ?wxOutRegion and ?wxInRegion are returned; a value ?wxInRegion then indicates that all or some part of the region is contained in this region.

contains/3

Specs

contains(This, X, Y) -> wx:wx_enum() when This :: wxRegion(), X :: integer(), Y :: integer().

Returns a value indicating whether the given point is contained within the region.

This method always returns wxOutRegion for an invalid region but may, nevertheless, be safely called in this case.

Return: The return value is one of wxOutRegion and wxInRegion.

contains/5

Specs

contains(This, X, Y, Width, Height) -> wx:wx_enum()
            when
                This :: wxRegion(),
                X :: integer(),
                Y :: integer(),
                Width :: integer(),
                Height :: integer().

Returns a value indicating whether the given rectangle is contained within the region.

This method always returns wxOutRegion for an invalid region but may, nevertheless, be safely called in this case.

Return: One of ?wxOutRegion, ?wxPartRegion or ?wxInRegion.

Note: On Windows, only ?wxOutRegion and ?wxInRegion are returned; a value ?wxInRegion then indicates that all or some part of the region is contained in this region.

convertToBitmap/1

Specs

convertToBitmap(This) -> wxBitmap:wxBitmap() when This :: wxRegion().

Convert the region to a black and white bitmap with the white pixels being inside the region.

This method can't be used for invalid region.

destroy/1

Specs

destroy(This :: wxRegion()) -> ok.

Destructor.

See reference-counted object destruction for more info.

getBox/1

Specs

getBox(This) -> {X :: integer(), Y :: integer(), W :: integer(), H :: integer()}
          when This :: wxRegion().

intersect/2

Specs

intersect(This, Rect) -> boolean()
             when
                 This :: wxRegion(),
                 Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()};
         (This, Region) -> boolean() when This :: wxRegion(), Region :: wxRegion().

Finds the intersection of this region and another region.

This method always fails, i.e. returns false, if this region is invalid but may nevertheless be safely used even in this case.

Return: true if successful, false otherwise.

Remark: Creates the intersection of the two regions, that is, the parts which are in both regions. The result is stored in this region.

intersect/5

Specs

intersect(This, X, Y, Width, Height) -> boolean()
             when
                 This :: wxRegion(),
                 X :: integer(),
                 Y :: integer(),
                 Width :: integer(),
                 Height :: integer().

Finds the intersection of this region and another, rectangular region, specified using position and size.

This method always fails, i.e. returns false, if this region is invalid but may nevertheless be safely used even in this case.

Return: true if successful, false otherwise.

Remark: Creates the intersection of the two regions, that is, the parts which are in both regions. The result is stored in this region.

isEmpty/1

Specs

isEmpty(This) -> boolean() when This :: wxRegion().

Returns true if the region is empty, false otherwise.

Always returns true if the region is invalid.

new/0

Specs

new() -> wxRegion().

Default constructor.

This constructor creates an invalid, or null, object, i.e. calling IsOk() on it returns false and isEmpty/1 returns true.

new/1

Specs

new(Rect) -> wxRegion()
       when Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()};
   (Bmp) -> wxRegion() when Bmp :: wxBitmap:wxBitmap().

Constructs a region using a bitmap.

See union/5 for more details.

new/2

Specs

new(TopLeft, BottomRight) -> wxRegion()
       when
           TopLeft :: {X :: integer(), Y :: integer()},
           BottomRight :: {X :: integer(), Y :: integer()}.

Constructs a rectangular region from the top left point and the bottom right point.

new/4

Specs

new(X, Y, Width, Height) -> wxRegion()
       when X :: integer(), Y :: integer(), Width :: integer(), Height :: integer().

Constructs a rectangular region with the given position and size.

offset/2

Specs

offset(This, Pt) -> boolean() when This :: wxRegion(), Pt :: {X :: integer(), Y :: integer()}.

offset/3

Specs

offset(This, X, Y) -> boolean() when This :: wxRegion(), X :: integer(), Y :: integer().

Moves the region by the specified offsets in horizontal and vertical directions.

This method can't be called if the region is invalid as it doesn't make sense to offset it then. Attempts to do it will result in assert failure.

Return: true if successful, false otherwise (the region is unchanged then).

subtract/2

Specs

subtract(This, Rect) -> boolean()
            when
                This :: wxRegion(),
                Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()};
        (This, Region) -> boolean() when This :: wxRegion(), Region :: wxRegion().

Subtracts a region from this region.

This method always fails, i.e. returns false, if this region is invalid but may nevertheless be safely used even in this case.

Return: true if successful, false otherwise.

Remark: This operation combines the parts of 'this' region that are not part of the second region. The result is stored in this region.

union/2

Specs

union(This, Region) -> boolean()
         when This :: wxRegion(), Region :: wxRegion:wxRegion() | wxBitmap:wxBitmap();
     (This, Rect) -> boolean()
         when
             This :: wxRegion(),
             Rect :: {X :: integer(), Y :: integer(), W :: integer(), H :: integer()}.

Finds the union of this region and another, rectangular region.

This method can be used even if this region is invalid and has the natural behaviour in this case, i.e. makes this region equal to the given rectangle.

Return: true if successful, false otherwise.

Remark: This operation creates a region that combines all of this region and the second region. The result is stored in this region.

union/3

Specs

union(This, Bmp, TransColour) -> boolean()
         when This :: wxRegion(), Bmp :: wxBitmap:wxBitmap(), TransColour :: wx:wx_colour().

union/4

Specs

union(This, Bmp, TransColour, [Option]) -> boolean()
         when
             This :: wxRegion(),
             Bmp :: wxBitmap:wxBitmap(),
             TransColour :: wx:wx_colour(),
             Option :: {tolerance, integer()}.

Finds the union of this region and the non-transparent pixels of a bitmap.

Colour to be treated as transparent is specified in the transColour argument, along with an optional colour tolerance value.

Return: true if successful, false otherwise.

Remark: This operation creates a region that combines all of this region and the second region. The result is stored in this region.

union/5

Specs

union(This, X, Y, Width, Height) -> boolean()
         when
             This :: wxRegion(),
             X :: integer(),
             Y :: integer(),
             Width :: integer(),
             Height :: integer().

Finds the union of this region and another, rectangular region, specified using position and size.

This method can be used even if this region is invalid and has the natural behaviour in this case, i.e. makes this region equal to the given rectangle.

Return: true if successful, false otherwise.

Remark: This operation creates a region that combines all of this region and the second region. The result is stored in this region.