wxSashWindow (wx v2.4.2)

Functions for wxSashWindow class

wxSashWindow allows any of its edges to have a sash which can be dragged to resize the window. The actual content window will be created by the application as a child of wxSashWindow.

The window (or an ancestor) will be notified of a drag via a wxSashEvent notification.

Styles

This class supports the following styles:

See: wxSashEvent, wxSashLayoutWindow, Overview events

This class is derived (and can use functions) from: wxWindow wxEvtHandler

wxWidgets docs: wxSashWindow

Events

Event types emitted from this class: sash_dragged

Types

wxSashWindow()

-type wxSashWindow() :: wx:wx_object().

Functions

destroy(This)

-spec destroy(This :: wxSashWindow()) -> ok.

Destructor.

getMaximumSizeX(This)

-spec getMaximumSizeX(This) -> integer() when This :: wxSashWindow().

Gets the maximum window size in the x direction.

getMaximumSizeY(This)

-spec getMaximumSizeY(This) -> integer() when This :: wxSashWindow().

Gets the maximum window size in the y direction.

getMinimumSizeX(This)

-spec getMinimumSizeX(This) -> integer() when This :: wxSashWindow().

Gets the minimum window size in the x direction.

getMinimumSizeY(This)

-spec getMinimumSizeY(This) -> integer() when This :: wxSashWindow().

Gets the minimum window size in the y direction.

getSashVisible(This, Edge)

-spec getSashVisible(This, Edge) -> boolean() when This :: wxSashWindow(), Edge :: wx:wx_enum().

Returns true if a sash is visible on the given edge, false otherwise.

See: setSashVisible/3

new()

-spec new() -> wxSashWindow().

Default ctor.

new(Parent)

-spec new(Parent) -> wxSashWindow() when Parent :: wxWindow:wxWindow().

new/2

-spec new(Parent, [Option]) -> wxSashWindow()
       when
           Parent :: wxWindow:wxWindow(),
           Option ::
               {id, integer()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Constructs a sash window, which can be a child of a frame, dialog or any other non-control window.

setMaximumSizeX(This, Min)

-spec setMaximumSizeX(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the maximum window size in the x direction.

setMaximumSizeY(This, Min)

-spec setMaximumSizeY(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the maximum window size in the y direction.

setMinimumSizeX(This, Min)

-spec setMinimumSizeX(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the minimum window size in the x direction.

setMinimumSizeY(This, Min)

-spec setMinimumSizeY(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the minimum window size in the y direction.

setSashVisible(This, Edge, Visible)

-spec setSashVisible(This, Edge, Visible) -> ok
                  when This :: wxSashWindow(), Edge :: wx:wx_enum(), Visible :: boolean().

Call this function to make a sash visible or invisible on a particular edge.

See: getSashVisible/2