wxSashWindow (wx v2.0.1)

wxSashWindow allows any of its edges to have a sash which can be dragged to resize the window. The actual content window will be created by the application as a child of wxSashWindow.

The window (or an ancestor) will be notified of a drag via a wxSashEvent notification.

Styles

This class supports the following styles:

See: wxSashEvent, wxSashLayoutWindow, Overview events

This class is derived (and can use functions) from: wxWindow wxEvtHandler

wxWidgets docs: wxSashWindow

Events

Event types emitted from this class: sash_dragged

Types

-type wxSashWindow() :: term().

Specs

wxSashWindow() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxSashWindow()) -> ok.

Destructor.

getMaximumSizeX/1

Specs

getMaximumSizeX(This) -> integer() when This :: wxSashWindow().

Gets the maximum window size in the x direction.

getMaximumSizeY/1

Specs

getMaximumSizeY(This) -> integer() when This :: wxSashWindow().

Gets the maximum window size in the y direction.

getMinimumSizeX/1

Specs

getMinimumSizeX(This) -> integer() when This :: wxSashWindow().

Gets the minimum window size in the x direction.

getMinimumSizeY/1

Specs

getMinimumSizeY(This) -> integer() when This :: wxSashWindow().

Gets the minimum window size in the y direction.

getSashVisible/2

Specs

getSashVisible(This, Edge) -> boolean() when This :: wxSashWindow(), Edge :: wx:wx_enum().

Returns true if a sash is visible on the given edge, false otherwise.

See: setSashVisible/3

new/0

Specs

new() -> wxSashWindow().

Default ctor.

new/1

Specs

new(Parent) -> wxSashWindow() when Parent :: wxWindow:wxWindow().

new/2

Specs

new(Parent, [Option]) -> wxSashWindow()
       when
           Parent :: wxWindow:wxWindow(),
           Option ::
               {id, integer()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()}.

Constructs a sash window, which can be a child of a frame, dialog or any other non-control window.

setMaximumSizeX/2

Specs

setMaximumSizeX(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the maximum window size in the x direction.

setMaximumSizeY/2

Specs

setMaximumSizeY(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the maximum window size in the y direction.

setMinimumSizeX/2

Specs

setMinimumSizeX(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the minimum window size in the x direction.

setMinimumSizeY/2

Specs

setMinimumSizeY(This, Min) -> ok when This :: wxSashWindow(), Min :: integer().

Sets the minimum window size in the y direction.

setSashVisible/3

Specs

setSashVisible(This, Edge, Visible) -> ok
                  when This :: wxSashWindow(), Edge :: wx:wx_enum(), Visible :: boolean().

Call this function to make a sash visible or invisible on a particular edge.

See: getSashVisible/2