wxShowEvent (wx v2.4.2)

Functions for wxShowEvent class

An event being sent when the window is shown or hidden. The event is triggered by calls to wxWindow:show/2, and any user action showing a previously hidden window or vice versa (if allowed by the current platform and/or window manager). Notice that the event is not triggered when the application is iconized (minimized) or restored under wxMSW.

See: Overview events, wxWindow:show/2, wxWindow:isShown/1

This class is derived (and can use functions) from: wxEvent

wxWidgets docs: wxShowEvent

Events

Use wxEvtHandler:connect/3 with wxShowEventType to subscribe to events of this type.

Types

wxShow()

-type wxShow() :: #wxShow{type :: wxShowEvent:wxShowEventType(), show :: boolean()}.

wxShowEvent()

-type wxShowEvent() :: wx:wx_object().

wxShowEventType()

-type wxShowEventType() :: show.

Functions

isShown(This)

-spec isShown(This) -> boolean() when This :: wxShowEvent().

Return true if the window has been shown, false if it has been hidden.

setShow(This, Show)

-spec setShow(This, Show) -> ok when This :: wxShowEvent(), Show :: boolean().

Set whether the windows was shown or hidden.