wxSingleChoiceDialog (wx v2.0.1)

This class represents a dialog that shows a list of strings, and allows the user to select one. Double-clicking on a list item is equivalent to single-clicking and then pressing OK.

Styles

This class supports the following styles:

See: Overview cmndlg, wxMultiChoiceDialog

This class is derived (and can use functions) from: wxDialog wxTopLevelWindow wxWindow wxEvtHandler

wxWidgets docs: wxSingleChoiceDialog

Types

-type wxSingleChoiceDialog() :: term().

Specs

wxSingleChoiceDialog() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxSingleChoiceDialog()) -> ok.

Destroys the object.

getSelection/1

Specs

getSelection(This) -> integer() when This :: wxSingleChoiceDialog().

Returns the index of selected item.

getStringSelection/1

Specs

getStringSelection(This) -> unicode:charlist() when This :: wxSingleChoiceDialog().

Returns the selected string.

new/4

Specs

new(Parent, Message, Caption, Choices) -> wxSingleChoiceDialog()
       when
           Parent :: wxWindow:wxWindow(),
           Message :: unicode:chardata(),
           Caption :: unicode:chardata(),
           Choices :: [unicode:chardata()].

new/5

Specs

new(Parent, Message, Caption, Choices, [Option]) -> wxSingleChoiceDialog()
       when
           Parent :: wxWindow:wxWindow(),
           Message :: unicode:chardata(),
           Caption :: unicode:chardata(),
           Choices :: [unicode:chardata()],
           Option :: {style, integer()} | {pos, {X :: integer(), Y :: integer()}}.

Constructor, taking an array of wxString (not implemented in wx) choices and optional client data.

Remark: Use wxDialog:showModal/1 to show the dialog.

setSelection/2

Specs

setSelection(This, Selection) -> ok when This :: wxSingleChoiceDialog(), Selection :: integer().

Sets the index of the initially selected item.