wxSpinCtrl (wx v2.0.1)

wxSpinCtrl combines wxTextCtrl and wxSpinButton in one control.

Styles

This class supports the following styles:

See: wxSpinButton, wxSpinCtrlDouble (not implemented in wx), wxControl

This class is derived (and can use functions) from: wxControl wxWindow wxEvtHandler

wxWidgets docs: wxSpinCtrl

Events

Event types emitted from this class: command_spinctrl_updated

Types

-type wxSpinCtrl() :: term().

Specs

wxSpinCtrl() :: wx:wx_object().

Functions

create/2

Specs

create(This, Parent) -> boolean() when This :: wxSpinCtrl(), Parent :: wxWindow:wxWindow().

create/3

Specs

create(This, Parent, [Option]) -> boolean()
          when
              This :: wxSpinCtrl(),
              Parent :: wxWindow:wxWindow(),
              Option ::
                  {id, integer()} |
                  {value, unicode:chardata()} |
                  {pos, {X :: integer(), Y :: integer()}} |
                  {size, {W :: integer(), H :: integer()}} |
                  {style, integer()} |
                  {min, integer()} |
                  {max, integer()} |
                  {initial, integer()}.

Creation function called by the spin control constructor.

See new/2 for details.

destroy/1

Specs

destroy(This :: wxSpinCtrl()) -> ok.

Destroys the object.

getMax/1

Specs

getMax(This) -> integer() when This :: wxSpinCtrl().

Gets maximal allowable value.

getMin/1

Specs

getMin(This) -> integer() when This :: wxSpinCtrl().

Gets minimal allowable value.

getValue/1

Specs

getValue(This) -> integer() when This :: wxSpinCtrl().

Gets the value of the spin control.

new/0

Specs

new() -> wxSpinCtrl().

Default constructor.

new/1

Specs

new(Parent) -> wxSpinCtrl() when Parent :: wxWindow:wxWindow().

new/2

Specs

new(Parent, [Option]) -> wxSpinCtrl()
       when
           Parent :: wxWindow:wxWindow(),
           Option ::
               {id, integer()} |
               {value, unicode:chardata()} |
               {pos, {X :: integer(), Y :: integer()}} |
               {size, {W :: integer(), H :: integer()}} |
               {style, integer()} |
               {min, integer()} |
               {max, integer()} |
               {initial, integer()}.

Constructor, creating and showing a spin control.

If value is non-empty, it will be shown in the text entry part of the control and if it has numeric value, the initial numeric value of the control, as returned by getValue/1 will also be determined by it instead of by initial. Hence, it only makes sense to specify initial if value is an empty string or is not convertible to a number, otherwise initial is simply ignored and the number specified by value is used.

See: create/3

setRange/3

Specs

setRange(This, MinVal, MaxVal) -> ok
            when This :: wxSpinCtrl(), MinVal :: integer(), MaxVal :: integer().

Sets range of allowable values.

Notice that calling this method may change the value of the control if it's not inside the new valid range, e.g. it will become minVal if it is less than it now. However no wxEVT_SPINCTRL event is generated, even if it the value does change.

Note: Setting a range including negative values is silently ignored if current base is set to 16.

setSelection/3

Specs

setSelection(This, From, To) -> ok when This :: wxSpinCtrl(), From :: integer(), To :: integer().

Select the text in the text part of the control between positions from (inclusive) and to (exclusive).

This is similar to wxTextCtrl:setSelection/3.

Note: this is currently only implemented for Windows and generic versions of the control.

setValue/2

Specs

setValue(This, Value) -> ok when This :: wxSpinCtrl(), Value :: integer();
        (This, Text) -> ok when This :: wxSpinCtrl(), Text :: unicode:chardata().

Sets the value of the spin control.

It is recommended to use the overload taking an integer value instead.

Notice that, unlike wxTextCtrl:setValue/2, but like most of the other setter methods in wxWidgets, calling this method does not generate any events as events are only generated for the user actions.