wxSplitterEvent (wx v2.0.1)

This class represents the events generated by a splitter control.

Also there is only one event class, the data associated to the different events is not the same and so not all accessor functions may be called for each event. The documentation mentions the kind of event(s) for which the given accessor function makes sense: calling it for other types of events will result in assert failure (in debug mode) and will return meaningless results.

See: wxSplitterWindow, Overview events

This class is derived (and can use functions) from: wxNotifyEvent wxCommandEvent wxEvent

wxWidgets docs: wxSplitterEvent

Events

Use wxEvtHandler:connect/3 with wxSplitterEventType to subscribe to events of this type.

Types

-type wxSplitter() :: term().

Specs

wxSplitter() :: #wxSplitter{}.

-type wxSplitterEvent() :: term().

Specs

wxSplitterEvent() :: wx:wx_object().

-type wxSplitterEventType() :: term().

Specs

wxSplitterEventType() ::
    command_splitter_sash_pos_changed | command_splitter_sash_pos_changing |
    command_splitter_doubleclicked | command_splitter_unsplit.

Functions

getSashPosition/1

Specs

getSashPosition(This) -> integer() when This :: wxSplitterEvent().

Returns the new sash position.

May only be called while processing wxEVT_SPLITTER_SASH_POS_CHANGING and wxEVT_SPLITTER_SASH_POS_CHANGED events.

getWindowBeingRemoved/1

Specs

getWindowBeingRemoved(This) -> wxWindow:wxWindow() when This :: wxSplitterEvent().

Returns a pointer to the window being removed when a splitter window is unsplit.

May only be called while processing wxEVT_SPLITTER_UNSPLIT events.

getX/1

Specs

getX(This) -> integer() when This :: wxSplitterEvent().

Returns the x coordinate of the double-click point.

May only be called while processing wxEVT_SPLITTER_DOUBLECLICKED events.

getY/1

Specs

getY(This) -> integer() when This :: wxSplitterEvent().

Returns the y coordinate of the double-click point.

May only be called while processing wxEVT_SPLITTER_DOUBLECLICKED events.

setSashPosition/2

Specs

setSashPosition(This, Pos) -> ok when This :: wxSplitterEvent(), Pos :: integer().

In the case of wxEVT_SPLITTER_SASH_POS_CHANGED events, sets the new sash position.

In the case of wxEVT_SPLITTER_SASH_POS_CHANGING events, sets the new tracking bar position so visual feedback during dragging will represent that change that will actually take place. Set to -1 from the event handler code to prevent repositioning.

May only be called while processing wxEVT_SPLITTER_SASH_POS_CHANGING and wxEVT_SPLITTER_SASH_POS_CHANGED events.