wxStdDialogButtonSizer (wx v2.0.1)

This class creates button layouts which conform to the standard button spacing and ordering defined by the platform or toolkit's user interface guidelines (if such things exist). By using this class, you can ensure that all your standard dialogs look correct on all major platforms. Currently it conforms to the Windows, GTK+ and macOS human interface guidelines.

When there aren't interface guidelines defined for a particular platform or toolkit, wxStdDialogButtonSizer reverts to the Windows implementation.

To use this class, first add buttons to the sizer by calling addButton/2 (or setAffirmativeButton/2, setNegativeButton/2 or setCancelButton/2) and then call Realize in order to create the actual button layout used. Other than these special operations, this sizer works like any other sizer.

If you add a button with wxID_SAVE, on macOS the button will be renamed to "Save" and the wxID_NO button will be renamed to "Don't Save" in accordance with the macOS Human Interface Guidelines.

See: wxSizer, Overview sizer, wxDialog:createButtonSizer/2

This class is derived (and can use functions) from: wxBoxSizer wxSizer

wxWidgets docs: wxStdDialogButtonSizer

Types

-type wxStdDialogButtonSizer() :: term().

Specs

wxStdDialogButtonSizer() :: wx:wx_object().

Functions

addButton/2

Specs

addButton(This, Button) -> ok when This :: wxStdDialogButtonSizer(), Button :: wxButton:wxButton().

Adds a button to the wxStdDialogButtonSizer.

The button must have one of the following identifiers:

destroy/1

Specs

destroy(This :: wxStdDialogButtonSizer()) -> ok.

Destroys the object.

new/0

Specs

new() -> wxStdDialogButtonSizer().

Constructor for a wxStdDialogButtonSizer.

realize/1

Specs

realize(This) -> ok when This :: wxStdDialogButtonSizer().

Rearranges the buttons and applies proper spacing between buttons to make them match the platform or toolkit's interface guidelines.

setAffirmativeButton/2

Specs

setAffirmativeButton(This, Button) -> ok
                        when This :: wxStdDialogButtonSizer(), Button :: wxButton:wxButton().

Sets the affirmative button for the sizer.

This allows you to use identifiers other than the standard identifiers outlined above.

setCancelButton/2

Specs

setCancelButton(This, Button) -> ok
                   when This :: wxStdDialogButtonSizer(), Button :: wxButton:wxButton().

Sets the cancel button for the sizer.

This allows you to use identifiers other than the standard identifiers outlined above.

setNegativeButton/2

Specs

setNegativeButton(This, Button) -> ok
                     when This :: wxStdDialogButtonSizer(), Button :: wxButton:wxButton().

Sets the negative button for the sizer.

This allows you to use identifiers other than the standard identifiers outlined above.