wxStyledTextEvent (wx v2.0.1)

The type of events sent from wxStyledTextCtrl.

This class is derived (and can use functions) from: wxCommandEvent wxEvent

wxWidgets docs: wxStyledTextEvent

Events

Use wxEvtHandler:connect/3 with wxStyledTextEventType to subscribe to events of this type.

Types

-type wxStyledText() :: term().

Specs

wxStyledText() :: #wxStyledText{}.

-type wxStyledTextEvent() :: term().

Specs

wxStyledTextEvent() :: wx:wx_object().

-type wxStyledTextEventType() :: term().

Specs

wxStyledTextEventType() ::
    stc_autocomp_cancelled | stc_autocomp_char_deleted | stc_autocomp_selection |
    stc_calltip_click | stc_change | stc_charadded | stc_do_drop | stc_doubleclick |
    stc_drag_over | stc_dwellend | stc_dwellstart | stc_hotspot_click | stc_hotspot_dclick |
    stc_hotspot_release_click | stc_indicator_click | stc_indicator_release | stc_macrorecord |
    stc_marginclick | stc_modified | stc_needshown | stc_painted | stc_romodifyattempt |
    stc_savepointleft | stc_savepointreached | stc_start_drag | stc_styleneeded | stc_updateui |
    stc_userlistselection | stc_zoom.

Functions

getAlt/1

Specs

getAlt(This) -> boolean() when This :: wxStyledTextEvent().

Returns true if the Alt key is pressed.

This method is valid for the following event types:

getControl/1

Specs

getControl(This) -> boolean() when This :: wxStyledTextEvent().

Returns true if the Control key is pressed.

This method is valid for the following event types:

getDragAllowMove/1

Specs

getDragAllowMove(This) -> boolean() when This :: wxStyledTextEvent().

getDragResult/1

Specs

getDragResult(This) -> wx:wx_enum() when This :: wxStyledTextEvent().

Returns drag result for this event.

This method is valid for wxEVT_STC_DRAG_OVER and wxEVT_STC_DO_DROP events.

getDragText/1

Specs

getDragText(This) -> unicode:charlist() when This :: wxStyledTextEvent().

Deprecated: Use wxCommandEvent:getString/1 instead.

getFoldLevelNow/1

Specs

getFoldLevelNow(This) -> integer() when This :: wxStyledTextEvent().

Returns the current fold level for the line.

This method is valid for wxEVT_STC_MODIFIED events when the result of getModificationType/1 includes ?wxSTC_MOD_CHANGEFOLD.

getFoldLevelPrev/1

Specs

getFoldLevelPrev(This) -> integer() when This :: wxStyledTextEvent().

Returns previous fold level for the line.

This method is valid for wxEVT_STC_MODIFIED events when the result of getModificationType/1 includes ?wxSTC_MOD_CHANGEFOLD.

getKey/1

Specs

getKey(This) -> integer() when This :: wxStyledTextEvent().

Returns the key code of the key that generated this event.

This method is valid for the following event types:

getLParam/1

Specs

getLParam(This) -> integer() when This :: wxStyledTextEvent().

Returns the value of the LParam field for this event.

This method is valid for wxEVT_STC_MACRORECORD events.

getLength/1

Specs

getLength(This) -> integer() when This :: wxStyledTextEvent().

Returns the length (number of characters) of this event.

This method is valid for wxEVT_STC_MODIFIED and wxEVT_STC_NEEDSHOWN events.

getLine/1

Specs

getLine(This) -> integer() when This :: wxStyledTextEvent().

Returns zero-based line number for this event.

This method is valid for wxEVT_STC_DOUBLECLICK and wxEVT_STC_MODIFIED events.

getLinesAdded/1

Specs

getLinesAdded(This) -> integer() when This :: wxStyledTextEvent().

Returns the number of lines added or deleted with this event.

This method is valid for wxEVT_STC_MODIFIED events when the result of getModificationType/1 includes ?wxSTC_MOD_INSERTTEXT or ?wxSTC_MOD_DELETETEXT.

getListType/1

Specs

getListType(This) -> integer() when This :: wxStyledTextEvent().

Returns the list type for this event.

The list type is an integer passed to a list when it is created with the wxStyledTextCtrl:userListShow/3 method and can be used to distinguish lists if more than one is used.

This method is valid for wxEVT_STC_AUTOCOMP_SELECTION_CHANGE and wxEVT_STC_USERLISTSELECTION events.

getMargin/1

Specs

getMargin(This) -> integer() when This :: wxStyledTextEvent().

Returns the zero-based index of the margin that generated this event.

This method is valid for wxEVT_STC_MARGINCLICK and wxEVT_STC_MARGIN_RIGHT_CLICK events.

getMessage/1

Specs

getMessage(This) -> integer() when This :: wxStyledTextEvent().

Returns a message number while a macro is being recorded.

Many of the wxStyledTextCtrl methods such as wxStyledTextCtrl:insertText/3 and wxStyledTextCtrl:paste/1 have an event number associated with them. This method returns that number while a macro is being recorded so that the macro can be played back later.

This method is valid for wxEVT_STC_MACRORECORD events.

getModificationType/1

Specs

getModificationType(This) -> integer() when This :: wxStyledTextEvent().

Returns the modification type for this event.

The modification type is a bit list that describes the change that generated this event. It may contain one or more of the following values:

This method is valid for wxEVT_STC_MODIFIED events.

getModifiers/1

Specs

getModifiers(This) -> integer() when This :: wxStyledTextEvent().

Returns the modifiers of the key press or mouse click for this event.

The returned value is a bit list that may contain one or more of the following values:

In addition, the value can be checked for equality with ?wxSTC_KEYMOD_NORM to test if no modifiers are present.

This method is valid for the following event types:

getPosition/1

Specs

getPosition(This) -> integer() when This :: wxStyledTextEvent().

Returns the zero-based text position associated this event.

This method is valid for the following event types:

getShift/1

Specs

getShift(This) -> boolean() when This :: wxStyledTextEvent().

Returns true if the Shift key is pressed.

This method is valid for the following event types:

getText/1

Specs

getText(This) -> unicode:charlist() when This :: wxStyledTextEvent().

Deprecated: Use wxCommandEvent:getString/1 instead.

getWParam/1

Specs

getWParam(This) -> integer() when This :: wxStyledTextEvent().

Returns value of the WParam field for this event.

This method is valid for wxEVT_STC_MACRORECORD events.

getX/1

Specs

getX(This) -> integer() when This :: wxStyledTextEvent().

Returns the X coordinate of the mouse for this event.

This method is valid for the following event types:

getY/1

Specs

getY(This) -> integer() when This :: wxStyledTextEvent().

Returns the Y coordinate of the mouse for this event.

This method is valid for the following event types: