wxTextAttr (wx v2.0.1)

wxTextAttr represents the character and paragraph attributes, or style, for a range of text in a wxTextCtrl or wxRichTextCtrl (not implemented in wx).

When setting up a wxTextAttr object, pass a bitlist mask to setFlags/2 to indicate which style elements should be changed. As a convenience, when you call a setter such as SetFont, the relevant bit will be set.

See: wxTextCtrl, wxRichTextCtrl (not implemented in wx)

wxWidgets docs: wxTextAttr

Types

-type wxTextAttr() :: term().

Specs

wxTextAttr() :: wx:wx_object().

Functions

destroy/1

Specs

destroy(This :: wxTextAttr()) -> ok.

Destroys the object.

getAlignment/1

Specs

getAlignment(This) -> wx:wx_enum() when This :: wxTextAttr().

Returns the alignment flags.

See ?wxTextAttrAlignment for a list of available styles.

getBackgroundColour/1

Specs

getBackgroundColour(This) -> wx:wx_colour4() when This :: wxTextAttr().

Returns the background colour.

getFlags/1

Specs

getFlags(This) -> integer() when This :: wxTextAttr().

Returns flags indicating which attributes are applicable.

See setFlags/2 for a list of available flags.

getFont/1

Specs

getFont(This) -> wxFont:wxFont() when This :: wxTextAttr().

Creates and returns a font specified by the font attributes in the wxTextAttr object.

Note that wxTextAttr does not store a wxFont object, so this is only a temporary font.

For greater efficiency, access the font attributes directly.

getLeftIndent/1

Specs

getLeftIndent(This) -> integer() when This :: wxTextAttr().

Returns the left indent in tenths of a millimetre.

getLeftSubIndent/1

Specs

getLeftSubIndent(This) -> integer() when This :: wxTextAttr().

Returns the left sub-indent in tenths of a millimetre.

getRightIndent/1

Specs

getRightIndent(This) -> integer() when This :: wxTextAttr().

Returns the right indent in tenths of a millimeter.

getTabs/1

Specs

getTabs(This) -> [integer()] when This :: wxTextAttr().

Returns an array of tab stops, each expressed in tenths of a millimeter.

Each stop is measured from the left margin and therefore each value must be larger than the last.

getTextColour/1

Specs

getTextColour(This) -> wx:wx_colour4() when This :: wxTextAttr().

Returns the text foreground colour.

hasBackgroundColour/1

Specs

hasBackgroundColour(This) -> boolean() when This :: wxTextAttr().

Returns true if the attribute object specifies a background colour.

hasFont/1

Specs

hasFont(This) -> boolean() when This :: wxTextAttr().

Returns true if the attribute object specifies any font attributes.

hasTextColour/1

Specs

hasTextColour(This) -> boolean() when This :: wxTextAttr().

Returns true if the attribute object specifies a text foreground colour.

isDefault/1

Specs

isDefault(This) -> boolean() when This :: wxTextAttr().

Returns false if we have any attributes set, true otherwise.

new/0

Specs

new() -> wxTextAttr().

Constructors.

new/1

Specs

new(ColText) -> wxTextAttr() when ColText :: wx:wx_colour();
   (Attr) -> wxTextAttr() when Attr :: wxTextAttr().

new/2

Specs

new(ColText, [Option]) -> wxTextAttr()
       when
           ColText :: wx:wx_colour(),
           Option ::
               {colBack, wx:wx_colour()} | {font, wxFont:wxFont()} | {alignment, wx:wx_enum()}.

setAlignment/2

Specs

setAlignment(This, Alignment) -> ok when This :: wxTextAttr(), Alignment :: wx:wx_enum().

Sets the paragraph alignment.

See ?wxTextAttrAlignment enumeration values.

Of these, wxTEXT_ALIGNMENT_JUSTIFIED is unimplemented. In future justification may be supported when printing or previewing, only.

setBackgroundColour/2

Specs

setBackgroundColour(This, ColBack) -> ok when This :: wxTextAttr(), ColBack :: wx:wx_colour().

Sets the background colour.

setFlags/2

Specs

setFlags(This, Flags) -> ok when This :: wxTextAttr(), Flags :: integer().

Sets the flags determining which styles are being specified.

The ?wxTextAttrFlags values can be passed in a bitlist.

setFont/2

Specs

setFont(This, Font) -> ok when This :: wxTextAttr(), Font :: wxFont:wxFont().

setFont/3

Specs

setFont(This, Font, [Option]) -> ok
           when This :: wxTextAttr(), Font :: wxFont:wxFont(), Option :: {flags, integer()}.

Sets the attributes for the given font.

Note that wxTextAttr does not store an actual wxFont object.

setLeftIndent/2

Specs

setLeftIndent(This, Indent) -> ok when This :: wxTextAttr(), Indent :: integer().

setLeftIndent/3

Specs

setLeftIndent(This, Indent, [Option]) -> ok
                 when This :: wxTextAttr(), Indent :: integer(), Option :: {subIndent, integer()}.

Sets the left indent and left subindent in tenths of a millimetre.

The sub-indent is an offset from the left of the paragraph, and is used for all but the first line in a paragraph.

A positive value will cause the first line to appear to the left of the subsequent lines, and a negative value will cause the first line to be indented relative to the subsequent lines.

wxRichTextBuffer (not implemented in wx) uses indentation to render a bulleted item. The left indent is the distance between the margin and the bullet. The content of the paragraph, including the first line, starts at leftMargin + leftSubIndent. So the distance between the left edge of the bullet and the left of the actual paragraph is leftSubIndent.

setRightIndent/2

Specs

setRightIndent(This, Indent) -> ok when This :: wxTextAttr(), Indent :: integer().

Sets the right indent in tenths of a millimetre.

setTabs/2

Specs

setTabs(This, Tabs) -> ok when This :: wxTextAttr(), Tabs :: [integer()].

Sets the tab stops, expressed in tenths of a millimetre.

Each stop is measured from the left margin and therefore each value must be larger than the last.

setTextColour/2

Specs

setTextColour(This, ColText) -> ok when This :: wxTextAttr(), ColText :: wx:wx_colour().

Sets the text foreground colour.