wxWebViewEvent (wx v2.0.1)

A navigation event holds information about events associated with wxWebView objects.

Since: 2.9.3

See: wxWebView

This class is derived (and can use functions) from: wxNotifyEvent wxCommandEvent wxEvent

wxWidgets docs: wxWebViewEvent

Events

Use wxEvtHandler:connect/3 with wxWebViewEventType to subscribe to events of this type.

Types

-type wxWebView() :: term().

Specs

wxWebView() :: #wxWebView{}.

-type wxWebViewEvent() :: term().

Specs

wxWebViewEvent() :: wx:wx_object().

-type wxWebViewEventType() :: term().

Specs

wxWebViewEventType() ::
    webview_navigating | webview_navigated | webview_loaded | webview_error | webview_newwindow |
    webview_title_changed.

Functions

getInt/1

Specs

getInt(This) -> integer() when This :: wxWebViewEvent().

Returns the integer identifier corresponding to a listbox, choice or radiobox selection (only if the event was a selection, not a deselection), or a boolean value representing the value of a checkbox.

For a menu item, this method returns -1 if the item is not checkable or a boolean value (true or false) for checkable items indicating the new state of the item.

getString/1

Specs

getString(This) -> unicode:charlist() when This :: wxWebViewEvent().

Returns item string for a listbox or choice selection event.

If one or several items have been deselected, returns the index of the first deselected item. If some items have been selected and others deselected at the same time, it will return the index of the first selected item.

getTarget/1

Specs

getTarget(This) -> unicode:charlist() when This :: wxWebViewEvent().

Get the name of the target frame which the url of this event has been or will be loaded into.

This may return an empty string if the frame is not available.

getURL/1

Specs

getURL(This) -> unicode:charlist() when This :: wxWebViewEvent().

Get the URL being visited.