wx_object behaviour (wx v2.0.1)

wx_object - Generic wx object behaviour

This is a behaviour module that can be used for "sub classing" wx objects. It works like a regular gen_server module and creates a server per object.

NOTE: Currently no form of inheritance is implemented.

The user module should export:

init(Args) should return
{wxObject, State} | {wxObject, State, Timeout} | ignore | {stop, Reason}

Asynchronous window event handling:
handle_event(#wx{}, State) should return
{noreply, State} | {noreply, State, Timeout} | {stop, Reason, State}

The user module can export the following callback functions:

handle_call(Msg, {From, Tag}, State) should return
{reply, Reply, State} | {reply, Reply, State, Timeout} | {noreply, State} | {noreply, State, Timeout} | {stop, Reason, Reply, State}

handle_cast(Msg, State) should return
{noreply, State} | {noreply, State, Timeout} | {stop, Reason, State}

If the above are not exported but called, the wx_object process will crash. The user module can also export:

Info is message e.g. {'EXIT', P, R}, {nodedown, N}, ...
handle_info(Info, State) should return , ...
{noreply, State} | {noreply, State, Timeout} | {stop, Reason, State}

If a message is sent to the wx_object process when handle_info is not exported, the message will be dropped and ignored.

When stop is returned in one of the functions above with Reason = normal | shutdown | Term, terminate(State) is called. It lets the user module clean up, it is always called when server terminates or when wx_object() in the driver is deleted. If the Parent process terminates the Module:terminate/2 function is called.
terminate(Reason, State)

Example:

  -module(myDialog).
  -export([new/2, show/1, destroy/1]).  %% API
  -export([init/1, handle_call/3, handle_event/2,
           handle_info/2, code_change/3, terminate/2]).
           new/2, showModal/1, destroy/1]).  %% Callbacks
 
  %% Client API
  new(Parent, Msg) ->
     wx_object:start(?MODULE, [Parent,Id], []).
 
  show(Dialog) ->
     wx_object:call(Dialog, show_modal).
 
  destroy(Dialog) ->
     wx_object:call(Dialog, destroy).
 
  %% Server Implementation ala gen_server
  init([Parent, Str]) ->
     Dialog = wxDialog:new(Parent, 42, "Testing", []),
     ...
     wxDialog:connect(Dialog, command_button_clicked),
     {Dialog, MyState}.
 
  handle_call(show, _From, State) ->
     wxDialog:show(State#state.win),
     {reply, ok, State};
  ...
  handle_event(#wx{}, State) ->
     io:format("Users clicked button~n",[]),
     {noreply, State};
  ...

DATA TYPES

request_id() = term()

server_ref() = wx:wx_object() | atom() | pid()

Functions

call(Obj, Request) -> term()

Specs

call(Obj, Request) -> term() when Obj :: wx:wx_object() | atom() | pid(), Request :: term().

Obj = wx:wx_object() | atom() | pid()

Request = term()

Make a call to a wx_object server. The call waits until it gets a result. Invokes handle_call(Request, From, State) in the server

call(Obj, Request, Timeout) -> term()

Specs

call(Obj, Request, Timeout) -> term()
        when Obj :: wx:wx_object() | atom() | pid(), Request :: term(), Timeout :: integer().

Obj = wx:wx_object() | atom() | pid()

Request = term()

Timeout = integer()

Make a call to a wx_object server with a timeout. Invokes handle_call(Request, From, State) in server

cast(Obj, Request) -> ok

Specs

cast(Obj, Request) -> ok when Obj :: wx:wx_object() | atom() | pid(), Request :: term().

Obj = wx:wx_object() | atom() | pid()

Request = term()

Make a cast to a wx_object server. Invokes handle_cast(Request, State) in the server

check_response(Msg::term(), Key:: request_id() ) -> {reply, Reply::term()} | false | {error, {term(), server_ref() }}

Specs

check_response(Msg :: term(), Key :: request_id()) ->
                  {reply, Reply :: term()} | false | {error, {term(), server_ref()}}.

Check if a received message was a reply to a RequestId

get_pid(Obj) -> pid()

Specs

get_pid(Obj) -> pid() when Obj :: wx:wx_object() | atom() | pid().

Obj = wx:wx_object() | atom() | pid()

Get the pid of the object handle.

reply(X1::{pid(), Tag::term()}, Reply::term()) -> pid()

Specs

reply({pid(), Tag :: term()}, Reply :: term()) -> pid().

Get the pid of the object handle.

send_request(Obj, Request::term()) -> request_id()

Specs

send_request(Obj, Request :: term()) -> request_id() when Obj :: wx:wx_object() | atom() | pid().

Obj = wx:wx_object() | atom() | pid()

Make an send_request to a generic server. and return a RequestId which can/should be used with wait_response/[1|2]. Invokes handle_call(Request, From, State) in server.

set_pid(Obj, Pid::pid()) -> wx:wx_object()

Specs

set_pid(Obj, pid()) -> wx:wx_object() when Obj :: wx:wx_object() | atom() | pid().

Obj = wx:wx_object() | atom() | pid()

Sets the controlling process of the object handle.

start(Name, Mod, Args, Options) -> wxWindow:wxWindow() | {error, term()}

Specs

start(Name, Mod, Args, Options) -> wxWindow:wxWindow() | {error, term()}
         when
             Name :: {local, atom()},
             Mod :: atom(),
             Args :: term(),
             Flag :: trace | log | {logfile, string()} | statistics | debug,
             Options :: [{timeout, timeout()} | {debug, [Flag]}].

Name = {local, atom()}

Mod = atom()

Args = term()

Flag = trace | log | {logfile, string()} | statistics | debug

Options = [{timeout, timeout()} | {debug, [Flag]}]

Starts a generic wx_object server and invokes Mod:init(Args) in the new process.

start_link(Mod, Args, Options) -> wxWindow:wxWindow() | {error, term()}

Specs

start_link(Mod, Args, Options) -> wxWindow:wxWindow() | {error, term()}
              when
                  Mod :: atom(),
                  Args :: term(),
                  Flag :: trace | log | {logfile, string()} | statistics | debug,
                  Options :: [{timeout, timeout()} | {debug, [Flag]}].

Mod = atom()

Args = term()

Flag = trace | log | {logfile, string()} | statistics | debug

Options = [{timeout, timeout()} | {debug, [Flag]}]

Starts a generic wx_object server and invokes Mod:init(Args) in the new process.

start_link(Name, Mod, Args, Options) -> wxWindow:wxWindow() | {error, term()}

Specs

start_link(Name, Mod, Args, Options) -> wxWindow:wxWindow() | {error, term()}
              when
                  Name :: {local, atom()},
                  Mod :: atom(),
                  Args :: term(),
                  Flag :: trace | log | {logfile, string()} | statistics | debug,
                  Options :: [{timeout, timeout()} | {debug, [Flag]}].

Name = {local, atom()}

Mod = atom()

Args = term()

Flag = trace | log | {logfile, string()} | statistics | debug

Options = [{timeout, timeout()} | {debug, [Flag]}]

Starts a generic wx_object server and invokes Mod:init(Args) in the new process.

stop(Obj) -> ok

Specs

stop(Obj) -> ok when Obj :: wx:wx_object() | atom() | pid().

Obj = wx:wx_object() | atom() | pid()

Stops a generic wx_object server with reason 'normal'. Invokes terminate(Reason,State) in the server. The call waits until the process is terminated. If the process does not exist, an exception is raised.

stop(Obj, Reason, Timeout) -> ok

Specs

stop(Obj, Reason, Timeout) -> ok
        when Obj :: wx:wx_object() | atom() | pid(), Reason :: term(), Timeout :: timeout().

Obj = wx:wx_object() | atom() | pid()

Reason = term()

Timeout = timeout()

Stops a generic wx_object server with the given Reason. Invokes terminate(Reason,State) in the server. The call waits until the process is terminated. If the call times out, or if the process does not exist, an exception is raised.

wait_response(RequestId:: request_id() ) -> {reply, Reply::term()} | {error, {term(), server_ref() }}

Specs

wait_response(RequestId :: request_id()) ->
                 {reply, Reply :: term()} | {error, {term(), server_ref()}}.

Wait infinitely for a reply from a generic server.

wait_response(Key:: request_id() , Timeout::timeout()) -> {reply, Reply::term()} | timeout | {error, {term(), server_ref() }}

Specs

wait_response(Key :: request_id(), timeout()) ->
                 {reply, Reply :: term()} | timeout | {error, {term(), server_ref()}}.

Wait 'timeout' for a reply from a generic server.